% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areNonExisting.R
\name{areNonExisting}
\alias{areNonExisting}
\alias{are_non_existing}
\alias{assertAreNonExisting}
\alias{assert_are_non_existing}
\title{Does the Variable Not Exist?}
\usage{
areNonExisting(x, envir = parent.frame(), inherits = TRUE)

are_non_existing(x, envir = parent.frame(), inherits = TRUE)

assertAreNonExisting(x, envir = parent.frame(), inherits = TRUE,
  .var.name = vname(x), add = NULL)

assert_are_non_existing(x, envir = parent.frame(), inherits = TRUE,
  .var.name = vname(x), add = NULL)
}
\arguments{
\item{x}{\code{character}. Variable names to check in \code{environment}.}

\item{envir}{\code{environment}. Environment.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Variable Not Exist?
}
\examples{
areNonExisting(c("XXX", "YYY"))
}
\seealso{
\itemize{
\item \code{checkmate::makeAssertionFunction()}.
}
}
