% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areSameLength.R
\name{areSameLength}
\alias{areSameLength}
\title{Do the Inputs Have the Same Length?}
\usage{
areSameLength(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Secondary object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop}) on failure.
}
}
\description{
Do the Inputs Have the Same Length?
}
\note{
Non-zero lengths for \code{x} and \code{y} are required, otherwise the check
function will intentionally error.
}
\examples{
x <- list(a = 1, b = 2)
y <- list(c = 3, d = 4)
areSameLength(x = x, y = y)
}
