% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areSamplesUnique.R
\name{areSamplesUnique}
\alias{areSamplesUnique}
\title{Does the Object Contain Unique Samples?}
\usage{
areSamplesUnique(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{base::stop()} or \code{base::stopifnot()} for details).
}
}
\description{
Checks a matrix for duplicated columns, which reprent samples. Duplicate
rows are allowed here, because many genes (rows) can contain all zeros.
}
\section{SummarizedExperiment}{


If the object is a \code{SummarizedExperiment}, then the primary \code{assay()} matrix
is checked for duplicated columns.
}

\examples{
## Unique.
x <- matrix(data = seq_len(20L), ncol = 2L)
areSamplesUnique(x)

## Duplicated.
x <- matrix(data = rep(seq_len(10L), times = 2L), ncol = 2L)
areSamplesUnique(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
