% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formals.R
\name{assertFormalCompress}
\alias{assertFormalCompress}
\title{Check Compression Formal}
\usage{
assertFormalCompress(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Check Compression Formal
}
\examples{
assertFormalCompress("xz")
}
\seealso{
\itemize{
\item \code{checkmate::makeAssertionFunction()}.
}
}
