% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertHasAggregateInfo.R
\name{assertHasAggregateInfo}
\alias{assertHasAggregateInfo}
\alias{hasAggregateInfo}
\title{Assert Has Aggregate Information}
\usage{
assertHasAggregateInfo(object)

hasAggregateInfo(object, stop = FALSE)
}
\description{
Determine whether an object can be aggregated automatically.
}
