% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isScalarOrNULL.R
\name{assertIsAnIntegerOrNULL}
\alias{assertIsAnIntegerOrNULL}
\title{Is an Integer (or NULL)?}
\usage{
assertIsAnIntegerOrNULL(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions  (see \code{stop()}for details).
}
}
\description{
Is an Integer (or NULL)?
}
\examples{
assertIsAnIntegerOrNULL(1L)
assertIsAnIntegerOrNULL(NULL)
}
