% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsANumberOrNULL.R
\name{assertIsANumberOrNULL}
\alias{assertIsANumberOrNULL}
\title{Assert Is a Number or NULL}
\usage{
assertIsANumberOrNULL(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions  (see \code{stop()}for details).
}
}
\description{
Assert Is a Number or NULL
}
\examples{
assertIsANumberOrNULL(1.1)
assertIsANumberOrNULL(NULL)
}
