% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsAStringOrNULL.R
\name{assertIsAStringOrNULL}
\alias{assertIsAStringOrNULL}
\title{Assert Is a String or NULL}
\usage{
assertIsAStringOrNULL(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions  (see \code{stop()}for details).
}
}
\description{
Assert Is a String or NULL
}
\examples{
assertIsAStringOrNULL("hello world")
assertIsAStringOrNULL(NULL)
}
