% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertIsFile.R
\name{assertIsFile}
\alias{assertIsFile}
\title{Assert Is File}
\usage{
assertIsFile(file)
}
\value{
Always stop on error.
}
\description{
Assert Is File
}
\examples{
assertIsFile(system.file("extdata/example.rds", package = "basejump"))
}
