% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{assertIsHexColorFunctionOrNULL}
\alias{assertIsHexColorFunctionOrNULL}
\title{Is Hex Color Function (or NULL)?}
\usage{
assertIsHexColorFunctionOrNULL(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Is Hex Color Function (or NULL)?
}
\examples{
hex <- function(n) {
    colors <- c("#FFFFFF", "#000000")
    colors[seq_len(n)]
}
assertIsHexColorFunctionOrNULL(hex)
}
