% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isScalarOrNULL.R
\name{assertIsStringOrNULL}
\alias{assertIsStringOrNULL}
\title{Is String (or NULL)?}
\usage{
assertIsStringOrNULL(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions  (see \code{stop()}for details).
}
}
\description{
Is String (or NULL)?
}
\examples{
assertIsStringOrNULL("hello world")
assertIsStringOrNULL(NULL)
}
