% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-bapply.R
\name{bapply}
\alias{bapply}
\title{Apply a Function over a List or Vector}
\usage{
bapply(x, predicate, ...)
}
\arguments{
\item{x}{\code{atomic} or \code{list}.}

\item{predicate}{\code{function}. A predicate function that returns \code{logical(1)}
(boolean flag) to \code{apply}.}

\item{...}{Additional arguments passed to \link{vapply}.}
}
\value{
\code{logical}.
}
\description{
\code{bapply} is a wrapper to \code{vapply} for functions that return \code{logical(1)}.
}
\note{
\code{USE.NAMES} is always set to TRUE.
}
\examples{
bapply(list(a = "example", b = 1), is.character)
}
\seealso{
\itemize{
\item \code{assertive.base::bapply}.
\item \code{vapply}.
}
}
