% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGE.R
\docType{package}
\name{basejump.assertions-package}
\alias{basejump.assertions}
\alias{basejump.assertions-package}
\title{basejump.assertions}
\description{
Assertive check functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://steinbaugh.com/basejump}
  \item \url{https://github.com/steinbaugh/basejump}
  \item Report bugs at \url{https://github.com/steinbaugh/basejump/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

}
