% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-basenameSansExt.R
\name{basenameSansExt}
\alias{basenameSansExt}
\title{Base name without the file extension}
\usage{
basenameSansExt(path)
}
\arguments{
\item{path}{\code{character}.
File path(s).
This function is vectorized and supports multiple files.}
}
\value{
\code{character}.
Character vector of same length as \code{path} input, with file extension removed.
Returns \code{NA} if no extension is detected.
}
\description{
Base name without the file extension
}
\note{
This function intentionally doesn't check whether a file exists.

Updated 2019-10-12.
}
\examples{
basenameSansExt(c("dir/foo.txt", "dir/bar.tar.gz", "dir/"))
}
\seealso{
\itemize{
\item \code{\link[base:basename]{basename()}}.
\item \code{tools::file_path_sans_ext()}.
}
}
