% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-R.utils-capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize}
\usage{
capitalize(x)
}
\arguments{
\item{x}{\code{vector} of \code{character} strings to be capitalized.}
}
\description{
Capitalize each character string in a vector.
}
\examples{
capitalize(c("hello", "world"))
}
\seealso{
\code{R.utils::capitalize()}.\preformatted{getS3method(
    f = "capitalize",
    class = "default",
    envir = asNamespace("R.utils")
)
}
}
\keyword{internal}
