% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-assertive-cause.R
\name{cause}
\alias{cause}
\alias{cause<-}
\title{Get or set the \code{cause} attribute}
\usage{
cause(x)

cause(x) <- value
}
\value{
\code{character(1)}. Cause attribute.
}
\description{
Gets or sets the \code{cause} (of failure) \link[base:attributes]{attribute} of a
variable.
}
\examples{
## FIXME

## ## Scalar case
## yn <- assertive.base::is_identical_to_true(FALSE)
## cause(yn)

## ## Vector case
## yn <- assertive.base::is_true(c(TRUE, FALSE, NA))
## cause(yn)
}
\seealso{
\itemize{
\item \code{assertive.base::cause()}.
\item \code{attributes()}.
\item \code{attr()}.
}
}
