% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-allAreAtomic.R
\name{check-scalar-allAreAtomic}
\alias{check-scalar-allAreAtomic}
\alias{allAreAtomic}
\title{Does the input contain elements that are all atomic?}
\usage{
allAreAtomic(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain elements that are all atomic?
}
\note{
Updated 2023-09-29.
}
\examples{
## TRUE ====
allAreAtomic(data.frame(a = "foo", b = "bar"))
allAreAtomic(list(a = "foo", b = "bar"))

## FALSE ====
allAreAtomic(data.frame())
allAreAtomic(list(a = "x", b = list()))
}
\seealso{
\code{is.atomic()}.
}
