% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-formalCompress.R
\name{check-scalar-formalCompress}
\alias{check-scalar-formalCompress}
\alias{formalCompress}
\title{Check the \code{compress} formal argument}
\usage{
formalCompress(compress)
}
\arguments{
\item{compress}{\code{logical(1)} or \code{character(1)}.
These character strings are currently allowed for \code{\link[base:save]{save()}}:
\code{"gzip"}, \code{"bzip2"}, or \code{"xz"}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Check the \code{compress} formal argument
}
\note{
Updated 2019-08-10.
}
\examples{
## TRUE ====
formalCompress("gzip")
formalCompress(TRUE)

## FALSE ====
formalCompress(NA)
formalCompress("xxx")
}
