% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasClusters.R
\name{check-scalar-hasClusters}
\alias{check-scalar-hasClusters}
\alias{hasClusters}
\title{Does the input object contain clusters?}
\usage{
hasClusters(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input object contain clusters?
}
\note{
Updated 2022-03-04.
}
\examples{
data(SingleCellExperiment_Seurat, package = "AcidTest")

## TRUE ====
if (requireNamespace("AcidSingleCell", quietly = TRUE)) {
    hasClusters(SingleCellExperiment_Seurat)
}

## FALSE ====
hasClusters(list())
}
\seealso{
\code{AcidGenerics::clusters}.
}
