% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-hasCPU.R
\name{check-scalar-hasCPU}
\alias{check-scalar-hasCPU}
\alias{hasCPU}
\title{Does the current machine have a sufficient number of CPU cores?}
\usage{
hasCPU(n)
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the current machine have a sufficient number of CPU cores?
}
\note{
Updated 2023-08-10.
}
\examples{
## TRUE ====
hasCPU(n = 1L)

## FALSE ====
hasCPU(n = Inf)
}
