% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasDims.R
\name{check-scalar-hasDims}
\alias{check-scalar-hasDims}
\alias{hasDims}
\alias{hasRows}
\alias{hasCols}
\title{Does the input have dimensions?}
\usage{
hasDims(x, .xname = getNameInParent(x))

hasRows(x, .xname = getNameInParent(x))

hasCols(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input have dimensions?
}
\note{
Updated 2019-07-29.
}
\examples{
## TRUE ====
x <- datasets::mtcars
hasDims(x)
hasRows(x)
hasCols(x)

## Note that dims don't have to be non-zero, just not NULL.
hasDims(data.frame())

## FALSE ====
x <- data.frame()
hasDims(list())
hasRows(x)
hasCols(x)
}
\seealso{
\itemize{
\item \code{assertive.properties::has_dims()}.
\item \code{assertive.properties::has_rows()}.
\item \code{assertive.properties::has_cols()}.
}
}
