% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasDuplicates.R
\name{check-scalar-hasDuplicates}
\alias{check-scalar-hasDuplicates}
\alias{hasDuplicates}
\alias{hasNoDuplicates}
\title{Does the input have duplicates?}
\usage{
hasDuplicates(x)

hasNoDuplicates(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input have duplicates?
}
\note{
Updated 2023-01-12.
}
\examples{
## TRUE ====
hasDuplicates(c("a", "a"))
hasNoDuplicates(c("a", "b"))

## FALSE ====
hasDuplicates(c("a", "b"))
hasNoDuplicates(c("a", "a", "b", "b"))
}
\seealso{
\itemize{
\item \code{assertive.properties::has_duplicates()}.
\item \code{assertive.properties::has_no_duplicates()}.
}
}
