% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasMetrics.R
\name{check-scalar-hasMetrics}
\alias{check-scalar-hasMetrics}
\title{Does the object contain quality control metrics?}
\arguments{
\item{colnames}{\code{character}.
Column names in \code{\link[SummarizedExperiment:colData]{colData()}} containing
expected quality control metrics.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the object contain quality control metrics?
}
\note{
Updated 2019-08-12.
}
\examples{
if (requireNamespace("SummarizedExperiment", quietly = TRUE)) {
    data(RangedSummarizedExperiment, package = "acidtest")
    x <- RangedSummarizedExperiment
    hasMetrics(x)
}
}
