% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasRownames.R
\name{check-scalar-hasRownames}
\alias{check-scalar-hasRownames}
\alias{hasRownames}
\title{Does the input have row names?}
\usage{
hasRownames(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input have row names?
}
\note{
Updated 2019-07-29.
}
\section{data.frame}{


Standard \code{data.frame} class objects cannot have \code{NULL} row names defined.
Here we are checking to see if a \code{data.frame} has soft \code{NULL} row names,
meaning that they return as a sequence that is identical to the number of
rows.
}

\examples{
## TRUE ====
x <- data.frame(
    "sample1" = c(1L, 2L),
    "sample2" = c(3L, 4L),
    row.names = c("gene1", "gene2")
)
print(x)
hasRownames(x)

## FALSE ====
x <- data.frame(a = seq_len(2L))
print(x)
# Standard data frame doesn't allow NULL row names.
rownames(x)
hasRownames(x)

x <- S4Vectors::DataFrame(a = seq_len(2L))
print(x)
# S4 data frame does allow NULL row names.
rownames(x)
hasRownames(x)
}
