% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasSubset.R
\name{check-scalar-hasSubset}
\alias{check-scalar-hasSubset}
\alias{hasSubset}
\title{Does the object contain a subset of data?}
\usage{
hasSubset(x, metadata = "subset")
}
\arguments{
\item{x}{\code{Annotated}.}

\item{metadata}{\code{character}.
Names in \code{\link[S4Vectors:Annotated-class]{metadata()}} that denote object is subset.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the object contain a subset of data?
}
\note{
Updated 2023-09-29.
}
\examples{
if (requireNamespace("SummarizedExperiment", quietly = TRUE)) {
    data(RangedSummarizedExperiment, package = "AcidTest")
    x <- RangedSummarizedExperiment
    hasSubset(x)
}
}
