% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasValidNames.R
\name{check-scalar-hasValidNames}
\alias{check-scalar-hasValidNames}
\alias{hasValidNames}
\alias{hasValidDimnames}
\title{Does the input object have syntactically valid names?}
\usage{
hasValidNames(x, .xname = getNameInParent(x))

hasValidDimnames(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input object have syntactically valid names?
}
\examples{
## TRUE ====
x <- list(a = 1, b = 2)
names(x)
hasValidNames(x)

x <- datasets::iris
lapply(dimnames(x), head)
hasValidDimnames(x)

## FALSE ====
x <- list(
    `1`       = 1,  # can't start with number
    `foo bar` = 2,  # no spaces
    `foo-bar` = 3   # no hyphens
)
print(x)
hasValidNames(x)

x <- datasets::mtcars
# Note the spaces in the row names here.
lapply(dimnames(x), head)
hasValidDimnames(x)
}
\seealso{
\code{validNames()}.
}
