% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-isBiocDevel.R
\name{check-scalar-isBiocDevel}
\alias{check-scalar-isBiocDevel}
\alias{isBiocDevel}
\title{Is the current Bioconductor installation under development?}
\usage{
isBiocDevel()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the current Bioconductor installation under development?
}
\note{
Updated 2023-09-29.
}
\examples{
isBiocDevel()
}
