% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isCharacter.R
\name{check-scalar-isCharacter}
\alias{check-scalar-isCharacter}
\alias{isCharacter}
\title{Does the input contain a (non-empty) character vector?}
\usage{
isCharacter(x, nullOK = FALSE, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input contain a (non-empty) character vector?
}
\section{Enforced parameters}{

\itemize{
\item Must have length.
\item Cannot contain \code{NA} strings.
\item Cannot contain empty (\code{""}) strings.
}
}

\examples{
## TRUE ====
isCharacter("a")

## FALSE ====
isCharacter(NULL)
isCharacter(character())
isCharacter("")
isCharacter(NA_character_)
}
\seealso{
\code{assertive.strings::is_non_missing_nor_empty_character()}.
}
