% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-isCleanSystemLibrary.R
\name{check-scalar-isCleanSystemLibrary}
\alias{check-scalar-isCleanSystemLibrary}
\alias{isCleanSystemLibrary}
\title{Is the R system library clean?}
\usage{
isCleanSystemLibrary()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the R system library clean?
}
\note{
Updated 2023-10-06.
}
\examples{
isCleanSystemLibrary()
}
