% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-system-scalar-isCondaEnabled.R
\name{check-scalar-isCondaEnabled}
\alias{check-scalar-isCondaEnabled}
\alias{isCondaEnabled}
\title{Is conda enabled (active) in the current R session?}
\usage{
isCondaEnabled(ignoreBase = TRUE)
}
\arguments{
\item{ignoreBase}{\code{logical(1)}.
If \code{TRUE}, don't consider conda to be "active" if only the base
environment is loaded. This is useful to avoid some false-positive
situations when handling the main conda \code{activate} script in shell
sessions.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is conda enabled (active) in the current R session?
}
\details{
This checks \code{CONDA_DEFAULT_ENV} and \code{CONDA_SHLVL} system environment
variables internally.
}
\note{
Updated 2021-08-19.
}
\examples{
vars <- c("CONDA_DEFAULT_ENV", "CONDA_SHLVL")
Sys.unsetenv(vars)

## TRUE ====
Sys.setenv(
    "CONDA_DEFAULT_ENV" = "test",
    "CONDA_SHLVL" = "2"
)
isCondaEnabled()
Sys.unsetenv(vars)

Sys.setenv(
    "CONDA_DEFAULT_ENV" = "base",
    "CONDA_SHLVL" = "1"
)
isCondaEnabled(ignoreBase = FALSE)
Sys.unsetenv(vars)

## FALSE ====
isCondaEnabled()

Sys.setenv(
    "CONDA_DEFAULT_ENV" = "base",
    "CONDA_SHLVL" = "1"
)
isCondaEnabled(ignoreBase = TRUE)
Sys.unsetenv(vars)
}
