% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isDark.R
\name{check-scalar-isDark}
\alias{check-scalar-isDark}
\alias{isDark}
\title{Is dark mode preferred?}
\usage{
isDark()
}
\value{
\code{logical(1)}.
}
\description{
Is dark mode preferred?
}
\details{
Checks for \code{acid.dark} option in current session.

Useful for changing the appearance of ggplot2 output, especially for single
cell RNA-seq clustering plots.
}
\note{
Updated 2022-03-08.
}
\examples{
## TRUE ====
options("acid.dark" = TRUE)
isDark()

## FALSE ====
options("acid.dark" = NULL)
isDark()
}
