% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-system-scalar-isDevel.R
\name{check-scalar-isDevel}
\alias{check-scalar-isDevel}
\alias{isDevel}
\title{Is the current session running inside R-devel?}
\usage{
isDevel()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the current session running inside R-devel?
}
\note{
Updated 2020-10-29.
}
\examples{
isDevel()
}
