% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-isDocker.R
\name{check-scalar-isDocker}
\alias{check-scalar-isDocker}
\alias{isDocker}
\title{Is the R session running inside Docker?}
\usage{
isDocker()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the R session running inside Docker?
}
\note{
Updated 2021-01-04.
}
\examples{
isDocker()
}
