% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-system-scalar-isDockerEnabled.R
\name{check-scalar-isDockerEnabled}
\alias{check-scalar-isDockerEnabled}
\alias{isDockerEnabled}
\title{Is Docker enabled (running) on the current machine?}
\usage{
isDockerEnabled()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is Docker enabled (running) on the current machine?
}
\note{
Updated 2021-09-21.
}
\examples{
isDockerEnabled()
}
