% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isFileSystemCaseSensitive.R
\name{check-scalar-isFileSystemCaseSensitive}
\alias{check-scalar-isFileSystemCaseSensitive}
\alias{isFileSystemCaseSensitive}
\title{Is the file system case sensitive?}
\usage{
isFileSystemCaseSensitive(dir = ".")
}
\arguments{
\item{dir}{\code{character(1)}.
Directory path.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the file system case sensitive?
}
\details{
Linux partitions in general are case sensitive by default.
macOS APFS and HFS partitions still default to case insensitive.
Windows partitions default to case insensitive.

This function checks for case sensitivity internally by creating (touching)
two invisible files with names that only differ by case. If 2 files are
detected, the file system is case sensitive.
}
\note{
Updated 2019-10-21.
}
\examples{
isFileSystemCaseSensitive()
}
\seealso{
\verb{_koopa_is_file_system_case_sensitive} shell function in koopa.
}
