% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isFlag.R
\name{check-scalar-isFlag}
\alias{check-scalar-isFlag}
\alias{isFlag}
\title{Does the input contain a boolean flag?}
\usage{
isFlag(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a boolean flag?
}
\note{
Updated 2019-08-10.
}
\examples{
## TRUE ====
isFlag(TRUE)
isFlag(FALSE)

## FALSE ====
isFlag(c(TRUE, TRUE))
isFlag(1)
isFlag(NA)
}
