% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isHexColorFunction.R
\name{check-scalar-isHexColorFunction}
\alias{check-scalar-isHexColorFunction}
\alias{isHexColorFunction}
\title{Does the argument contain a function that returns hexadecimal colors?}
\usage{
isHexColorFunction(x, nullOK = FALSE, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
This assert check is intended primarily to check for RColorBrewer or viridis
hexadecimal color value return.
}
\note{
Updated 2019-07-29.
}
\examples{
## TRUE ====
x <- viridis::viridis
isHexColorFunction(x)

## FALSE ====
x <- ggplot2::scale_colour_manual
isHexColorFunction(x)
}
\seealso{
RColorBrewer, viridis packages.
}
