% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-isLinux.R
\name{check-scalar-isLinux}
\alias{check-scalar-isLinux}
\alias{isLinux}
\title{Is the operating system Linux?}
\usage{
isLinux()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the operating system Linux?
}
\note{
Updated 2020-04-12.
}
\examples{
isLinux()
}
