% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isMacOS.R
\name{check-scalar-isMacOS}
\alias{check-scalar-isMacOS}
\alias{isMacOS}
\title{Is the operating system macOS?}
\usage{
isMacOS()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the operating system macOS?
}
\note{
Updated 2020-04-12.
}
\examples{
isMacOS()
}
