% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isOfDimension.R
\name{check-scalar-isOfDimension}
\alias{check-scalar-isOfDimension}
\title{Does the input contain specific dimensions?}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain specific dimensions?
}
\note{
Updated 2019-10-04.
}
\examples{
x <- data.frame(a = c(TRUE, FALSE), b = c(FALSE, TRUE))
dim(x)

## TRUE ====
isOfDimension(x, n = c(2L, 2L))

## FALSE ====
isOfDimension(x, n = c(1L, 2L))
isOfDimension(x, n = c(2L, 1L))
}
\seealso{
\itemize{
\item \code{assertive.properties::is_of_dimension()}.
}
}
