% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-isRStudio.R
\name{check-scalar-isRStudio}
\alias{check-scalar-isRStudio}
\alias{isRStudio}
\title{Is the current R session running inside RStudio?}
\usage{
isRStudio()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the current R session running inside RStudio?
}
\note{
Updated 2020-04-12.
}
\examples{
isRStudio()
}
