% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isString.R
\name{check-scalar-isString}
\alias{check-scalar-isString}
\alias{isString}
\title{Does the input contain a (non-empty) character string?}
\usage{
isString(x, nullOk = FALSE)
}
\arguments{
\item{x}{Object.}

\item{nullOk}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a (non-empty) character string?
}
\note{
Updated 2022-12-14.
}
\examples{
## TRUE ====
isString("hello")

## FALSE ====
isString(1)
isString("")
isString(NA_character_)
}
