% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isTximport.R
\name{check-scalar-isTximport}
\alias{check-scalar-isTximport}
\alias{isTximport}
\title{Does the input contain a tximport list?}
\usage{
isTximport(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a tximport list?
}
\note{
Updated 2023-07-21.
}
\examples{
## TRUE ====
if (allAreInstalled(c("tximport", "tximportData"))) {
    dir <- system.file("extdata", package = "tximportData")
    samples <- read.table(file.path(dir, "samples.txt"), header = TRUE)
    files <- file.path(dir, "salmon", samples[["run"]], "quant.sf.gz")
    names(files) <- paste0("sample", seq(from = 1L, to = length(files)))
    object <- tximport::tximport(
        files = files,
        type = "salmon",
        txIn = TRUE,
        txOut = TRUE
    )
    isTximport(object)
}

## FALSE ====
object <- list()
isTximport(object)
}
\seealso{
\code{tximport::tximport}.
}
