% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isUnix.R
\name{check-scalar-isUnix}
\alias{check-scalar-isUnix}
\alias{isUnix}
\title{Is the operating system Unix-based?}
\usage{
isUnix()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the operating system Unix-based?
}
\details{
This check will return \code{TRUE} on Linux and macOS but \code{FALSE} on Windows.
}
\note{
Updated 2020-04-07.
}
\examples{
isUnix()
}
