% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isVanilla.R
\name{check-scalar-isVanilla}
\alias{check-scalar-isVanilla}
\alias{isVanilla}
\title{Is the R session running as vanilla?}
\usage{
isVanilla()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the R session running as vanilla?
}
\note{
Updated 2020-08-11.
}
\examples{
isVanilla()
}
