% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isVector.R
\name{check-scalar-isVector}
\alias{check-scalar-isVector}
\alias{isVector}
\title{Does the input contain a (non-empty) vector?}
\usage{
isVector(x, nullOK = FALSE, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a (non-empty) vector?
}
\note{
Updated 2022-12-14.
}
\examples{
## TRUE ====
isVector(c("aaa", "bbb", "ccc"))
isVector(c(1L, 2L, 3L))
isVector(c(TRUE, FALSE))

## FALSE ====
isVector(character())
isVector(matrix())
isVector(data.frame())
isVector("")
isVector(NA)
}
