% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isVectorish.R
\name{check-scalar-isVectorish}
\alias{check-scalar-isVectorish}
\alias{isVectorish}
\title{Is the input vector(ish)?}
\usage{
isVectorish(x, nullOk = FALSE)
}
\arguments{
\item{x}{Object.}

\item{nullOk}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the input vector(ish)?
}
\details{
Intentionally returns \code{TRUE} for some other vector-like classes, including
\code{factor}, and \code{Rle}.
}
\note{
Updated 2022-12-14.
}
\examples{
## TRUE ====
isVectorish(character())
isVectorish(factor())
isVectorish(integer())
isVectorish(logical())
isVectorish(S4Vectors::Rle())

## FALSE ====
isVectorish(data.frame())
isVectorish(matrix())
isVectorish(S4Vectors::DataFrame())
}
