% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isWindows.R
\name{check-scalar-isWindows}
\alias{check-scalar-isWindows}
\alias{isWindows}
\title{Is the operating system Windows?}
\usage{
isWindows()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the operating system Windows?
}
\note{
Updated 2020-04-07.
}
\examples{
isWindows()
}
