% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-validNames.R
\name{check-scalar-validNames}
\alias{check-scalar-validNames}
\alias{validNames}
\title{Are these valid names?}
\usage{
validNames(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Are these valid names?
}
\note{
Updated 2022-05-13.
}
\examples{
## TRUE ====
## Dots (periods) and underscores are valid.
validNames(c("sample.1", "sample_1"))

## FALSE ====
## Can't begin with a number.
validNames("293cells")
## Cannot contain duplicates.
validNames(c("a", "a"))
## Spaces, dashes (hyphens), and other non-alphanumerics aren't valid.
validNames("sample 1")
validNames("cell-AAAAAAAA")
validNames("GFP+")
}
\seealso{
\itemize{
\item \code{make.names()}.
}
}
