% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isAwsS3Uri.R
\name{check-vector-isAwsS3Uri}
\alias{check-vector-isAwsS3Uri}
\alias{isAwsS3Uri}
\alias{isAnAwsS3Uri}
\alias{allAreAwsS3Uris}
\title{Does the input contain an AWS S3 URI?}
\usage{
isAwsS3Uri(x)

isAnAwsS3Uri(x)

allAreAwsS3Uris(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain an AWS S3 URI?
}
\details{
This assert check is intended to be simple and does not check to see if the
URL exists (is active). For that, refer to \code{isExistingUrl} instead.
}
\section{Functions}{
\itemize{
\item \code{isAwsS3Uri()}: Vectorized.

\item \code{isAnAwsS3Uri()}: Scalar. Requires a single URI.

\item \code{allAreAwsS3Uris()}: Scalar. Checks that all strings are URIs.

}}
\note{
Updated 2023-11-03.
}
\examples{
uris <- c("s3://acidgenomics.com/", "s3://koopa.acidgenomics.com/")

## TRUE ====
isAwsS3Uri(uris)
isAnAwsS3Uri(uris[[1L]])
allAreAwsS3Uris(urls)

## FALSE ====
isAwsS3Uri("xxx")
isAnAwsS3Uri(uri)
}
