% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isDirectory.R
\name{check-vector-isDirectory}
\alias{check-vector-isDirectory}
\alias{isDirectory}
\alias{isADirectory}
\alias{allAreDirectories}
\alias{isDir}
\alias{isADir}
\alias{allAreDirs}
\title{Does the input contain a directory?}
\usage{
isDirectory(x)

isADirectory(x, nullOK = FALSE)

allAreDirectories(x)

isDir(x)

isADir(x, nullOK = FALSE)

allAreDirs(x)
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a directory?
}
\section{Functions}{
\itemize{
\item \code{isDirectory}: Vectorized.

\item \code{isADirectory}: Scalar.

\item \code{allAreDirectories}: Scalar.

\item \code{isDir}: Short alias for \code{\link[=isDirectory]{isDirectory()}}.

\item \code{isADir}: Short alias for \code{\link[=isADirectory]{isADirectory()}}.

\item \code{allAreDirs}: Short alias for \code{\link[=allAreDirectories]{allAreDirectories()}}.
}}

\note{
Updated 2019-08-05.
}
\examples{
## TRUE ====
isDirectory(c("~", "~"))
isADirectory("~")

## FALSE ====
isDirectory(1L)
}
\seealso{
\itemize{
\item \code{dir.exists()}.
\item \code{R.utils::isDirectory()}.
}
}
