% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isDuplicate.R
\name{check-vector-isDuplicate}
\alias{check-vector-isDuplicate}
\alias{isDuplicate}
\title{Does the input contain a duplicate value?}
\usage{
isDuplicate(x)
}
\arguments{
\item{x}{\code{vector}.
Any vector type (e.g. \code{character}, \code{logical}, \code{numeric}).}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a duplicate value?
}
\details{
This check is designed to serve as a complement to the base \code{duplicated}
function, which only returns \code{TRUE} after the first observed duplicate value.
Our check here returns \code{TRUE} for all values that are duplicated.
}
\section{Functions}{
\itemize{
\item \code{isDuplicate()}: Vectorized.

}}
\note{
Updated 2023-09-19.
}
\examples{
## TRUE ====
x <- c("a", "a", "b", "b", "c", "d")
isDuplicate(x)
duplicated(x)

## FALSE ====
isDuplicate(c("a", "b", "c"))
}
