% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isEqual.R
\name{check-vector-isEqual}
\alias{check-vector-isEqual}
\alias{isEqualTo}
\alias{isNotEqualTo}
\alias{isGreaterThan}
\alias{isGreaterThanOrEqualTo}
\alias{isLessThan}
\alias{isLessThanOrEqualTo}
\alias{allAreEqualTo}
\alias{allAreNotEqualTo}
\alias{allAreGreaterThan}
\alias{allAreGreaterThanOrEqualTo}
\alias{allAreLessThan}
\alias{allAreLessThanOrEqualTo}
\title{How does the input relate to a value?}
\usage{
isEqualTo(x, y)

isNotEqualTo(x, y)

isGreaterThan(x, y)

isGreaterThanOrEqualTo(x, y)

isLessThan(x, y)

isLessThanOrEqualTo(x, y)

allAreEqualTo(x, y)

allAreNotEqualTo(x, y)

allAreGreaterThan(x, y)

allAreGreaterThanOrEqualTo(x, y)

allAreLessThan(x, y)

allAreLessThanOrEqualTo(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
How does the input relate to a value?
}
\section{Functions}{
\itemize{
\item \code{isEqualTo()}: Vectorized.

\item \code{isNotEqualTo()}: Vectorized.

\item \code{isGreaterThan()}: Vectorized.

\item \code{isGreaterThanOrEqualTo()}: Vectorized.

\item \code{isLessThan()}: Vectorized.

\item \code{isLessThanOrEqualTo()}: Vectorized.

\item \code{allAreEqualTo()}: Scalar.

\item \code{allAreNotEqualTo()}: Scalar.

\item \code{allAreGreaterThan()}: Scalar.

\item \code{allAreGreaterThanOrEqualTo()}: Scalar.

\item \code{allAreLessThan()}: Scalar.

\item \code{allAreLessThanOrEqualTo()}: Scalar.

}}
\note{
Updated 2023-10-02.
}
\examples{
## TRUE ====
isEqualTo(x = 1L, y = 1)
isNotEqualTo(x = 2, y = 1)
isGreaterThan(x = 1, y = 0)
isGreaterThanOrEqualTo(x = seq_len(2), y = 1)
isLessThan(x = -1, y = 0)
isLessThanOrEqualTo(x = seq_len(2), y = 3)

## FALSE ====
isEqualTo(x = seq_len(2), y = 1)
}
\seealso{
\itemize{
\item Primitive operators: \code{==}, \code{>}, \code{>=}, \code{<}, \code{<=}.
\item \code{assertive.numbers::is_equal_to()}.
\item \code{assertive.numbers::is_not_equal_to()}.
\item \code{assertive.numbers::is_greater_than()}.
\item \code{assertive.numbers::is_greater_than_or_equal_to()}.
\item \code{assertive.numbers::is_less_than()}.
\item \code{assertive.numbers::is_less_than_or_equal_to()}.
}
}
