% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isExistingAwsS3Uri.R
\name{check-vector-isExistingAwsS3Uri}
\alias{check-vector-isExistingAwsS3Uri}
\alias{isExistingAwsS3Uri}
\alias{isAnExistingAwsS3Uri}
\alias{allAreExistingAwsS3Uris}
\title{Does the input contain an existing (active) AWS S3 URI?}
\usage{
isExistingAwsS3Uri(x, profile = "default")

isAnExistingAwsS3Uri(x, profile = "default")

allAreExistingAwsS3Uris(x, profile = "default")
}
\arguments{
\item{x}{Object.}

\item{profile}{\code{character(1)}.
AWS profile name.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain an existing (active) AWS S3 URI?
}
\details{
Requires the AWS CLI to be installed.
}
\section{Functions}{
\itemize{
\item \code{isExistingAwsS3Uri()}: Vectorized.

\item \code{isAnExistingAwsS3Uri()}: Scalar. Requires a single URI.

\item \code{allAreExistingAwsS3Uris()}: Scalar. Checks that all strings
are existing URis.

}}
\note{
Updated 2023-11-03.
}
\examples{
## TRUE ====
## > isAnExistingAwsS3Uri(
## >     x = "s3://koopa.acidgenomics.com/install",
## >     profile = "acidgenomics"
## > )

## FALSE ====
## > isAnExistingAwsS3Uri(
## >     x = "s3://koopa.acidgenomics.com/foo",
## >     profile = "acidgenomics"
## > )
}
\seealso{
\itemize{
\item \verb{aws help}.
\item https://www.learnaws.org/2023/01/30/aws-s3-cli-check-file/
\item koopa function \code{koopa_is_existing_aws_s3_uri}.
}
}
