% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isFile.R,
%   R/check-vector-isSymlink.R
\name{check-vector-isFile}
\alias{check-vector-isFile}
\alias{isFile}
\alias{isAFile}
\alias{allAreFiles}
\title{Does the input contain a file?}
\usage{
isFile(x)

isAFile(x, nullOK = FALSE)

allAreFiles(x)
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a file?

Does the input contain a symbolic link?
}
\section{Functions}{
\itemize{
\item \code{isFile}: Vectorized.

\item \code{isAFile}: Scalar.

\item \code{allAreFiles}: Scalar.
}}

\note{
Updated 2020-06-26.

Updated 2020-06-26.
}
\examples{
## TRUE ====
x <- "example.txt"
file.create(x)
isAFile(x)
unlink(x)

## FALSE ====
isFile(c("~", "."))
from <- "from.txt"
to <- "to.txt"
file.create(from)
file.symlink(from = from, to = to)
isSymlink(c(from, to))
unlink(c(from, to))
}
\seealso{
\itemize{
\item \code{file.exists()}.
\item \code{R.utils::isFile()}.
}

\itemize{
\item \code{file.info()}.
\item \code{Sys.readlink()}.
}
}
