% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isGitRepo.R
\name{check-vector-isGitRepo}
\alias{check-vector-isGitRepo}
\alias{isGitRepo}
\alias{isAGitRepo}
\alias{allAreGitRepos}
\title{Does the input contain a Git repository?}
\usage{
isGitRepo(x)

isAGitRepo(x)

allAreGitRepos(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a Git repository?
}
\section{Functions}{
\itemize{
\item \code{isGitRepo}: Vectorized.

\item \code{isAGitRepo}: Scalar.

\item \code{allAreGitRepos}: Scalar.
}}

\note{
Updated 2022-05-13.
}
\examples{
## TRUE ====
isGitRepo(file.path("~", "git", "monorepo"))

## FALSE ====
isGitRepo("~")
}
