% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isMatching.R
\name{check-vector-isMatching}
\alias{check-vector-isMatching}
\alias{isMatchingFixed}
\alias{isMatchingRegex}
\alias{isNotMatchingFixed}
\alias{isNotMatchingRegex}
\alias{allAreMatchingFixed}
\alias{allAreMatchingRegex}
\alias{allAreNotMatchingFixed}
\alias{allAreNotMatchingRegex}
\title{Does the string match a pattern?}
\usage{
isMatchingFixed(x, pattern, .xname = getNameInParent(x))

isMatchingRegex(x, pattern, .xname = getNameInParent(x))

isNotMatchingFixed(x, pattern, .xname = getNameInParent(x))

isNotMatchingRegex(x, pattern, .xname = getNameInParent(x))

allAreMatchingFixed(x, pattern, .xname = getNameInParent(x))

allAreMatchingRegex(x, pattern, .xname = getNameInParent(x))

allAreNotMatchingFixed(x, pattern, .xname = getNameInParent(x))

allAreNotMatchingRegex(x, pattern, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{logical}.
}
\description{
Does the string match a pattern?
}
\section{Functions}{
\itemize{
\item \code{isMatchingFixed}: Vectorized.

\item \code{isMatchingRegex}: Vectorized.

\item \code{isNotMatchingFixed}: Vectorized.

\item \code{isNotMatchingRegex}: Vectorized.

\item \code{allAreMatchingFixed}: Scalar.

\item \code{allAreMatchingRegex}: Scalar.

\item \code{allAreNotMatchingFixed}: Scalar.

\item \code{allAreNotMatchingRegex}: Scalar.
}}

\note{
Updated 2019-08-10.
}
\examples{
## TRUE ====
isMatchingRegex(x = "foobar", pattern = "^f")
isNotMatchingRegex(x = "foobar", pattern = "^b")

isMatchingFixed(x = "foobar", pattern = "bar")
isNotMatchingFixed(x = "foo", pattern = "bar")
}
\seealso{
\itemize{
\item \code{grepl()}.
\item \code{stringi::stri_detect_regex()}.
\item \code{stringi::stri_detect_fixed()}.
\item \code{stringr::str_detect()}.
\item \code{assertive.strings::is_matching_regex()}.
\item \code{assertive.strings::is_matching_fixed()}.
\item \code{assertive.strings::is_not_matching_regex()}.
\item \code{assertive.strings::is_not_matching_fixed()}.
}
}
