% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-system-vector-isPackageVersion.R
\name{check-vector-isPackageVersion}
\alias{check-vector-isPackageVersion}
\alias{isPackageVersion}
\title{Is the package installed and a specific version?}
\usage{
isPackageVersion(x, op = ">=")
}
\arguments{
\item{x}{\code{character}.
Named character vector.
Name corresponds to package name, and value corresponds to minimum version.}

\item{op}{\code{character(1)}.
Mathematical operator.
Defaults to less than or equal to.}
}
\description{
Is the package installed and a specific version?
}
\section{Functions}{
\itemize{
\item \code{isPackageVersion()}: Vectorized.

}}
\note{
Updated 2022-05-13.
}
\examples{
## TRUE ====
isPackageVersion(
    x = c(
        "base" = packageVersion("base"),
        "utils" = packageVersion("utils")
    ),
    op = "=="
)

## FALSE ====
isPackageVersion(
    x = c(
        "base" = packageVersion("base"),
        "utils" = packageVersion("utils")
    ),
    op = ">"
)
}
