% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isSymlink.R
\name{check-vector-isSymlink}
\alias{check-vector-isSymlink}
\alias{isSymlink}
\alias{isASymlink}
\alias{allAreSymlinks}
\title{Does the input contain a symbolic link?}
\usage{
isSymlink(x)

isASymlink(x, nullOk = FALSE)

allAreSymlinks(x)
}
\arguments{
\item{x}{Object.}

\item{nullOk}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a symbolic link?
}
\details{
Supported on Linux and macOS but not Windows.
}
\section{Functions}{
\itemize{
\item \code{isSymlink()}: Vectorized.

\item \code{isASymlink()}: Scalar.

\item \code{allAreSymlinks()}: Scalar.

}}
\note{
Updated 2023-10-02.
}
\examples{
if (!isWindows()) {
    from <- "from.txt"
    to <- "to.txt"
    file.create(from)
    file.symlink(from = from, to = to)
    isSymlink(c(from, to))
    unlink(c(from, to))
}
}
\seealso{
\itemize{
\item \code{file.info()}.
\item \code{Sys.readlink()}.
}
}
