% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isTempFile.R
\name{check-vector-isTempFile}
\alias{check-vector-isTempFile}
\alias{isTempFile}
\alias{isATempFile}
\alias{allAreTempFiles}
\title{Does the input contain a temporary file?}
\usage{
isTempFile(x)

isATempFile(x, nullOK = FALSE)

allAreTempFiles(x)
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a temporary file?
}
\section{Functions}{
\itemize{
\item \code{isTempFile()}: Vectorized.

\item \code{isATempFile()}: Scalar.

\item \code{allAreTempFiles()}: Scalar.

}}
\note{
Updated 2023-07-13.
}
\examples{
## TRUE ====
x <- "example.txt"
file.create(x)
isAFile(x)
unlink(x)

## FALSE ====
isFile(c("~", "."))
}
\seealso{
\itemize{
\item \code{tempfile()}.
}
}
