% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isUrl.R
\name{check-vector-isUrl}
\alias{check-vector-isUrl}
\alias{isUrl}
\alias{isAUrl}
\alias{allAreUrls}
\title{Does the input contain a URL?}
\usage{
isUrl(x)

isAUrl(x)

allAreUrls(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a URL?
}
\details{
This assert check is intended to be simple and does not check to see if the
URL exists (is active). For that, refer to \code{isExistingUrl} instead.
}
\section{Functions}{
\itemize{
\item \code{isUrl()}: Vectorized.

\item \code{isAUrl()}: Scalar. Requires a single URL.

\item \code{allAreUrls()}: Scalar. Checks that all strings are URLs.

}}
\note{
Updated 2023-10-06.
}
\examples{
urls <- c("https://www.r-project.org/", "ftp://r-project.org/")

## TRUE ====
isUrl(urls)
isAUrl(urls[[1L]])
allAreUrls(urls)

## FALSE ====
isUrl("xxx")
isAUrl(urls)
}
