% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAllAreAtomic.R
\name{checkAllAreAtomic}
\alias{checkAllAreAtomic}
\alias{allAreAtomic}
\alias{check_all_are_atomic}
\alias{testAllAreAtomic}
\alias{test_all_are_atomic}
\alias{assertAllAreAtomic}
\alias{assert_all_are_atomic}
\alias{expect_all_are_atomic}
\title{Check If an Object Contain Elements That Are All Atomic}
\usage{
checkAllAreAtomic(x)

check_all_are_atomic(x)

testAllAreAtomic(x)

test_all_are_atomic(x)

assertAllAreAtomic(x, .var.name = vname(x), add = NULL)

assert_all_are_atomic(x, .var.name = vname(x), add = NULL)

expect_all_are_atomic(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Check If an Object Contain Elements That Are All Atomic
}
\examples{
checkAllAreAtomic(datasets::mtcars)
}
\seealso{
\itemize{
\item \code{checkmate::makeTestFunction()}.
\item \code{checkmate::makeAssertionFunction()}.
\item \code{checkmate::makeExpectationFunction()}.
}
}
