% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnyClass.R
\name{checkAnyClass}
\alias{checkAnyClass}
\alias{anyClass}
\alias{check_any_class}
\alias{testAnyClass}
\alias{test_any_class}
\alias{assertAnyClass}
\alias{assert_any_class}
\alias{expect_any_class}
\title{Does an Object Belong to Any of These Classes?}
\usage{
checkAnyClass(x, classes)

check_any_class(x, classes)

testAnyClass(x, classes)

test_any_class(x, classes)

assertAnyClass(x, classes, .var.name = checkmate::vname(x), add = NULL)

assert_any_class(x, classes, .var.name = checkmate::vname(x),
  add = NULL)

expect_any_class(x, classes, info = NULL, label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{classes}{\code{character}. Object classes.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}
}
\description{
Does an Object Belong to Any of These Classes?
}
\examples{
# Define an integer.
x <- 1L

testClass(x, classes = c("integer", "numeric", "character", "atomic"))

# This should pass, but fails...very strict.
checkClass(x, "numeric")
checkClass(x, "character")


}
\seealso{
\itemize{
\item \code{checkmate::checkClass()}.
\item \code{checkmate::checkMultiClass()}.
\item \code{assertive::is2()}.
\item \code{assertive::assert_is_any_of()}.
}
}
