% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAreNonExisting.R
\name{checkAreNonExisting}
\alias{checkAreNonExisting}
\alias{areNonExisting}
\alias{check_are_non_existing}
\alias{testAreNonExisting}
\alias{test_are_non_existing}
\alias{assertAreNonExisting}
\alias{assert_are_non_existing}
\alias{expect_are_non_existing}
\title{Does the Requested Variable Exist?}
\usage{
checkAreNonExisting(x)

check_are_non_existing(x)

testAreNonExisting(x)

test_are_non_existing(x)

assertAreNonExisting(x, .var.name = checkmate::vname(x), add = NULL)

assert_are_non_existing(x, .var.name = checkmate::vname(x), add = NULL)

expect_are_non_existing(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{\code{character}. Variable names to check in \code{environment}.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Requested Variable Exist?
}
\examples{
checkAreNonExisting(c("XXX", "YYY"))
}
\seealso{
\itemize{
\item \code{checkmate::makeTestFunction()}.
\item \code{checkmate::makeAssertionFunction()}.
\item \code{checkmate::makeExpectationFunction()}.
}
}
