% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHasInternet.R
\name{checkHasInternet}
\alias{checkHasInternet}
\alias{hasInternet}
\alias{has_internet}
\alias{check_has_internet}
\alias{testHasInternet}
\alias{test_has_internet}
\alias{assertHasInternet}
\alias{assert_has_internet}
\title{Does the Current Session Have an Internet Connection?}
\usage{
checkHasInternet()

check_has_internet()

testHasInternet()

test_has_internet()

assertHasInternet()

assert_has_internet()
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Current Session Have an Internet Connection?
}
\examples{
checkHasInternet()
}
