% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHasValidDimnames.R
\name{checkHasValidDimnames}
\alias{checkHasValidDimnames}
\alias{hasValidDimnames}
\alias{has_valid_dimnames}
\alias{check_has_valid_dimnames}
\alias{testHasValidDimnames}
\alias{test_has_valid_dimnames}
\alias{assertHasValidDimnames}
\alias{assert_has_valid_dimnames}
\alias{expect_has_valid_dimnames}
\title{Does the Object Have Syntactically Valid Dimnames?}
\usage{
checkHasValidDimnames(x)

check_has_valid_dimnames(x)

testHasValidDimnames(x)

test_has_valid_dimnames(x)

assertHasValidDimnames(x, .var.name = checkmate::vname(x), add = NULL)

assert_has_valid_dimnames(x, .var.name = checkmate::vname(x),
  add = NULL)

expect_has_valid_dimnames(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}

\item{label}{XXX}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Object Have Syntactically Valid Dimnames?
}
\examples{
## Pass ====
x <- datasets::iris
lapply(dimnames(x), head)
checkHasValidDimnames(x)

## Fail ====
x <- datasets::mtcars
# Note the spaces in the row names here.
lapply(dimnames(x), head)
checkHasValidDimnames(x)
}
