% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHexColorFunction.R
\name{checkHexColorFunction}
\alias{checkHexColorFunction}
\alias{hexColorFunction}
\alias{hex_color_function}
\title{Does the Argument Contain a Hex Color Function?}
\usage{
checkHexColorFunction(x, null.ok = FALSE)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Argument Contain a Hex Color Function?
}
\examples{
x <- function(n) {
    colors <- c("#FFFFFF", "#000000")
    colors[seq_len(n)]
}
checkHexColorFunction(x)
}
