% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIdentical.R
\name{checkIdentical}
\alias{checkIdentical}
\alias{identical}
\alias{check_identical}
\alias{testIdentical}
\alias{test_identical}
\alias{assertIdentical}
\alias{assert_identical}
\title{Are Two Objects Identical?}
\usage{
checkIdentical(x, y)

check_identical(x, y)

testIdentical(x, y)

test_identical(x, y)

assertIdentical(x, y, .var.name = checkmate::vname(x), add = NULL)

assert_identical(x, y, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{Object.}

\item{y}{Secondary object.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Are Two Objects Identical?
}
\examples{
## Pass ====
checkIdentical(x = 1L, y = 1L)

## Fail ====
checkIdentical(x = 1L, y = 1)
}
\seealso{
\code{identical()}.
}
