% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNonEmpty.R
\name{checkNonEmpty}
\alias{checkNonEmpty}
\alias{nonEmpty}
\alias{non_empty}
\alias{check_non_empty}
\alias{testNonEmpty}
\alias{test_non_empty}
\alias{assertNonEmpty}
\alias{assert_non_empty}
\alias{expect_non_empty}
\title{Does an Object Contain Any Information?}
\usage{
checkNonEmpty(x)

check_non_empty(x)

testNonEmpty(x)

test_non_empty(x)

assertNonEmpty(x, .var.name = checkmate::vname(x), add = NULL)

assert_non_empty(x, .var.name = checkmate::vname(x), add = NULL)

expect_non_empty(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{.var.name}{\code{character(1)}. Name of the checked object to print in
assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}

\item{label}{\code{character(1)}. Name of the checked object to print in messages.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does an Object Contain Any Information?
}
\examples{
## Pass ====
checkNonEmpty(c("foo", "bar"))
checkNonEmpty(matrix(data = seq_len(100L), nrow = 10L))

## Fail ====
checkNonEmpty(NULL)
checkNonEmpty(character())
checkNonEmpty(matrix())
}
