% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkScale.R
\name{checkScale}
\alias{checkScale}
\alias{scale}
\alias{check_scale}
\alias{testScale}
\alias{test_scale}
\alias{assertScale}
\alias{assert_scale}
\alias{expect_scale}
\title{Does the Argument Contain a ggplot2 Scale?}
\usage{
checkScale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE)

check_scale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE)

testScale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE)

test_scale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE)

assertScale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE, .var.name = checkmate::vname(x),
  add = NULL)

assert_scale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE, .var.name = checkmate::vname(x),
  add = NULL)

expect_scale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), null.ok = FALSE, info = NULL, label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{scale}{\code{string}. Type of scale, either \code{"continuous"} or \code{"discrete"}.}

\item{aes}{\code{string}. Aesthetic mapping, either \code{"colour"} or \code{"fill"}. Note
that ggplot2 prefers British spelling, so we're enforcing that convention
here.}

\item{null.ok}{\code{logical(1)}. If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{.var.name}{\code{character(1)}. Name of the checked object to print in
assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}

\item{label}{\code{character(1)}. Name of the checked object to print in messages.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Argument Contain a ggplot2 Scale?
}
\examples{
library(ggplot2)

colour_c <- scale_colour_gradient(low = "red", high = "blue")
class(colour_c)

colour_d <- scale_colour_manual(values = c("red", "blue"))
class(colour_d)

fill_c <- scale_fill_gradient(low = "red", high = "blue")
class(fill_c)

fill_d <- scale_fill_manual(values = c("red", "blue"))
class(fill_d)

checkScale(x = colour_c, scale = "continuous", aes = "colour")
checkScale(x = colour_d, scale = "discrete", aes = "colour")
checkScale(x = fill_c, scale = "continuous", aes = "fill")
checkScale(x = fill_d, scale = "discrete", aes = "fill")
}
