% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSubset.R
\name{checkSubset}
\alias{checkSubset}
\alias{check_subset}
\alias{testSubset}
\alias{test_subset}
\alias{assertSubset}
\alias{assert_subset}
\alias{expect_subset}
\title{Does an Argument Match Against a Secondary Argument?}
\usage{
checkSubset(x, y)

check_subset(x, y)

testSubset(x, y)

test_subset(x, y)

assertSubset(x, y, .var.name = checkmate::vname(x), add = NULL)

assert_subset(x, y, .var.name = checkmate::vname(x), add = NULL)

expect_subset(x, y, info = NULL, label = vname(x))
}
\description{
Does an Argument Match Against a Secondary Argument?
}
\examples{
checkSubset(x = "foo", y = c("foo", "bar"))
}
\seealso{
\itemize{
\item \code{match()}, \code{\%in\%}.
\item \code{checkNames()}.
}
}
