% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkURL.R
\name{checkURL}
\alias{checkURL}
\alias{url}
\alias{check_url}
\alias{testURL}
\alias{test_url}
\alias{assertURL}
\alias{assert_url}
\alias{expect_url}
\alias{checkURLs}
\alias{check_urls}
\alias{testURLs}
\alias{test_urls}
\alias{assertAreURLs}
\alias{assert_are_urls}
\alias{expect_are_urls}
\title{Does the Argument Contain a URL?}
\usage{
checkURL(x)

check_url(x)

testURL(x)

test_url(x)

assertURL(x, .var.name = checkmate::vname(x), add = NULL)

assert_url(x, .var.name = checkmate::vname(x), add = NULL)

expect_url(x, info = NULL, label = vname(x))

checkURLs(x)

check_urls(x)

testURLs(x)

test_urls(x)

assertAreURLs(x, .var.name = checkmate::vname(x), add = NULL)

assert_are_urls(x, .var.name = checkmate::vname(x), add = NULL)

expect_are_urls(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Argument Contain a URL?
}
\examples{
urls <- c("https://www.r-project.org", "ftp://r-project.org")

## Single URL.
checkURL(urls[[1L]])
checkURL(urls)  # fail

## Multiple URLs.
checkURLs(urls)
checkURLs(urls[[1L]])  # fail
}
\seealso{
\itemize{
\item \code{checkmate::makeTestFunction()}.
\item \code{checkmate::makeAssertionFunction()}.
\item \code{checkmate::makeExpectationFunction()}.
}
}
