% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-containsAlpha.R
\name{containsAlpha}
\alias{containsAlpha}
\alias{isAlpha}
\title{Does the input contain an alpha level?}
\usage{
containsAlpha(x, .xname = getNameInParent(x))

isAlpha(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
An alpha level must be \code{numeric(1)} greater than 0 and less than 1.
}
\examples{
## Pass ====
containsAlpha(0.05)

## Fail ====
containsAlpha("xxx")
containsAlpha(1L)
}
