% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-containsHexColors.R
\name{containsHexColors}
\alias{containsHexColors}
\alias{areHexColors}
\title{Does the input contain hexadecimal colors?}
\usage{
containsHexColors(x, .xname = getNameInParent(x))

areHexColors(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input contain hexadecimal colors?
}
\examples{
## Pass ====
x <- viridis::viridis(n = 2L)
class(x)
print(x)
containsHexColors(x)

## Fail ====
x <- ggplot2::scale_colour_manual
class(x)
containsHexColors(x)
}
\seealso{
\code{assertive::is_hex_color()}.
}
