% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containsURL.R
\name{containsURL}
\alias{containsURL}
\alias{containsAURL}
\title{Does the Input Contain a URL?}
\usage{
containsURL(x)

containsAURL(x)
}
\arguments{
\item{x}{Object.}

\item{string}{\code{logical(1)}. Require match against string?}
}
\value{
\code{logical}.
}
\description{
Does the Input Contain a URL?
}
\section{Functions}{
\itemize{
\item \code{containsURL}: Supports multiple URLs.

\item \code{containsAURL}: Requires a single URL.
}}

\examples{
urls <- c("https://www.r-project.org", "ftp://r-project.org")

## Pass ====
containsURL(urls)
containsAURL(urls[[1L]])

## Fail ====
containsURL("xxx")
containsAURL(urls)
}
