% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-bapply.R
\name{engine-bapply}
\alias{engine-bapply}
\alias{bapply}
\title{Boolean apply}
\usage{
bapply(X, FUN, ..., USE.NAMES = TRUE)
}
\arguments{
\item{X}{\code{atomic} or \code{list}.}

\item{FUN}{\code{function}.
An assert check function that returns \code{logical(1)} boolean flag
(\code{TRUE}/\code{FALSE}) to \link[base:apply]{apply}.}

\item{...}{Additional arguments passed to \link[base:lapply]{vapply()}.}

\item{USE.NAMES}{\code{logical(1)}.
If \code{TRUE} and \code{X} is character, use \code{X} as \code{\link[base:names]{names}} for the
result, unless it has names already.}
}
\value{
\code{logical}.
}
\description{
\code{\link[=bapply]{bapply()}} is a wrapper for \code{\link[base:lapply]{vapply()}} that enforces
\code{FUN.VALUE = logical(1)}.
}
\note{
\itemize{
\item \code{FUN.VALUE} is always set to \code{logical(1)}.
\item \code{USE.NAMES} is always set to \code{TRUE}.
}

Updated 2021-02-23.
}
\examples{
bapply(X = list(a = "example", b = 1), FUN = is.character)
}
\seealso{
\itemize{
\item \code{assertive.base::bapply()}.
\item \code{vapply()}.
}
}
