% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-safeDeparse.R
\name{engine-safeDeparse}
\alias{engine-safeDeparse}
\alias{safeDeparse}
\title{Expression deparsing}
\usage{
safeDeparse(expr, ...)
}
\arguments{
\item{expr}{\code{expression}.
Any R expression.}

\item{...}{Passed to \code{\link[base:deparse]{deparse()}}.}
}
\value{
\code{character(1)}.
}
\description{
Turn unevaluated expressions into character strings.
}
\details{
\code{\link[=safeDeparse]{safeDeparse()}} is modified version of \code{\link[base:deparse]{deparse()}} that
always returns \code{character(1)}.
}
\note{
Updated 2020-01-04.
}
\examples{
safeDeparse(is.character("a"))
}
\seealso{
\itemize{
\item \code{assertive.base::safe_deparse()}.
\item \code{deparse()}.
}
}
