% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-shorten.R
\name{engine-shorten}
\alias{engine-shorten}
\alias{shorten}
\title{Shorten (truncate) input to a specified width}
\usage{
shorten(x, width = getOption("width"))
}
\arguments{
\item{x}{\code{atomic}.}
}
\value{
\code{character}.
}
\description{
Automatically generate a substring and add trailing "..." if necessary.
}
\details{
Note that return matches the desired width.
}
\note{
Updated 2020-01-04.
}
\examples{
x <- "the quick brown fox"
x <- shorten(x, width = 10L)
nchar(x)
}
\seealso{
\code{assertive.base:::truncate()}.
}
