% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existing.R
\name{existing}
\alias{existing}
\alias{isExisting}
\alias{areNonExisting}
\title{Does the Requested Input Exist in the Environment?}
\usage{
isExisting(x, envir = parent.frame(), inherits = TRUE,
  .xname = get_name_in_parent(x))

areNonExisting(x, envir = parent.frame(), inherits = TRUE)
}
\arguments{
\item{x}{\code{character}. Variable names to check in \code{environment}.}

\item{envir}{\code{environment}. Environment.}

\item{inherits}{\code{logical(1)}. Should the enclosing frames of the
\code{environment} be searched?}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Requested Input Exist in the Environment?
}
\note{
\code{exists()} only supports \code{character(1)}, so we are exporting
\code{isExisting()} as a convenience function to check multiple variables in a
single call.
}
\examples{
suppressWarnings(rm(x, y))
a <- 1L
b <- 2L

## Pass ====
isExisting(c("a", "b"))
areNonExisting(c("x", "y"))

## Fail ====
isExisting(c("x", "y"))
areNonExisting(c("a", "b"))
}
