% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\name{extPattern}
\alias{extPattern}
\alias{compressExtPattern}
\title{File extension pattern matching}
\usage{
extPattern

compressExtPattern
}
\description{
File extension pattern matching
}
\note{
Updated 2023-05-31.
}
\examples{
extPattern
compressExtPattern
}
\keyword{internal}
