% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false.R
\name{false}
\alias{false}
\title{Set a false goalie check with cause attribute}
\usage{
false(...)
}
\arguments{
\item{...}{Elements to pass to internal \code{\link[base:sprintf]{sprintf()}} call.}
}
\value{
\code{goalie}.
}
\description{
Set a false goalie check with cause attribute
}
\note{
Updated 2023-09-29.
}
\examples{
x <- false("{.var \%s} is invalid.", "xxx")
print(x)
print(cause(x))
}
