% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falseFromVector-methods.R
\name{falseFromVector}
\alias{falseFromVector}
\title{Set a scalar false goalie check with cause attribute}
\value{
\code{goalie}.
}
\description{
Set a scalar false goalie check with cause attribute
}
\note{
Updated 2021-02-23.
}
\examples{
x <- goalie(
    object = c("aaa" = FALSE, "bbb" = TRUE, "ccc" = FALSE),
    cause = c("xxx", NA, "yyy")
)
print(x)
print(cause(x))
xx <- falseFromVector(x)
print(xx)
print(cause(xx))
}
