% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-formalCompress.R
\name{formalCompress}
\alias{formalCompress}
\title{Check the \code{compress} formal argument}
\usage{
formalCompress(compress)
}
\arguments{
\item{compress}{\code{logical(1)} or \code{character(1)}.
These character strings are currently allowed for \code{save()}:
\code{"gzip"}, \code{"bzip2"}, or \code{"xz"}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Check the \code{compress} formal argument
}
\examples{
## Pass ====
formalCompress("gzip")
formalCompress(TRUE)

## Fail ====
formalCompress(NA)
formalCompress("xxx")
}
