% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNameInParent.R
\name{getNameInParent}
\alias{getNameInParent}
\alias{vname}
\title{Get the name of a variable defined in the parent frame}
\usage{
getNameInParent(x, escapePercent = TRUE)

vname(x, escapePercent = TRUE)
}
\arguments{
\item{x}{\code{symbol}.
Variable to get the name of.}

\item{escapePercent}{\code{logical(1)}.
If \code{TRUE}, percent signs are doubled, making the value suitable for use
with \link[base:sprintf]{sprintf()}.}
}
\description{
Get the name of a variable defined in the parent frame
}
\examples{
getNameInParent(test)
}
\seealso{
\itemize{
\item \code{assertive.base::get_name_in_parent()}.
\item \code{checkmate::vname()}.
}
}
