% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{goalie-class}
\alias{goalie-class}
\title{goalie check}
\value{
\code{goalie}.
}
\description{
goalie check
}
\details{
Contains a \code{logical} with \code{cause} of \code{character(1)} if any elements are
\code{FALSE}.
}
\note{
Updated 2021-02-23.
}
