% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{goalie}
\alias{goalie}
\title{Define a goalie check classed return}
\usage{
goalie(object, cause)
}
\arguments{
\item{object}{\code{logical}.}

\item{cause}{\code{character}.
Only applies when \code{object} vector contains any \code{FALSE} values.
The \code{cause} attribute must be named.
Cause values corresponding to \code{TRUE} must be set \code{NA}.}
}
\value{
All \code{TRUE}: \code{logical}.
Any \code{FALSE}: \code{goalie}, which extends \code{logical}.
}
\description{
Define a goalie check classed return
}
\note{
Updated 2023-09-29.
}
\examples{
## Any FALSE.
x <- goalie(
    object = c(FALSE, TRUE),
    cause = c(
        "1" = "failed check",
        "2" = NA_character_
    )
)
print(x)
print(cause(x))

## All TRUE.
x <- goalie(rep(TRUE, 2L))
print(x)
}
