% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-hasAccess.R
\name{hasAccess}
\alias{hasAccess}
\title{Check file system access rights}
\usage{
hasAccess(x, access = "r")
}
\arguments{
\item{x}{\code{character(1)}.
File or directory path.}

\item{access}{\code{character(1)}.
String containing any of these characters, including in combination:
\itemize{
\item \code{r} read.
\item \code{w}: write.
\item \code{x}: execute.
}

Write and executable cannot be checked on Windows.}
}
\description{
Works for either file or directory paths.
}
\examples{
hasAccess("~")
}
\seealso{
\code{checkmate::checkAccess()}.
}
