% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-hasDimnames.R
\name{hasDimnames}
\alias{hasDimnames}
\title{Does the input have dimnames?}
\usage{
hasDimnames(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input have dimnames?
}
\examples{
## Pass ====
x <- datasets::mtcars
hasDimnames(x)
hasRownames(x)
hasColnames(x)

## Fail ====
x <- data.frame()
hasDimnames(x)
hasRownames(x)
hasColnames(x)
}
\seealso{
\itemize{
\item \code{hasRownames()}.
\item \code{assertive.properties::has_dimnames()}.
\item \code{assertive.properties::has_colnames()}.
}
}
