% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-hasDuplicates.R
\name{hasDuplicates}
\alias{hasDuplicates}
\alias{hasNoDuplicates}
\title{Does the input have duplicates?}
\usage{
hasDuplicates(x, .xname = getNameInParent(x))

hasNoDuplicates(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input have duplicates?
}
\examples{
## Pass ====
hasDuplicates(c("a", "a"))
hasNoDuplicates(c("a", "b"))

## Fail ====
hasDuplicates(c("a", "b"))
hasNoDuplicates(c("a", "a", "b", "b"))
}
\seealso{
\itemize{
\item \code{assertive.properties::has_duplicates()}.
\item \code{assertive.properties::has_no_duplicates()}.
}
}
