% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-hasInternet.R
\name{hasInternet}
\alias{hasInternet}
\title{Does the Current Session Have an Internet Connection?}
\usage{
hasInternet()
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
Does the Current Session Have an Internet Connection?
}
\examples{
hasInternet()
}
