% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasName.R
\name{hasName}
\alias{hasName}
\title{Does an Object Have an Element with This Name?}
\usage{
hasName(x, name)
}
\arguments{
\item{x}{Object.}

\item{name}{\code{string}. Name.}
}
\description{
Does an Object Have an Element with This Name?
}
