% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-flag-hasNames.R
\name{hasNames}
\alias{hasNames}
\title{Does the input have names?}
\usage{
hasNames(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input have names?
}
\examples{
## Pass ====
hasNames(datasets::mtcars)

## Fail ====
hasNames(matrix())
hasNames(data.frame())
}
\seealso{
\code{assertive.properties::has_names()}.
}
