% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasNonZeroRowsAndCols.R
\name{hasNonZeroRowsAndCols}
\alias{hasNonZeroRowsAndCols}
\title{Does the Input Contain Non-Zero Rows and Columns?}
\usage{
hasNonZeroRowsAndCols(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop}) on failure.
}
}
\description{
Useful for quickly checking to see if we have dropped rows or columns
containing all zeros.
}
\details{
This is a common check when handling RNA-seq data prior to generating a
heatmap or applying a log transformation, for example.

Currently requires a \code{matrix} or \code{sparseMatrix} as input.
}
\examples{
## Pass ====
x <- matrix(data = seq_len(4), nrow = 2)
print(x)
hasNonZeroRowsAndCols(x)

## Fail ====
x <- matrix(nrow = 0, ncol = 0)
print(x)
hasNonZeroRowsAndCols(x)

x <- matrix(data = rep(0L, times = 2), byrow = FALSE)
print(x)
hasNonZeroRowsAndCols(x)

x <- matrix(data = rep(0L, times = 2), byrow = TRUE)
print(x)
hasNonZeroRowsAndCols(x)
}
