% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasRownames.R
\name{hasRownames}
\alias{hasRownames}
\alias{has_rownames}
\alias{assertHasRownames}
\alias{assert_has_rownames}
\title{Does the Object Have Rownames?}
\usage{
hasRownames(x)

has_rownames(x)

assertHasRownames(x, .var.name = vname(x), add = NULL)

assert_has_rownames(x, .var.name = vname(x), add = NULL)
}
\arguments{
\item{x}{Object.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
A stricter alternative to the assertive version that works properly with
data frames.
}
\examples{
x <- S4Vectors::DataFrame(
    "sample1" = c(1L, 2L),
    "sample2" = c(3L, 4L),
    row.names = c("gene1", "gene2")
)
print(x)
hasRownames(x)
}
\seealso{
\itemize{
\item \code{checkmate::makeAssertionFunction()}.
}
}
