% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-hasUMAP.R
\name{hasUMAP}
\alias{hasUMAP}
\title{Is the Python UMAP module available in the current session?}
\usage{
hasUMAP()
}
\description{
Is the Python UMAP module available in the current session?
}
\examples{
hasUMAP()
}
\seealso{
\itemize{
\item \href{https://cran.r-project.org/package=reticulate}{reticulate}
\item \href{https://github.com/lmcinnes/umap}{umap-learn}
}
}
