% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasUniqueCols.R
\name{hasUniqueCols}
\alias{hasUniqueCols}
\title{Does the Input Have Columns with Unique Values?}
\usage{
hasUniqueCols(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop}) on failure.
}
}
\description{
Checks a matrix for duplicated columns, which reprent samples. Duplicate
rows are allowed here, because many genes (rows) can contain all zeros.
}
\section{SummarizedExperiment}{


If the object is a \code{SummarizedExperiment}, then the primary \code{assay} matrix
is checked for duplicated columns.
}

\examples{
## Pass ====
x <- matrix(data = seq_len(20L), ncol = 2L)
hasUniqueCols(x)

## Fail ====
x <- matrix(data = rep(seq_len(10L), times = 2L), ncol = 2L)
hasUniqueCols(x)
}
