% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasValidNames.R
\name{hasValidNames}
\alias{hasValidNames}
\alias{hasValidDimnames}
\title{Does the Input Object Have Syntactically Valid Names?}
\usage{
hasValidNames(x)

hasValidDimnames(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop}) on failure.
}
}
\description{
Does the Input Object Have Syntactically Valid Names?
}
\examples{
## Pass ====
x <- list(a = 1, b = 2)
names(x)
hasValidNames(x)

x <- datasets::iris
lapply(dimnames(x), head)
hasValidDimnames(x)

## Fail ====
x <- list(
    `1`       = 1,  # can't start with number
    `foo bar` = 2,  # no spaces
    `foo-bar` = 3   # no hyphens
)
print(x)
hasValidNames(x)

x <- datasets::mtcars
# Note the spaces in the row names here.
lapply(dimnames(x), head)
hasValidDimnames(x)
}
\seealso{
\code{validNames}.
}
