% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-is2.R
\name{is2}
\alias{is2}
\title{Is an Object from a Class?}
\usage{
is2(x, class, .xname = getNameInParent(x))
}
\description{
Alternate version of \code{methods::is}.
}
\details{
If a function named \code{is.class} exists, call \code{is.class(x)}.
If not, call \code{is(x, class)}.
}
\examples{
is2(1:5, "character")
is2(matrix(1:5), "character")
is2(1:5, c("character", "list", "numeric"))
is2(mean, c("function", "data.frame"))
}
\seealso{
\itemize{
\item \code{assertive.base::is2}.
\item \code{methods::is}.
}
}
