% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-flag-isAll.R
\name{isAll}
\alias{isAll}
\title{Does the Input Belong to or Inherit All of These Classes?}
\usage{
isAll(x, classes, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{classes}{\code{character}.
Object classes.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the Input Belong to or Inherit All of These Classes?
}
\examples{
x <- 1L

## Pass ====
isAll(x, classes = c("integer", "numeric"))

## Fail ====
isAll(x, classes = c("integer", "NULL"))
}
