% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAlpha.R
\name{isAlpha}
\alias{isAlpha}
\alias{assertIsAlpha}
\title{Is Alpha}
\usage{
isAlpha(x)

assertIsAlpha(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions  (see \code{stop()}for details).
}
}
\description{
An alpha level must be a \code{numeric scalar} greater than 0 and less than 1.
}
\examples{
isAlpha(0.05)
}
