% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAny.R
\name{isAny}
\alias{isAny}
\title{Does an Object Belong to Any of These Classes?}
\usage{
isAny(x, classes)
}
\arguments{
\item{x}{Object.}

\item{classes}{\code{character}. Object classes.}
}
\description{
Does an Object Belong to Any of These Classes?
}
