% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCharacter.R
\name{isCharacter}
\alias{isCharacter}
\title{Does the Argument Contain a (Non-Empty) Character?}
\usage{
isCharacter(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Argument Contain a (Non-Empty) Character?
}
\section{Enforced parameters}{

\itemize{
\item Must have length.
\item Cannot contain \code{NA} strings.
\item Cannot contain empty (\code{""}) strings.
}
}

\examples{
## Pass ====
isCharacter("a")

## Fail ====
isCharacter(NULL)
isCharacter(character())
isCharacter("")
isCharacter(NA_character_)
}
