% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isDark.R
\name{isDark}
\alias{isDark}
\title{Is dark mode preferred?}
\usage{
isDark()
}
\value{
\code{logical(1)}.
}
\description{
Is dark mode preferred?
}
\details{
Checks for \code{acid.dark} option in current session.
}
\note{
Updated 2022-03-04.
}
\examples{
isDark()
}
