% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDir.R
\name{isDir}
\alias{isDir}
\title{Does the Argument Contain a Single (Existing) Directory?}
\usage{
isDir(x)
}
\arguments{
\item{x}{\code{character(1)}. String specifying a local directory path that
must exist and be accessible.}
}
\description{
Does the Argument Contain a Single (Existing) Directory?
}
\examples{
isDir("~")
}
