% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-isDirectory.R
\name{isDirectory}
\alias{isDirectory}
\alias{isDir}
\alias{isADirectory}
\alias{isADir}
\title{Does the Input Contain a Directory?}
\usage{
isDirectory(x)

isDir(x)

isADirectory(x)

isADir(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{logical}.
}
\description{
Does the Input Contain a Directory?
}
\examples{
isADirectory("~")
}
\seealso{
\code{R.utils::isDirectory}.
}
