% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEqualTo.R
\name{isEqualTo}
\alias{isEqualTo}
\alias{isNotEqualTo}
\alias{isGreaterThan}
\alias{isGreaterThanOrEqualTo}
\alias{isLessThan}
\alias{isLessThanOrEqualTo}
\title{How Does the Input Relate to a Value?}
\usage{
isEqualTo(x, y, tol = 100 * .Machine$double.eps,
  .xname = get_name_in_parent(x), .yname = get_name_in_parent(x))

isNotEqualTo(x, y, tol = 100 * .Machine$double.eps,
  .xname = get_name_in_parent(x), .yname = get_name_in_parent(x))

isGreaterThan(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(x))

isGreaterThanOrEqualTo(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(x))

isLessThan(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(x))

isLessThanOrEqualTo(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(x))
}
\arguments{
\item{x}{Object.}

\item{y}{Secondary object.}

\item{tol}{Values within \code{tol} are considered equal.}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{.yname}{\emph{Not intended to be used directly.}}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
How Does the Input Relate to a Value?
}
