% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFile.R
\name{isFile}
\alias{isFile}
\alias{assertIsFile}
\alias{areFiles}
\alias{assertAreFiles}
\title{Is File?}
\usage{
isFile(x)

assertIsFile(x)

areFiles(x)

assertAreFiles(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{base::stop()} or \code{base::stopifnot()} for details).
}
}
\description{
Is File?
}
\examples{
x <- system.file("extdata/example.rds", package = "basejump")

## Requires scalar.
isFile(x)

## Parameterized.
areFiles(rep(x, times = 2L))
}
