% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFlag.R
\name{isFlag}
\alias{isFlag}
\title{Does the Argument Contain a Boolean Flag?}
\usage{
isFlag(x, na.ok = FALSE, null.ok = FALSE)
}
\arguments{
\item{x}{Object.}

\item{na.ok}{\code{logical(1)}. If set to \code{TRUE}, \code{x} may also be \code{NA}.}

\item{null.ok}{\code{logical(1)}. If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Argument Contain a Boolean Flag?
}
\examples{
## Pass ====
isFlag(TRUE)
isFlag(FALSE)

## Fail ====
isFlag("xxx")
}
