% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-flag-isHeaderLevel.R
\name{isHeaderLevel}
\alias{isHeaderLevel}
\alias{containsHeaderLevel}
\title{Does the input contain a Markdown header level?}
\usage{
isHeaderLevel(x, .name = getNameInParent(x))

containsHeaderLevel(x, .name = getNameInParent(x))
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Markdown supports header levels \code{1}-\code{7} (\code{<H1>}-\code{<H7>}).
}
\examples{
## Pass ====
isHeaderLevel(1)

## Fail ====
isHeaderLevel(0)
}
