% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isHexColor.R
\name{isHexColor}
\alias{isHexColor}
\alias{allAreHexColors}
\title{Does the input contain hexadecimal colors?}
\usage{
isHexColor(x, .xname = getNameInParent(x))

allAreHexColors(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input contain hexadecimal colors?
}
\section{Functions}{
\itemize{
\item \code{isHexColor}: Vectorized.

\item \code{allAreHexColors}: Scalar.
}}

\examples{
## Pass ====
x <- viridis::viridis(n = 2L)
class(x)
print(x)
isHexColor(x)

## Fail ====
x <- ggplot2::scale_colour_manual
class(x)
isHexColor(x)
}
\seealso{
\code{assertive::is_hex_color()}.
}
