% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isHexColorFunction.R
\name{isHexColorFunction}
\alias{isHexColorFunction}
\title{Does the Argument Contain a Function That Returns Hexadecimal Colors?}
\usage{
isHexColorFunction(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
This assert check is intended primarily to check for RColorBrewer or viridis
hexadecimal color value return.
}
\examples{
## Pass ====
x <- viridis::viridis
isHexColorFunction(x)

## Fail ====
x <- ggplot2::scale_colour_manual
isHexColorFunction(x)
}
\seealso{
RColorBrewer, viridis packages.
}
