% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isImplicitInteger.R
\name{isImplicitInteger}
\alias{isImplicitInteger}
\alias{assertIsImplicitInteger}
\alias{assertIsImplicitIntegerOrNULL}
\alias{isAnImplicitInteger}
\alias{assertIsAnImplicitInteger}
\alias{assertIsAnImplicitIntegerOrNULL}
\title{Is Implicit Integer (Integerish)?}
\usage{
isImplicitInteger(x)

assertIsImplicitInteger(x)

assertIsImplicitIntegerOrNULL(x)

isAnImplicitInteger(x)

assertIsAnImplicitInteger(x)

assertIsAnImplicitIntegerOrNULL(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Check to see if a numeric vector is an implicit integer
(e.g. \code{1} instead of \code{1L}).
}
\examples{
isImplicitInteger(c(1, 2))
isAnImplicitInteger(1)
}
\seealso{
\itemize{
\item \code{checkmate::makeAssertionFunction()}.
}
}
