% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isPackageVersion.R
\name{isPackageVersion}
\alias{isPackageVersion}
\title{Is the package installed and a specific version?}
\usage{
isPackageVersion(x, op = ">=")
}
\arguments{
\item{x}{\code{character}.
Named character vector.
Name corresponds to package name, and value corresponds to minimum version.}

\item{op}{\code{character(1)}.
Mathematical operator.
Defaults to less than or equal to.}
}
\value{
\code{logical}.
Whether package passes version check.
}
\description{
Is the package installed and a specific version?
}
\note{
Updated 2020-08-11.
}
\examples{
isPackageVersion(
    x = c(
        "basejump" = "0.1.0",
        "goalie" = "0.1.0"
    )
)
}
