% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isScalar.R
\name{isScalar}
\alias{isScalar}
\alias{isScalarInteger}
\alias{isScalarIntegerish}
\title{Is Scalar?}
\usage{
isScalar(x, metric = c("length", "elements"),
  .xname = get_name_in_parent(x))

isScalarInteger(x)

isScalarIntegerish(x, finite = NULL)
}
\arguments{
\item{x}{Object.}

\item{metric}{A string. Should be length or the number of elements be used to
determine if the object is empty/non-empty/scalar?}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{finite}{Whether all values of the vector are finite. The
non-finite values are \code{NA}, \code{Inf}, \code{-Inf} and \code{NaN}. Setting this
to something other than \code{NULL} can be expensive because the whole
vector needs to be traversed and checked.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Is Scalar?
}
\seealso{
\code{help(topic = "scalar-type-predicates", package = "rlang")}.
}
