% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isString.R
\name{isString}
\alias{isString}
\title{Is the Argument a String?}
\usage{
isString(x, na.ok = FALSE, min.chars = NULL, pattern = NULL,
  fixed = NULL, ignore.case = FALSE, null.ok = FALSE)
}
\arguments{
\item{x}{Object.}

\item{na.ok}{\code{logical(1)}. If set to \code{TRUE}, \code{x} may also be \code{NA}.}

\item{min.chars}{[\code{integer(1)}]\cr
Minimum number of characters for each element of \code{x}.}

\item{pattern}{\code{character(1)}. Pattern to use for matching.}

\item{fixed}{[\code{character(1)}]\cr
Substring to detect in \code{x}. Will be used as \code{pattern} in \code{\link[base]{grepl}}
with option \code{fixed} set to \code{TRUE}.
All non-missing elements of \code{x} must contain this substring.}

\item{ignore.case}{[\code{logical(1)}]\cr
See \code{\link[base]{grepl}}. Default is \code{FALSE}.}

\item{null.ok}{\code{logical(1)}. If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Is the Argument a String?
}
\examples{
## Pass ====
isString("a")

## Fail ====
isString(c("a", "b"))
}
