% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isURL.R
\name{isURL}
\alias{isURL}
\alias{assertIsURL}
\alias{areURLs}
\alias{assertAreURLs}
\title{Is URL?}
\usage{
isURL(x)

assertIsURL(x)

areURLs(x)

assertAreURLs(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{base::stop()} or \code{base::stopifnot()} for details).
}
}
\description{
Is URL?
}
\examples{
urls <- c("https://www.r-project.org", "ftp://r-project.org")

## Requires scalar.
isURL(urls[[1L]])

## Parameterized.
areURLs(urls)
}
