% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-checkmate-makeTest.R
\name{makeTest}
\alias{makeTest}
\alias{makeTestFunction}
\title{Make Test Function}
\usage{
makeTest(res)

makeTestFunction(fun, envir = parent.frame())
}
\arguments{
\item{res}{\code{TRUE} or \code{character(1)}.
The result of a check function: \code{TRUE} for successful checks, or send an
error message as string otherwise.}

\item{fun}{\code{function}.
Check function, which returns either \code{TRUE} or \code{character(1)}.}

\item{envir}{\code{environment}.
The environment of the created function. Defaults to \code{parent.frame}.}
}
\description{
Make Test Function
}
\seealso{
\itemize{
\item \code{checkmate::makeTest}.
\item \code{checkmate::makeTestFunction}.
}
}
