% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-matchesUniqueGeneNames.R
\name{matchesUniqueGeneNames}
\alias{matchesUniqueGeneNames}
\title{Do the Input Gene Names Match Unique Values in the Corresponding Object?}
\usage{
matchesUniqueGeneNames(x, genes)
}
\arguments{
\item{x}{Object.}

\item{genes}{\code{character}.
Genes.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
This assert check determines if a user-defined gene name query is using only
unique (non-amgibuous) symbols. It is designed to be used for gene plotting
particularly when performing single-cell RNA-seq marker analysis.
}
\examples{
x <- SummarizedExperiment::SummarizedExperiment(
    assays = matrix(
        data = seq_len(16L),
        nrow = 4L,
        ncol = 4L,
        dimnames = list(
            paste0("gene", seq_len(4L)),
            paste0("sample", seq_len(4L))
        )
    ),
    rowData = S4Vectors::DataFrame(
        geneID = paste0("ENSG0000000000", seq_len(4L)),
        geneName = paste0("SYMBOL", seq_len(4L))
    )
)
genes <- SummarizedExperiment::rowData(x)$geneName

## Pass ====
matchesUniqueGeneNames(x = x, genes = genes)
}
