% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-methodDefinition.R
\name{methodDefinition}
\alias{methodDefinition}
\alias{methodFunction}
\alias{methodFormals}
\title{Access the data inside an S4 method definition}
\usage{
methodFunction(f, signature, package)

methodFormals(f, signature, package)
}
\arguments{
\item{f}{a generic function or the character-string name of one.}

\item{signature}{the signature of classes to match to the arguments
    of \code{f}.  See the details below.}

\item{package}{\code{character(1)}.
Package name.}
}
\value{
\itemize{
\item \code{methodFormals}: \code{list}. Extract the sealed formal arguments.
\item \code{methodFunction}: \code{function}. Extract the sealed function.
}
}
\description{
Access the data inside an S4 method definition
}
\note{
Updated 2019-10-21.
}
\examples{
## Function ====
x <- methodFunction(
    f = "as.data.frame",
    signature = "DataFrame",
    package = "BiocGenerics"
)
class(x)
formals(x)

## Formals ====
x <- methodFormals(
    f = "as.data.frame",
    signature = "DataFrame",
    package = "BiocGenerics"
)
class(x)
print(x)
}
\seealso{
\itemize{
\item Modified version of \href{https://goo.gl/ymX571}{John Chambers' code}.
\item \code{help(topic = "MethodDefinition-class", package = "methods")}.
\item \code{methods::selectMethod()}.
\item \code{methods::setMethod()}.
}
}
