% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nocause-methods.R
\name{nocause}
\alias{nocause}
\alias{nocause,goalie-method}
\title{Return standard logical without cause attribute}
\usage{
nocause(object, ...)

\S4method{nocause}{goalie}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Return standard logical without cause attribute
}
\note{
Updated 2021-02-23.
}
\examples{
x <- goalie(
    object = c(FALSE, TRUE),
    cause = c(
        "Element 1 is FALSE.",
        NA_character_
    )
)
class(x)
xx <- nocause(x)
print(xx)
class(xx)
}
