% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{class}{\code{character(1)}.
Object class.}

\item{classes}{\code{character}.
Object classes.}

\item{envir}{\code{environment}.
Environment.}

\item{genes}{\code{character}.
Genes.}

\item{inherits}{\code{logical(1)}.
Should the enclosing frames of the \code{environment} be searched?}

\item{lower}{\code{numeric(1)}.
Lower boundary.}

\item{metric}{XXX.}

\item{n}{XXX.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{object}{Object.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{upper}{\code{numeric(1)}.
Upper boundary.}

\item{value}{Value.}

\item{x}{Object.}

\item{y}{Secondary object.}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{.yname}{\emph{Not intended to be used directly.}}

\item{...}{Additional arguments.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Parameters
}
\keyword{internal}
