% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print}
\alias{print}
\alias{print.goalie}
\title{Print methods for objects with a cause attribute}
\usage{
\method{print}{goalie}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Print command and return invisibly.
}
\description{
Print methods for objects with a cause attribute
}
\note{
Updated 2019-07-29.
}
