% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printString.R
\name{printString}
\alias{printString}
\title{Print string}
\usage{
printString(x, ..., max = getOption("max.print", 100L))
}
\arguments{
\item{x}{An object used to select a \code{\link[base:print]{print()}} method.}

\item{...}{Passthrough arguments to \code{\link[base:print]{print()}}.}

\item{max}{\code{integer(1)}.
Maximum length of vector. Works internally by calling
\code{\link[utils:head]{head()}} on the print capture, prior to collapse using
\code{\link[base:paste]{paste()}}. Supports \code{getOption("max.print")} global
variable.}
}
\value{
\code{character(1)}.
}
\description{
Capture \code{\link[base:print]{print()}} output as a \code{character} string.
}
\details{
Useful for returning informative messages inside a function.
}
\note{
Updated 2019-07-29.
}
\examples{
printString(c("hello", "world"))
printString(datasets::mtcars, max = 2L)
}
\seealso{
\code{cat()}.
}
