% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ranges.R
\name{ranges}
\alias{ranges}
\alias{isInRange}
\alias{isInClosedRange}
\alias{isInOpenRange}
\alias{isInLeftOpenRange}
\alias{isInRightOpenRange}
\alias{isNegative}
\alias{isPositive}
\alias{isNonNegative}
\alias{isNonPositive}
\alias{isPercentage}
\alias{isProportion}
\title{Is the input in range?}
\usage{
isInRange(x, lower = -Inf, upper = Inf, closed = c(TRUE, TRUE),
  .xname = getNameInParent(x))

isInClosedRange(x, lower = -Inf, upper = Inf,
  .xname = getNameInParent(x))

isInOpenRange(x, lower = -Inf, upper = Inf,
  .xname = getNameInParent(x))

isInLeftOpenRange(x, lower = -Inf, upper = Inf,
  .xname = getNameInParent(x))

isInRightOpenRange(x, lower = -Inf, upper = Inf,
  .xname = getNameInParent(x))

isNegative(x, .xname = getNameInParent(x))

isPositive(x, .xname = getNameInParent(x))

isNonNegative(x, .xname = getNameInParent(x))

isNonPositive(x, .xname = getNameInParent(x))

isPercentage(x, .xname = getNameInParent(x))

isProportion(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{closed}{\code{logical(2)}.
Should the lower (1) and upper (2) bounaries be closed?}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Is the input in range?
}
\section{Intervals}{

\itemize{
\item Closed: Includes all its limit points, and is denoted with square brackets.
For example, \code{[0,1]} means greater than or equal to 0 and less than or
equal to 1.
\item Open: Does not include its endpoints, and is indicated with parentheses.
For example, \code{(0,1)} means greater than 0 and less than 1.
}
}

\examples{
## Pass ====
isInRange(0, lower = 0, upper = 1)
isInRange(1, lower = 0, upper = 1)
isInClosedRange(1, lower = 0, upper = 1)

isInOpenRange(0.5, lower = 0, upper = 1)
isInLeftOpenRange(1, lower = 0, upper = 1)
isInRightOpenRange(0, lower = 0, upper = 1)

isNegative(c(-2, -1))
isPositive(c(1, 2))

isNonNegative(c(0, 1))
isNonPositive(c(-1, 0))

isPercentage(c(0, 25, 50, 100))
isProportion(c(0, 0.01, 0.1, 1))

## Fail ====
isInRange(c(2, 3), lower = 0, upper = 1)
isInClosedRange(c(2, 3), lower = 0, upper = 1)

isInOpenRange(c(1, 2), lower = 0, upper = 1)
isInLeftOpenRange(0, lower = 0)
isInRightOpenRange(1, upper = 1)

isPositive(-1)
isNegative(1)

isPercentage(110)
isProportion(1.1)
}
\seealso{
\itemize{
\item \code{assertive.numbers::is_in_range()}.
\item \code{assertive.numbers::is_in_closed_range()}.
\item \code{assertive.numbers::is_in_open_range()}.
\item \code{assertive.numbers::is_in_left_open_range()}.
\item \code{assertive.numbers::is_in_right_open_range()}.
\item \code{assertive.numbers::is_negative()}.
\item \code{assertive.numbers::is_positive()}.
\item \code{assertive.numbers::is_non_negative()}.
\item \code{assertive.numbers::is_non_positive()}.
\item \code{assertive.numbers::is_percentage()}.
\item \code{assertive.numbers::is_proportion()}.
}
}
