% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFile.R, R/package.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{assert_that}
\alias{validate_that}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{assertthat}{\code{\link[assertthat]{assert_that}}, \code{\link[assertthat]{validate_that}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{R.utils}{\code{\link[R.utils]{isFile}}}
}}

