% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireNamespaces.R
\name{requireNamespaces}
\alias{requireNamespaces}
\title{Require package namespaces}
\usage{
requireNamespaces(packages)
}
\arguments{
\item{packages}{\code{character}.
Package names to load.}
}
\description{
Require package namespaces
}
\details{
This function intentionally does not attach packages, unlike \code{library},
\code{require}, or \code{attachNamespace} functions.

Intended for use inside package functions that relying upon optional
dependencies defined via \code{Suggests} instead of \code{Imports} or \code{Depends} in
\code{DESCRIPTION} file.

Unlike \code{requireNamespace}, this variant always errors on namespace failure.
}
\note{
Updated 2023-02-07.
}
\examples{
requireNamespaces(c("base", "utils"))
}
\seealso{
\itemize{
\item \code{requireNamespace}
\item \code{attachNamespace}
\item \code{loadNamespace}
}
}
