% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeDeparse.R
\name{safeDeparse}
\alias{safeDeparse}
\title{Expression deparsing}
\usage{
safeDeparse(expr, ...)
}
\arguments{
\item{expr}{\code{expression}.
Any R expression.}

\item{...}{Passed to \link[base:deparse]{deparse()}.}
}
\description{
Turn unevaluated expressions into character strings.
}
\details{
\code{\link[=safeDeparse]{safeDeparse()}} is modified version of \link[base:deparse]{deparse()} that always
returns \code{character(1)}.
}
\examples{
safeDeparse(is.character("a"))
}
\seealso{
\itemize{
\item \code{assertive.base::safe_deparse()}.
\item \code{deparse()}.
}
}
