% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale}
\alias{scale}
\alias{assertIsColorScaleContinuousOrNULL}
\alias{assertIsColorScaleDiscreteOrNULL}
\alias{assertIsFillScaleContinuousOrNULL}
\alias{assertIsFillScaleDiscreteOrNULL}
\title{Is Scale (or NULL)?}
\usage{
assertIsColorScaleContinuousOrNULL(x)

assertIsColorScaleDiscreteOrNULL(x)

assertIsFillScaleContinuousOrNULL(x)

assertIsFillScaleDiscreteOrNULL(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{base::stop()} or \code{base::stopifnot()} for details).
}
}
\description{
Is Scale (or NULL)?
}
\examples{
color_c <- ggplot2::scale_colour_gradient(low = "red", high = "blue")
class(color_c)

color_d <- ggplot2::scale_colour_manual(values = c("red", "blue"))
class(color_d)

fill_c <- ggplot2::scale_fill_gradient(low = "red", high = "blue")
class(fill_c)

fill_d <- ggplot2::scale_fill_manual(values = c("red", "blue"))
class(fill_d)

assertIsColorScaleContinuousOrNULL(color_c)
assertIsColorScaleDiscreteOrNULL(color_d)
assertIsFillScaleContinuousOrNULL(fill_c)
assertIsFillScaleDiscreteOrNULL(fill_d)
}
