% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setCause-methods.R
\name{setCause}
\alias{setCause}
\alias{setCause,goalie-method}
\alias{setCause,logical-method}
\title{Return \code{FALSE} vector with causes of failure}
\usage{
setCause(object, ...)

\S4method{setCause}{goalie}(object, ...)

\S4method{setCause}{logical}(object, false)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{false}{\code{character}.
A character vector to set the cause to, when \code{x} is \code{FALSE}.}
}
\value{
\code{goalie}.
}
\description{
Sets the \code{cause} \link[base:attributes]{attribute} of an object and returns that
object.
}
\note{
Updated 2023-10-02.
}
\examples{
x <- setCause(
    object = c(TRUE, TRUE, FALSE, FALSE),
    false = "false"
)
print(x)
print(cause(x))
}
\seealso{
\itemize{
\item \code{cause()}.
\item \code{assertive.base::set_cause()}.
\item \code{stats::setNames()}.
}
}
