% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\name{show}
\alias{show}
\alias{show,goalie-method}
\title{Show an object}
\usage{
show(object)

\S4method{show}{goalie}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Invisible \code{NULL}.
}
\description{
Show an object
}
\note{
Updated 2023-10-02.
}
\examples{
## Match the default logical print method.
x <- goalie(
    object = c(FALSE, TRUE),
    cause = c("xxx", NA_character_)
)
show(x)
}
\author{
Michael Steinbaugh
}
