% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCauseNames.R
\name{toCauseNames}
\alias{toCauseNames}
\alias{toCauseName}
\title{Sanitize vector input to cause names}
\usage{
toCauseNames(x)

toCauseName(x)
}
\arguments{
\item{x}{\code{atomic}.}
}
\value{
\code{character}.
}
\description{
Sanitize vector input to cause names
}
\details{
Names resulting from this function do not necessarily return valid, and will
not be identical to output from \code{\link[base:make.names]{make.names()}}.
}
\section{Functions}{
\itemize{
\item \code{toCauseName()}: Scalar variant.

}}
\note{
Updated 2023-09-29.
}
\examples{
## Non-character vectors are supported.
toCauseNames(1)
toCauseNames(complex(1L))
toCauseNames(NA)
toCauseNames(TRUE)

## Doesn't use 'make.names()' to sanitize.
toCauseNames(c("sample-1", "hello world"))
}
\seealso{
\itemize{
\item \code{assertive.base:::to_names()}.
\item https://stackoverflow.com/questions/26183735
}
}
