% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNames.R
\name{toNames}
\alias{toNames}
\title{Sanitize vector input to names}
\usage{
toNames(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{character}.
}
\description{
Sanitize vector input to names
}
\note{
Names resulting from this function do not necessarily return valid, and
will not be identical to output from \code{\link[base:make.names]{make.names()}}.
The output is intended for downstream use with the \code{\link[=cause]{cause()}} function.

Updated 2019-08-10.
}
\examples{
## Non-character vectors are supported.
toNames(1)
toNames(complex(1L))
toNames(NA)
toNames(TRUE)

## Doesn't use `make.names()` to sanitize.
toNames(c("sample-1", "hello world"))
}
\seealso{
\itemize{
\item \code{assertive.base:::to_names()}.
\item https://stackoverflow.com/questions/26183735
}
}
