% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-assertive-typeDescription.R
\name{typeDescription}
\alias{typeDescription}
\title{Set the class type description}
\usage{
typeDescription(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Set the class type description
}
\examples{
typeDescription("a")
typeDescription(1)
typeDescription(list())
}
\seealso{
\code{assertive.base:::type_description()}.
}
