% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate an S4 Class}
\usage{
validate(..., env = parent.frame(), msg = NULL)
}
\arguments{
\item{...}{Additional arguments.}
}
\description{
\code{validate()} is a variant of \code{assert()} that is specifically intended to be
used inside of an S4 validity method definition.
}
\details{
Like \code{assert()}, \code{validate()} returns \code{TRUE} on success. However, on failure
it returns a \code{character} instead of a \code{stop()} call. This is the current
recommended practice for defining S4 validity methods inside of a
\code{setValidity()} call. Refer to the documentation in the methods package,
specifically on \code{validObject()} for detailed information on S4 validity
methods.
}
\examples{
validate(
    is.atomic("example"),
    is.character("example")
)
}
\seealso{
\itemize{
\item \code{methods::setValidity()}.
\item \code{methods::validObject()}.
\item \code{assertthat::validate_that()}.
}
}
