% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validNames.R
\name{validNames}
\alias{validNames}
\title{Are These Valid Names?}
\usage{
validNames(names)
}
\arguments{
\item{names}{\code{character}. Element names (e.g. row names, column names).}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Are These Valid Names?
}
\examples{
## Dots (periods) and underscores are valid.
validNames(c("sample.1", "sample_1"))

## Can't begin with a number.
validNames("293cells")

## Spaces, dashes (hyphens), and other non-alphanumerics aren't valid.
validNames("sample 1")
validNames("cell-AAAAAAAA")
validNames("GFP+")
}
\seealso{
\itemize{
\item \code{make.names()}.
\item \code{basejump::makeNames()}.
}
}
