% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-vname.R
\name{vname}
\alias{vname}
\title{Look up a variable name}
\usage{
vname(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{character(1)}.
Variable name.
}
\description{
Tries to heuristically determine the variable name of \code{x} in the
\code{parent.frame} with a combination of \code{deparse} and \code{substitute}.
}
\examples{
test <- 1L
vname(test)
}
\seealso{
checkmate::vname.
}
