% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allAreAtomic.R
\name{allAreAtomic}
\alias{allAreAtomic}
\alias{all_are_atomic}
\title{Are All Items Atomic?}
\usage{
allAreAtomic(x)

all_are_atomic(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{base::stop()} or \code{base::stopifnot()} for details).
}
}
\description{
Are All Items Atomic?
}
\examples{
allAreAtomic(datasets::mtcars)
}
