% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-flag-allAreAtomic.R
\name{allAreAtomic}
\alias{allAreAtomic}
\title{Does the input contain elements that are all atomic?}
\usage{
allAreAtomic(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input contain elements that are all atomic?
}
\examples{
## Pass ====
allAreAtomic(data.frame(a = "foo", b = "bar"))
allAreAtomic(list(a = "foo", b = "bar"))

## Fail ====
allAreAtomic(data.frame())
allAreAtomic(list(a = "x", b = list()))
}
\seealso{
\code{is.atomic()}.
}
