% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allAreAtomic.R
\name{allAreAtomic}
\alias{allAreAtomic}
\title{Does the Input Contain Elements That Are All Atomic?}
\usage{
allAreAtomic(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
Does the Input Contain Elements That Are All Atomic?
}
\examples{
## Pass ====
allAreAtomic(data.frame(a = "foo", b = "bar"))
allAreAtomic(list(a = "foo", b = "bar"))

## Fail ====
allAreAtomic(list(a = "x", b = list()))
}
