% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendToBody.R
\name{appendToBody}
\alias{appendToBody}
\title{Append values to function body}
\usage{
appendToBody(fun, values, after = 1L)
}
\arguments{
\item{fun}{\code{function}.}

\item{values}{\code{call}.
Quoted values (i.e. call) to slot in the function body.}

\item{after}{\code{integer(1)}.
Where to append in the \code{body}.
The default of \code{1L} places directly after the opening curly bracket.}
}
\value{
\code{function}.
}
\description{
Append values to function body
}
\note{
Updated 2019-08-23.
}
\examples{
## Add a deprecation call into function body.
fun <- function() { print("hello world") }
body(fun)

## values: call ====
values <- as.call(quote(.Deprecated("XXX")))
x <- appendToBody(fun = fun, values = values)
body(x)

## values: list ====
values <- list(
    quote(print("AAA")),
    quote(print("BBB"))
)
x <- appendToBody(fun = fun, values = values)
body(x)
}
\seealso{
\href{https://stackoverflow.com/questions/38732663}{Stack Overflow}
}
