% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areNonExisting.R
\name{areNonExisting}
\alias{areNonExisting}
\alias{assertAreNonExisting}
\title{Does the Variable Not Exist?}
\usage{
areNonExisting(x, envir = parent.frame(), inherits = TRUE)

assertAreNonExisting(...)
}
\arguments{
\item{x}{\code{character}. Variable names to check in \code{environment}.}

\item{envir}{\code{environment}. Environment.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions  (see \code{stop()}for details).
}
}
\description{
Does the Variable Not Exist?
}
\examples{
areNonExisting(c("XXX", "YYY"))
}
