% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-flag-areSameLength.R
\name{areSameLength}
\alias{areSameLength}
\title{Do the inputs have the same length?}
\usage{
areSameLength(x, y, .xname = getNameInParent(x),
  .yname = getNameInParent(y))
}
\arguments{
\item{x}{Object.}

\item{y}{Secondary object.}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{.yname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Do the inputs have the same length?
}
\note{
Non-zero lengths for \code{x} and \code{y} are required, otherwise the check
function will intentionally error.
}
\examples{
## Pass ====
x <- list(a = 1, b = 2)
y <- list(c = 3, d = 4)
areSameLength(x = x, y = y)

## Fail ====
x <- list(a = 1)
y <- list(b = 2, c = 3)
areSameLength(x = x, y = y)
}
