% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert That Certain Conditions Are True}
\usage{
assert(..., env = parent.frame(), msg = NULL)
}
\arguments{
\item{...}{unnamed expressions that describe the conditions to be tested.
Rather than combining expressions with \code{&&}, separate them by commas
so that better error messages can be generated.}

\item{env}{(advanced use only) the environment in which to evaluate the
assertions.}

\item{msg}{a custom error message to be printed if one of the conditions is
false.}
}
\description{
\code{assert()} behaves like \code{stopifnot()} but supports more informative error
messages. The function supports multiple individual assertions must return a
\code{boolean flag} (i.e. a single \code{TRUE} or \code{FALSE} \code{logical}).
}
\examples{
assert(
    is.atomic("example"),
    is.character("example")
)
}
\seealso{
\itemize{
\item \code{stopifnot()}.
\item \code{assertthat::assert_that()}.
}
}
