% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertthat-assert.R
\name{assert}
\alias{assert}
\title{Assert That Certain Conditions Are True}
\usage{
assert(..., env = parent.frame(), msg = NULL)
}
\arguments{
\item{...}{[any]\cr
List of calls to check functions.}
}
\description{
\code{assert()} is a drop-in replacement for \code{stopifnot()}, but is designed to
return more informative error messages.
}
\details{
Currently, \code{assert()} reexports \code{assertthat::assert_that()}.
}
\examples{
assert(
    is.atomic("example"),
    is.character("example")
)
}
\seealso{
\itemize{
\item \code{stopifnot()}.
\item \code{assertthat::assert_that()}.
\item \code{checkmate::assert()}.
}
}
