% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert that certain conditions are true}
\usage{
assert(..., msg = NULL)
}
\arguments{
\item{...}{Any number of R expressions that return \code{logical(1)}, each of
which should evaluate to \code{TRUE}. Rather than combining expressions with
\code{&&}, separate them by commas so that better error messages can be
generated.}

\item{msg}{\code{NULL} or \code{character(1)}.
Custom message to return.}
}
\value{
\code{TRUE} on success, error on failure.
}
\description{
\code{assert()} is a drop-in replacement for \code{stopifnot()} that supports more
informative error messages.
}
\details{
If any of the expressions defined in \code{...} are \code{TRUE}, \code{stop()} is called,
producing an error message indicating the first expression which was not
\code{TRUE}.
}
\note{
Updated 2021-10-08.
}
\examples{
assert(
    is.atomic("example"),
    is.character("example")
)
}
\seealso{
\itemize{
\item \code{stopifnot()}.
\item \code{assertthat::assert_that()}.
\item \code{assertive.base::assert_engine()}.
\item \code{checkmate::assert()}.
}
}
