% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert That Certain Conditions Are True}
\usage{
assert(..., combine = "and", .var.name = NULL)
}
\arguments{
\item{...}{[any]\cr
List of calls to check functions.}

\item{combine}{[\code{character(1)}]\cr
\dQuote{or} or \dQuote{and} to combine the check functions with an OR
or AND, respectively.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in error messages. Defaults to
the heuristic implemented in \code{\link{vname}}.}
}
\description{
\code{assert()} is a drop-in replacement for \code{stopifnot()} but is designed to
return more informative error messages.
}
\details{
Currently \code{assert()} reexports \code{checkmate::assert()} with a stricter default
setting of the \code{combine} argument set to \code{"and"} instead of \code{"or"}. This
enables \code{assert()} to act as a substitute for \code{stopifnot()}.
}
\examples{
assert(
    is.atomic("example"),
    is.character("example")
)
}
\seealso{
\itemize{
\item \code{stopifnot()}.
\item \code{checkmate::assert()}.
\item \code{assertthat::assert_that()}.
}
}
