% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert that certain conditions are true}
\usage{
assert(...)
}
\arguments{
\item{...}{Any number of R expressions that return \code{logical(1)}, each of
which should evaluate to \code{TRUE}.}
}
\description{
\code{\link[=assert]{assert()}} is a drop-in replacement for \link[base:stopifnot]{stopifnot()}
supporting more informative error messages.
}
\examples{
assert(
    is.atomic("example"),
    is.character("example")
)
}
\seealso{
\itemize{
\item \code{stopifnot()}.
\item \code{assertthat::assert_that()}.
\item \code{assertive.base::assert_engine()}.
\item \code{checkmate::assert()}.
}
}
