% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert that certain conditions are true}
\usage{
assert(..., msg = NULL, traceback = getOption("goalie.traceback",
  TRUE))
}
\arguments{
\item{...}{Any number of R expressions that return \code{logical(1)}, each of
which should evaluate to \code{TRUE}. Rather than combining expressions with
\code{&&}, separate them by commas so that better error messages can be
generated.}

\item{msg}{\code{NULL} or \code{character(1)}.
Custom message to return on the event of any check failure.}

\item{traceback}{\code{logical(1)}.
Include traceback in error message.
See \code{\link[base:traceback]{traceback()}} for details.}
}
\value{
\code{TRUE} on success, error on failure.
}
\description{
\code{\link[=assert]{assert()}} is a drop-in replacement for \code{\link[base:stopifnot]{stopifnot()}}
supporting more informative error messages.
}
\examples{
assert(
    is.atomic("example"),
    is.character("example")
)
}
\seealso{
\itemize{
\item \code{stopifnot()}.
\item \code{assertthat::assert_that()}.
\item \code{assertive.base::assert_engine()}.
\item \code{checkmate::assert()}.
}
}
