% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formals.R
\name{assertFormalCompress}
\alias{assertFormalCompress}
\title{Assert Formal Compression}
\usage{
assertFormalCompress(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions  (see \code{stop()}for details).
}
}
\description{
Assert Formal Compression
}
\examples{
assertFormalCompress("xz")
}
