% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapply.R
\name{bapply}
\alias{bapply}
\title{Boolean apply}
\usage{
bapply(X, FUN, ...)
}
\arguments{
\item{X}{\code{atomic} or \code{list}.}

\item{FUN}{\code{function}.
An assert check function that returns \code{logical(1)} boolean flag
(\code{TRUE}/\code{FALSE}) to \link[base:apply]{apply}.}

\item{...}{Additional arguments passed to \link[base:vapply]{vapply()}.}
}
\value{
\code{logical}.
}
\description{
\code{\link[=bapply]{bapply()}} is a wrapper for \code{\link[base:vapply]{vapply()}} that enforces
\code{FUN.VALUE = logical(1)}.
}
\note{
\itemize{
\item \code{FUN.VALUE} is always set to \code{logical(1)}.
\item \code{USE.NAMES} is always set to \code{TRUE}.
}

Updated 2019-09-06.
}
\examples{
bapply(list(a = "example", b = 1), is.character)
}
\seealso{
\itemize{
\item \code{assertive.base::bapply()}.
\item \code{vapply()}.
}
}
