% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-assertive-bapply.R
\name{bapply}
\alias{bapply}
\title{Boolean apply}
\usage{
bapply(x, predicate, ...)
}
\arguments{
\item{x}{\code{atomic} or \code{list}.}

\item{predicate}{\code{function}.
A predicate function that returns \code{logical(1)} boolean flag
(\code{TRUE}/\code{FALSE}) to \link[base:apply]{apply}.}

\item{...}{Additional arguments passed to \link[base:vapply]{vapply()}.}
}
\value{
\code{logical}.
}
\description{
\code{\link[=bapply]{bapply()}} is a wrapper for \link[base:vapply]{vapply()} that enforces
\code{FUN.VALUE = logical(1)}.
}
\note{
\code{USE.NAMES} is always set to TRUE.
}
\examples{
bapply(list(a = "example", b = 1), is.character)
}
\seealso{
\itemize{
\item \code{assertive.base::bapply()}.
\item \code{vapply()}.
}
}
