% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-callAndName.R
\name{callAndName}
\alias{callAndName}
\title{Call a function and give the result names}
\usage{
callAndName(fun, x, ...)
}
\arguments{
\item{fun}{\code{function}.}

\item{x}{The first input to \code{fun}.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\description{
Calls a function and names the result with the first argument.
}
\seealso{
\code{assertive.base::call_and_name()}.
}
