% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callAndName.R
\name{callAndName}
\alias{callAndName}
\title{Call a function and give the result names}
\usage{
callAndName(fun, x, ...)
}
\arguments{
\item{fun}{\code{function}.}

\item{x}{The first input to \code{fun}.}

\item{...}{Additional arguments passed to function defined in \code{fun}.}
}
\value{
The result of \code{fun(x, ...)}, with the names given by the argument
\code{x}.
}
\description{
Calls a function and names the result with the first argument.
}
\note{
Updated 2019-07-29.
}
\examples{
callAndName(is.finite, c(1, Inf, NA))
}
\seealso{
\code{assertive.base::call_and_name()}.
}
