% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause.R
\name{cause}
\alias{cause}
\alias{cause<-}
\alias{nocause}
\title{Get or set the \code{cause} attribute}
\usage{
cause(x)

cause(x) <- value

nocause(x)
}
\arguments{
\item{x}{Object.}

\item{value}{Value.}
}
\value{
\code{character(1)}.
Cause attribute.
}
\description{
Gets or sets the \link{cause} (of failure) \link[base:attributes]{attribute} of a
variable.
}
\examples{
## Scalar cause.
x <- isFlag("xxx")
cause(x)

## Vector cause.
x <- isInRange(c(1L, 2L), lower = 3L)
cause(x)
}
\seealso{
\itemize{
\item \code{assertive.base::cause()}.
\item \code{attributes()}.
\item \code{attr()}.
}
}
