% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cause-methods.R
\name{cause}
\alias{cause}
\alias{cause,goalie-method}
\title{Cause attribute}
\usage{
cause(object, ...)

\S4method{cause}{goalie}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Cause attribute
}
\note{
Updated 2023-10-02.
}
\examples{
x <- goalie(
    object = c(FALSE, TRUE),
    cause = c(
        "1" = "Element 1 is FALSE.",
        "2" = NA_character_
    )
)
print(x)
x <- cause(x)
print(x)
}
