% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cause-methods.R
\name{cause}
\alias{cause}
\alias{cause,goalie-method}
\title{Cause attribute}
\usage{
cause(object)

\S4method{cause}{goalie}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character}.
}
\description{
Cause attribute
}
\note{
Updated 2021-02-23.
}
\examples{
x <- goalie(
    object = c(FALSE, TRUE),
    cause = c(
        "Element 1 is FALSE.",
        NA_character_
    )
)
print(x)
x <- cause(x)
print(x)
}
