% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-allAreAtomic.R
\name{check-scalar-allAreAtomic}
\alias{check-scalar-allAreAtomic}
\alias{allAreAtomic}
\title{Does the input contain elements that are all atomic?}
\usage{
allAreAtomic(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain elements that are all atomic?
}
\note{
Updated 2019-07-29.
}
\examples{
## TRUE ====
allAreAtomic(data.frame(a = "foo", b = "bar"))
allAreAtomic(list(a = "foo", b = "bar"))

## FALSE ====
allAreAtomic(data.frame())
allAreAtomic(list(a = "x", b = list()))
}
\seealso{
\code{is.atomic()}.
}
