% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasClusters.R
\name{check-scalar-hasClusters}
\alias{check-scalar-hasClusters}
\alias{hasClusters}
\title{Does the input object contain clusters?}
\usage{
hasClusters(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input object contain clusters?
}
\note{
Updated 2023-09-29.
}
\examples{
data(SingleCellExperiment_Seurat, package = "AcidTest")

## TRUE ====
if (requireNamespace("AcidSingleCell", quietly = TRUE)) {
    hasClusters(SingleCellExperiment_Seurat)
}

## FALSE ====
hasClusters(list())
}
\seealso{
\code{AcidGenerics::clusters}.
}
