% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-hasCpu.R
\name{check-scalar-hasCpu}
\alias{check-scalar-hasCpu}
\alias{hasCpu}
\title{Does the current machine have a sufficient number of CPU cores?}
\usage{
hasCpu(n)
}
\arguments{
\item{n}{\code{integer(1)}.
Minimum number of CPU cores.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the current machine have a sufficient number of CPU cores?
}
\note{
Updated 2023-08-10.
}
\examples{
## TRUE ====
hasCpu(n = 1L)

## FALSE ====
hasCpu(n = Inf)
}
