% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasDimnames.R
\name{check-scalar-hasDimnames}
\alias{check-scalar-hasDimnames}
\alias{hasDimnames}
\alias{hasColnames}
\title{Does the input have dimnames?}
\usage{
hasDimnames(x, .xname = getNameInParent(x))

hasColnames(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input have dimnames?
}
\note{
Updated 2021-10-07.
}
\examples{
## TRUE ====
x <- datasets::mtcars
hasDimnames(x)
hasRownames(x)
hasColnames(x)

## FALSE ====
x <- data.frame()
hasDimnames(x)
hasRownames(x)
hasColnames(x)
}
\seealso{
\itemize{
\item \code{hasRownames()}.
\item \code{assertive.properties::has_dimnames()}.
\item \code{assertive.properties::has_colnames()}.
}
}
