% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasElements.R
\name{check-scalar-hasElements}
\alias{check-scalar-hasElements}
\alias{hasElements}
\alias{nElements}
\title{Does the input have elements?}
\usage{
hasElements(x, n = NULL, .xname = getNameInParent(x))

nElements(x)
}
\arguments{
\item{x}{Object.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input have elements?
}
\section{Functions}{
\itemize{
\item \code{nElements}: Return the number of elements in object.
}}

\note{
Updated 2019-10-04.
}
\examples{
## TRUE ====
hasElements("hello")
hasElements("hello", n = 1)
hasElements(list(a = 1, b = 2), n = 2)

## FALSE ====
hasElements(NULL)
hasElements(list(), n = 1)
}
\seealso{
\itemize{
\item \code{prod()}, which returns the product of all values in its arguments. This
is called internally to check the number of elements.
\item \code{assertive.properties::has_elements()}.
\item \code{assertive.properties::is_of_dimension()}.
\item \code{assertive.properties:::n_elements()}.
}
}
