% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-system-scalar-hasGitHubPAT.R
\name{check-scalar-hasGitHubPAT}
\alias{check-scalar-hasGitHubPAT}
\alias{hasGitHubPAT}
\title{Does the current session have a GitHub personal access token?}
\usage{
hasGitHubPAT()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Required for package installs from GitHub, otherwise will hit rate limit.
}
\note{
Updated 2019-10-04.
}
\examples{
## TRUE ====
Sys.setenv("GITHUB_PAT" = "XXX")
hasGitHubPAT()

## FALSE ====
Sys.setenv("GITHUB_PAT" = "")
hasGitHubPAT()
}
