% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-hasGithubPat.R
\name{check-scalar-hasGithubPat}
\alias{check-scalar-hasGithubPat}
\alias{hasGithubPat}
\title{Does the current session have a GitHub personal access token?}
\usage{
hasGithubPat()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Required for package installs from GitHub, otherwise will hit rate limit.
}
\note{
Updated 2019-10-04.
}
\examples{
## TRUE ====
Sys.setenv("GITHUB_PAT" = "XXX")
hasGithubPat()

## FALSE ====
Sys.setenv("GITHUB_PAT" = "")
hasGithubPat()
}
