% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-hasInternet.R
\name{check-scalar-hasInternet}
\alias{check-scalar-hasInternet}
\alias{hasInternet}
\title{Does the current session have an internet connection?}
\usage{
hasInternet(url = "http://www.bioconductor.org/")
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the current session have an internet connection?
}
\note{
Updated 2020-05-11.
}
\examples{
hasInternet()
}
\seealso{
\itemize{
\item \code{Biobase::testBioCConnection()}.
\item \code{curl::has_internet()}.
\item \code{curl::nslookup()}.
\item \code{RCurl::getURL()}.
\item https://stackoverflow.com/a/17620732/3911732
}
}
