% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasMetrics.R
\name{check-scalar-hasMetrics}
\alias{check-scalar-hasMetrics}
\alias{hasMetrics}
\title{Does the object contain quality control metrics?}
\usage{
hasMetrics(x, colData = c("nCount", "nFeature"), .xname = getNameInParent(x))
}
\arguments{
\item{x}{\code{SummarizedExperiment}.}

\item{colData}{\code{character}.
Column names in \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} containing
expected quality control metrics.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the object contain quality control metrics?
}
\note{
Updated 2021-01-04.
}
\examples{
if (requireNamespace("SummarizedExperiment", quietly = TRUE)) {
    data(RangedSummarizedExperiment, package = "AcidTest")
    x <- RangedSummarizedExperiment
    hasMetrics(x)
}
}
