% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasMultipleSamples.R
\name{check-scalar-hasMultipleSamples}
\alias{check-scalar-hasMultipleSamples}
\alias{hasMultipleSamples}
\title{Does the input object contain multiple samples?}
\usage{
hasMultipleSamples(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input object contain multiple samples?
}
\note{
Updated 2022-03-04.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## TRUE ====
if (requireNamespace("AcidSingleCell", quietly = TRUE)) {
    (SingleCellExperiment_splatter)
}

## FALSE ====
hasMultipleSamples(list())
}
\seealso{
\code{sampleNames}.
}
