% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasNonZeroRowsAndCols.R
\name{check-scalar-hasNonZeroRowsAndCols}
\alias{check-scalar-hasNonZeroRowsAndCols}
\alias{hasNonZeroRowsAndCols}
\title{Does the input contain non-zero rows and columns?}
\usage{
hasNonZeroRowsAndCols(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Matrix.
Currently requires a \code{matrix} or \code{sparseMatrix} as input.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Useful for quickly checking to see if we have dropped rows or columns
containing all zeros.
}
\details{
This is a common check when handling RNA-seq data prior to generating a
heatmap or applying a log transformation, for example.
}
\note{
Updated 2019-07-29.
}
\examples{
## TRUE ====
x <- matrix(data = seq_len(4), nrow = 2)
print(x)
hasNonZeroRowsAndCols(x)

x <- matrix(data = rep(1, times = 2), byrow = TRUE)
print(x)
hasNonZeroRowsAndCols(x)

x <- matrix(data = rep(1, times = 2), byrow = FALSE)
print(x)
hasNonZeroRowsAndCols(x)

## FALSE ====
x <- matrix(nrow = 0, ncol = 0)
print(x)
hasNonZeroRowsAndCols(x)

x <- matrix(nrow = 1, ncol = 0)
print(x)
hasNonZeroRowsAndCols(x)

x <- matrix(nrow = 0, ncol = 1)
print(x)
hasNonZeroRowsAndCols(x)
}
