% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-hasRAM.R
\name{check-scalar-hasRAM}
\alias{check-scalar-hasRAM}
\alias{hasRAM}
\title{Does the current machine have a sufficient RAM?}
\usage{
hasRAM(n)
}
\arguments{
\item{n}{\code{integer(1)}.
Minimum amount of RAM in gigabytes (GB).}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the current machine have a sufficient RAM?
}
\note{
Updated 2023-08-10.
}
\examples{
## TRUE ====
hasRAM(n = 1L)

## FALSE ====
hasRAM(n = Inf)
}
