% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasSubset.R
\name{check-scalar-hasSubset}
\alias{check-scalar-hasSubset}
\alias{hasSubset}
\title{Does the object contain a subset of data?}
\usage{
hasSubset(x, metadata = "subset", .xname = getNameInParent(x))
}
\arguments{
\item{x}{\code{Annotated}.}

\item{metadata}{\code{character}.
Names in \code{\link[S4Vectors:Annotated-class]{metadata()}} that denote object is subset.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the object contain a subset of data?
}
\note{
Updated 2020-04-08.
}
\examples{
if (requireNamespace("SummarizedExperiment", quietly = TRUE)) {
    data(RangedSummarizedExperiment, package = "acidtest")
    x <- RangedSummarizedExperiment
    hasSubset(x)
}
}
