% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasValidNames.R
\name{check-scalar-hasValidNames}
\alias{check-scalar-hasValidNames}
\alias{hasValidNames}
\alias{hasValidDimnames}
\title{Does the input object have syntactically valid names?}
\usage{
hasValidNames(x)

hasValidDimnames(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input object have syntactically valid names?
}
\note{
Updated 2022-05-13.
}
\examples{
## TRUE ====
x <- list("a" = 1L, "b" = 2L)
names(x)
hasValidNames(x)

x <- datasets::iris
lapply(X = dimnames(x), FUN = head)
hasValidDimnames(x)

## FALSE ====
x <- list(
    "1" = 1, # can't start with number
    "foo bar" = 2, # no spaces
    "foo-bar" = 3 # no hyphens
)
print(x)
hasValidNames(x)

x <- datasets::mtcars
# Note the spaces in the row names here.
lapply(X = dimnames(x), FUN = head)
hasValidDimnames(x)
}
\seealso{
\code{validNames()}.
}
