% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isAlpha.R
\name{check-scalar-isAlpha}
\alias{check-scalar-isAlpha}
\alias{isAlpha}
\title{Does the input contain an alpha level?}
\usage{
isAlpha(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
An alpha level must be \code{numeric(1)} greater than 0 and less than 1.
}
\note{
Updated 2019-08-10.
}
\examples{
## TRUE ====
isAlpha(0.05)

## FALSE ====
isAlpha("xxx")
isAlpha(1L)
}
