% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isBiocParallelParam.R
\name{check-scalar-isBiocParallelParam}
\alias{check-scalar-isBiocParallelParam}
\alias{isBiocParallelParam}
\title{Does the input contain a BiocParallel parameter?}
\usage{
isBiocParallelParam(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a BiocParallel parameter?
}
\note{
Updated 2023-07-21.
}
\examples{
## TRUE ====
if (isInstalled("BiocParallel")) {
    isBiocParallelParam(BiocParallel::bpparam())
}

## FALSE ====
isBiocParallelParam(list())
}
\seealso{
\code{BiocParallel::bpparam}.
}
