% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isGgscale.R
\name{check-scalar-isGgscale}
\alias{check-scalar-isGgscale}
\alias{isGgscale}
\title{Does the input contain a ggplot2 scale?}
\usage{
isGgscale(
  x,
  scale = c("continuous", "discrete"),
  aes = c("color", "colour", "fill"),
  nullOk = FALSE
)
}
\arguments{
\item{x}{Object.}

\item{scale}{\code{character(1)}.
Type of scale, either \code{"continuous"} or \code{"discrete"}.}

\item{aes}{\code{character(1)}.
Aesthetic mapping, either  \code{"color"}/\code{"colour"} or \code{"fill"}.
Note that ggplot2 prefers British spelling.}

\item{nullOk}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a ggplot2 scale?
}
\note{
Updated 2019-09-14.
}
\examples{
library(ggplot2)

color_c <- scale_color_gradient(low = "red", high = "blue")
class(color_c)

color_d <- scale_color_manual(values = c("red", "blue"))
class(color_d)

fill_c <- scale_fill_gradient(low = "red", high = "blue")
class(fill_c)

fill_d <- scale_fill_manual(values = c("red", "blue"))
class(fill_d)

isGgscale(x = color_c, scale = "continuous", aes = "color")
isGgscale(x = color_d, scale = "discrete", aes = "color")
isGgscale(x = fill_c, scale = "continuous", aes = "fill")
isGgscale(x = fill_d, scale = "discrete", aes = "fill")
}
