% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isGGScale.R
\name{check-scalar-isGGScale}
\alias{check-scalar-isGGScale}
\alias{isGGScale}
\title{Does the input contain a ggplot2 scale?}
\usage{
isGGScale(x, scale = c("continuous", "discrete"), aes = c("colour",
  "fill"), nullOK = FALSE)
}
\arguments{
\item{x}{Object.}

\item{scale}{\code{character(1)}.
Type of scale, either \code{"continuous"} or \code{"discrete"}.}

\item{aes}{\code{character(1)}.
Aesthetic mapping, either \code{"colour"} or \code{"fill"}.
Note that ggplot2 prefers British spelling, so we're enforcing that
convention here.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input contain a ggplot2 scale?
}
\note{
Updated 2019-07-29.
}
\examples{
library(ggplot2)

colour_c <- scale_colour_gradient(low = "red", high = "blue")
class(colour_c)

colour_d <- scale_colour_manual(values = c("red", "blue"))
class(colour_d)

fill_c <- scale_fill_gradient(low = "red", high = "blue")
class(fill_c)

fill_d <- scale_fill_manual(values = c("red", "blue"))
class(fill_d)

isGGScale(x = colour_c, scale = "continuous", aes = "colour")
isGGScale(x = colour_d, scale = "discrete", aes = "colour")
isGGScale(x = fill_c, scale = "continuous", aes = "fill")
isGGScale(x = fill_d, scale = "discrete", aes = "fill")
}
