% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-isMacos.R
\name{check-scalar-isMacos}
\alias{check-scalar-isMacos}
\alias{isMacos}
\title{Is the operating system macOS?}
\usage{
isMacos()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the operating system macOS?
}
\note{
Updated 2020-04-12.
}
\examples{
isMacos()
}
