% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isOfDimension.R
\name{check-scalar-isOfDimension}
\alias{check-scalar-isOfDimension}
\alias{isOfDimension}
\title{Does the input contain specific dimensions?}
\usage{
isOfDimension(x, n, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain specific dimensions?
}
\note{
Updated 2021-10-08.
}
\examples{
x <- data.frame(a = c(TRUE, FALSE), b = c(FALSE, TRUE))
dim(x)

## TRUE ====
isOfDimension(x, n = c(2L, 2L))

## FALSE ====
isOfDimension(x, n = c(1L, 2L))
isOfDimension(x, n = c(2L, 1L))
}
\seealso{
\itemize{
\item \code{assertive.properties::is_of_dimension()}.
}
}
