% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isTximport.R
\name{check-scalar-isTximport}
\alias{check-scalar-isTximport}
\alias{isTximport}
\title{Does the input contain a tximport list?}
\usage{
isTximport(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a tximport list?
}
\note{
Updated 2023-07-21.
}
\examples{
## TRUE ====

## FALSE ====
object <- list()
isTximport(object)
}
\seealso{
\code{BiocParallel::bpparam}.
}
