% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-check-scalar-isVSCode.R
\name{check-scalar-isVSCode}
\alias{check-scalar-isVSCode}
\alias{isVSCode}
\title{Is the current R session running inside Visual Studio Code?}
\usage{
isVSCode()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the current R session running inside Visual Studio Code?
}
\note{
Updated 2022-10-18.
}
\examples{
isVSCode()
}
