% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isCompressedFile.R
\name{check-vector-isCompressedFile}
\alias{check-vector-isCompressedFile}
\alias{isCompressedFile}
\alias{isACompressedFile}
\alias{allAreCompressedFiles}
\title{Does the input contain a compressed file?}
\usage{
isCompressedFile(x)

isACompressedFile(x, nullOK = FALSE)

allAreCompressedFiles(x)
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a compressed file?
}
\details{
Currently only performs a simple check, based on file extension match.
}
\section{Functions}{
\itemize{
\item \code{isCompressedFile()}: Vectorized.

\item \code{isACompressedFile()}: Scalar.

\item \code{allAreCompressedFiles()}: Scalar.

}}
\note{
Updated 2022-12-14.
}
\examples{
## TRUE ====
x <- c("sample1.fastq.gz", "sample2.fastq.bz2")
file.create(x)
isCompressedFile(x)
unlink(x)

## FALSE ====
x <- c("sample1.fastq", "sample2.fastq")
file.create(x)
isCompressedFile(x)
unlink(x)
}
\seealso{
\itemize{
\item \code{\link[=isFile]{isFile()}}.
\item \code{compressFileExt}.
}
}
