% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isExisting.R
\name{check-vector-isExisting}
\alias{check-vector-isExisting}
\alias{isExisting}
\alias{isNonExisting}
\alias{allAreExisting}
\alias{allAreNonExisting}
\title{Does the requested input exist in the environment?}
\usage{
isExisting(x, envir = parent.frame(), inherits = FALSE,
  .xname = getNameInParent(x))

isNonExisting(x, envir = parent.frame(), inherits = FALSE,
  .xname = getNameInParent(x))

allAreExisting(x, envir = parent.frame(), inherits = FALSE,
  .xname = getNameInParent(x))

allAreNonExisting(x, envir = parent.frame(), inherits = FALSE,
  .xname = getNameInParent(x))
}
\arguments{
\item{x}{\code{character}.
Variable names to check in \code{environment}.}

\item{envir}{\code{environment}.
Environment to use for assignment.}

\item{inherits}{\code{logical(1)}.
Should the enclosing frames of the \code{environment} be searched?}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the requested input exist in the environment?
}
\section{Functions}{
\itemize{
\item \code{isExisting}: Vectorized.

\item \code{isNonExisting}: Vectorized.

\item \code{allAreExisting}: Scalar.

\item \code{allAreNonExisting}: Scalar.
}}

\note{
\code{\link[base:exists]{exists()}} only supports \code{character(1)}, so we are
exporting \code{\link[=isExisting]{isExisting()}} as a convenience function to check multiple
variables in a single call.

Updated 2019-07-29.
}
\examples{
a <- 1L
b <- 2L

## TRUE ====
isExisting(c("a", "b"))
allAreNonExisting(c("x", "y"))

## FALSE ====
isExisting(c("x", "y"))
}
\seealso{
\code{assertive.code::is_existing()}.
}
