% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isExistingURL.R
\name{check-vector-isExistingURL}
\alias{check-vector-isExistingURL}
\alias{isExistingURL}
\alias{isAnExistingURL}
\alias{allAreExistingURLs}
\title{Does the input contain an existing (active) URL?}
\usage{
isExistingURL(x, .xname = getNameInParent(x))

isAnExistingURL(x, .xname = getNameInParent(x))

allAreExistingURLs(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain an existing (active) URL?
}
\section{Functions}{
\itemize{
\item \code{isExistingURL()}: Vectorized.

\item \code{isAnExistingURL()}: Scalar. Requires a single URL.

\item \code{allAreExistingURLs()}: Scalar. Checks that all strings are
existing URLs.

}}
\note{
Updated 2023-08-24.
}
\examples{
## TRUE ====
isAnExistingURL("https://acidgenomics.com")

## FALSE ====
isAnExistingURL("https://failwhale.acidgenomics.com")
}
\seealso{
\itemize{
\item \code{open.connection()}.
\item \code{curl::has_internet()}.
\item \code{curl::nslookup()}.
\item \code{RCurl::getURL()}.
\item https://stackoverflow.com/questions/52911812
\item https://stackoverflow.com/a/17620732/3911732
}
}
