% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isExistingURL.R
\name{check-vector-isExistingURL}
\alias{check-vector-isExistingURL}
\alias{isExistingURL}
\alias{isAnExistingURL}
\alias{allAreExistingURLs}
\title{Does the input contain an existing (active) URL?}
\usage{
isExistingURL(x)

isAnExistingURL(x)

allAreExistingURLs(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain an existing (active) URL?
}
\details{
Supports HTTPS, HTTP, and FTP protocols.
}
\section{Functions}{
\itemize{
\item \code{isExistingURL()}: Vectorized.

\item \code{isAnExistingURL()}: Scalar. Requires a single URL.

\item \code{allAreExistingURLs()}: Scalar. Checks that all strings are
existing URLs.

}}
\note{
Updated 2023-09-19.
}
\examples{
## TRUE ====
isAnExistingURL("https://acidgenomics.com/")

## FALSE ====
isAnExistingURL("https://failwhale.acidgenomics.com/")
}
\seealso{
\itemize{
\item \code{curlGetHeaders()}: Amazing function with good timeout control.
\item \code{open.connection()}: Base method with no dependencies, but prone to hang
with poor timeout control.
\item \code{RCurl::url.exists()}: Seems to be good but adds external dependency.
\item urlchecker and curl packages.
\item https://github.com/r-lib/urlchecker/blob/main/inst/tools/urltools.R
\item https://stackoverflow.com/questions/52911812
\item https://stackoverflow.com/a/17620732/3911732
}
}
