% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isFile.R
\name{check-vector-isFile}
\alias{check-vector-isFile}
\alias{isFile}
\alias{isAFile}
\alias{allAreFiles}
\title{Does the input contain a file?}
\usage{
isFile(x)

isAFile(x, nullOk = FALSE)

allAreFiles(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a file?
}
\section{Functions}{
\itemize{
\item \code{isFile()}: Vectorized.

\item \code{isAFile()}: Scalar.

\item \code{allAreFiles()}: Scalar.

}}
\note{
Updated 2023-07-13.
}
\examples{
## TRUE ====
x <- "example.txt"
file.create(x)
isAFile(x)
unlink(x)

## FALSE ====
isFile(c("~", "."))
}
\seealso{
\itemize{
\item \code{file.exists()}.
\item \code{R.utils::isFile()}.
}
}
