% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isFile.R
\name{check-vector-isFile}
\alias{check-vector-isFile}
\alias{isFile}
\alias{isAFile}
\alias{allAreFiles}
\title{Does the input contain a file?}
\usage{
isFile(x)

isAFile(x, nullOK = FALSE)

allAreFiles(x)
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input contain a file?
}
\section{Functions}{
\itemize{
\item \code{isFile}: Vectorized.

\item \code{isAFile}: Scalar.

\item \code{allAreFiles}: Scalar.
}}

\note{
Updated 2019-07-29.
}
\examples{
## TRUE ====
x <- "example.txt"
file.create(x)
isAFile(x)
unlink(x)

## FALSE ====
isFile(c("~", "."))
}
\seealso{
\itemize{
\item \code{file.exists()}.
\item \code{R.utils::isFile()}.
}
}
