% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isInstalled.R
\name{check-vector-isInstalled}
\alias{check-vector-isInstalled}
\alias{isInstalled}
\alias{allAreInstalled}
\title{Is the package installed?}
\usage{
isInstalled(x)

allAreInstalled(x)
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Is the package installed?
}
\section{Functions}{
\itemize{
\item \code{isInstalled}: Vectorized.

\item \code{allAreInstalled}: Scalar.
}}

\note{
Updated 2020-04-07.
}
\examples{
## TRUE ====
isInstalled(c("base", "utils"))

## FALSE ====
isInstalled(c("AAA", "BBB"))
}
