% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isIntegerish.R
\name{check-vector-isIntegerish}
\alias{check-vector-isIntegerish}
\alias{isIntegerish}
\alias{isInt}
\title{Is the input integer(ish)?}
\usage{
isIntegerish(x, .xname = getNameInParent(x))

isInt(x, nullOK = FALSE)
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Check for valid input of either explicit (e.g. \code{1L}) and/or implict
(e.g. \code{1}) \code{integer}.
}
\section{Functions}{
\itemize{
\item \code{isIntegerish}: Vectorized.

\item \code{isInt}: Scalar.
}}

\note{
Updated 2019-07-29.
}
\examples{
## TRUE ====
isIntegerish(seq_len(2L))
isIntegerish(c(1, 2))

## FALSE ====
isIntegerish(0.1)
}
\seealso{
\itemize{
\item \code{isInt()} or \code{isScalarIntegerish()} for scalar.
\item \code{rlang::is_integerish()}.
\item \code{checkmate::checkIntegerish()}.
}
}
