% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isTempFile.R
\name{check-vector-isTempFile}
\alias{check-vector-isTempFile}
\alias{isTempFile}
\alias{isATempFile}
\alias{allAreTempFiles}
\title{Does the input contain a temporary file (that exists on disk)?}
\usage{
isTempFile(x)

isATempFile(x)

allAreTempFiles(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a temporary file (that exists on disk)?
}
\section{Functions}{
\itemize{
\item \code{isTempFile()}: Vectorized.

\item \code{isATempFile()}: Scalar.

\item \code{allAreTempFiles()}: Scalar.

}}
\note{
Updated 2023-07-13.
}
\examples{
## The temporary file must exist on disk.

## TRUE ====
x <- tempfile()
file.create(x)
file.exists(x)
isATempFile(x)
unlink(x)

## FALSE ====
x <- tempfile()
file.exists(x)
isATempFile(x)
}
\seealso{
\itemize{
\item \code{tempfile()}.
}
}
