% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isURL.R
\name{check-vector-isURL}
\alias{check-vector-isURL}
\alias{isURL}
\alias{isAURL}
\alias{allAreURLs}
\title{Does the input contain a URL?}
\usage{
isURL(x, .xname = getNameInParent(x))

isAURL(x, .xname = getNameInParent(x))

allAreURLs(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with cause set.
}
\description{
Does the input contain a URL?
}
\section{Functions}{
\itemize{
\item \code{isURL()}: Vectorized.

\item \code{isAURL()}: Scalar. Requires a single URL.

\item \code{allAreURLs()}: Scalar. Checks that all strings are URLs.

}}
\note{
Updated 2021-07-19.
}
\examples{
urls <- c("https://www.r-project.org", "ftp://r-project.org")

## TRUE ====
isURL(urls)
isAURL(urls[[1L]])
allAreURLs(urls)

## FALSE ====
isURL("xxx")
isAURL(urls)
}
