% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalie-checkAllAreAtomic.R
\name{checkAllAreAtomic}
\alias{checkAllAreAtomic}
\alias{allAreAtomic}
\alias{all_are_atomic}
\alias{check_all_are_atomic}
\alias{testAllAreAtomic}
\alias{test_all_are_atomic}
\alias{assertAllAreAtomic}
\alias{assert_all_are_atomic}
\alias{expect_all_are_atomic}
\title{Does an Object Contain Elements That Are All Atomic?}
\usage{
checkAllAreAtomic(x)

check_all_are_atomic(x)

testAllAreAtomic(x)

test_all_are_atomic(x)

assertAllAreAtomic(x, .var.name = checkmate::vname(x), add = NULL)

assert_all_are_atomic(x, .var.name = checkmate::vname(x), add = NULL)

expect_all_are_atomic(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{.var.name}{\code{character(1)}. Name of the checked object to print in
assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}

\item{label}{\code{character(1)}. Name of the checked object to print in messages.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does an Object Contain Elements That Are All Atomic?
}
\examples{
## Pass ====
checkAllAreAtomic(data.frame(a = "foo", b = "bar"))
checkAllAreAtomic(list(a = "foo", b = "bar"))

## Fail ====
checkAllAreAtomic(list(a = "x", b = list()))
}
