% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnyClass.R, R/verbs-is.R
\name{checkAnyClass}
\alias{checkAnyClass}
\alias{anyClass}
\alias{any_class}
\alias{check_any_class}
\alias{testAnyClass}
\alias{test_any_class}
\alias{assertAnyClass}
\alias{assert_any_class}
\alias{expect_any_class}
\alias{isAny}
\alias{is_any}
\title{Does an Object Belong to Any of These Classes?}
\usage{
checkAnyClass(x, classes)

check_any_class(x, classes)

testAnyClass(x, classes)

test_any_class(x, classes)

assertAnyClass(x, classes, .var.name = checkmate::vname(x), add = NULL)

assert_any_class(x, classes, .var.name = checkmate::vname(x),
  add = NULL)

expect_any_class(x, classes, info = NULL, label = vname(x))

isAny(x, classes)
}
\arguments{
\item{x}{Object.}

\item{classes}{\code{character}. Object classes.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}

\item{label}{XXX}
}
\description{
Does an Object Belong to Any of These Classes?
}
\section{Functions}{
\itemize{
\item \code{isAny}: Less strict variant that calls \code{assertive::is2()}
internally, and allows for inherited classes to pass.

\item \code{is_any}: snake alias.
}}

\examples{
x <- 1L

## Note that `checkAnyClass()` extends `checkClass()`, which  is very strict.
is(x, "numeric")
checkmate::testClass(x, "numeric")

## Pass ====
checkAnyClass(x, classes = c("integer", "character"))

## Fail ====
checkAnyClass(x, classes = c("atomic", "numeric"))
}
\seealso{
\itemize{
\item \code{checkmate::checkClass()}.
\item \code{checkmate::checkMultiClass()}.
\item \code{assertive::is2()}.
\item \code{assertive::assert_is_any_of()}.
}
}
