% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAreNonExisting.R
\name{checkAreNonExisting}
\alias{checkAreNonExisting}
\alias{areNonExisting}
\alias{check_are_non_existing}
\alias{testAreNonExisting}
\alias{test_are_non_existing}
\alias{assertAreNonExisting}
\alias{assert_are_non_existing}
\alias{expect_are_non_existing}
\title{Does the Requested Variable Exist?}
\usage{
checkAreNonExisting(x, envir = parent.frame(), inherits = TRUE)

check_are_non_existing(x, envir = parent.frame(), inherits = TRUE)

testAreNonExisting(x, envir = parent.frame(), inherits = TRUE)

test_are_non_existing(x, envir = parent.frame(), inherits = TRUE)

assertAreNonExisting(x, envir = parent.frame(), inherits = TRUE,
  .var.name = checkmate::vname(x), add = NULL)

assert_are_non_existing(x, envir = parent.frame(), inherits = TRUE,
  .var.name = checkmate::vname(x), add = NULL)

expect_are_non_existing(x, envir = parent.frame(), inherits = TRUE,
  info = NULL, label = vname(x))
}
\arguments{
\item{x}{\code{character}. Variable names to check in \code{environment}.}

\item{envir}{\code{environment}. Environment.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Requested Variable Exist?
}
\examples{
## Pass ====
a <- NULL
b <- NULL
checkAreNonExisting(c("a", "b"))

## Fail ====
a <- 1L
b <- NULL
checkAreNonExisting(c("a", "b"))
}
\seealso{
\itemize{
\item \code{checkmate::makeTestFunction()}.
\item \code{checkmate::makeAssertionFunction()}.
\item \code{checkmate::makeExpectationFunction()}.
}
}
