% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHexColorFunction.R
\name{checkHexColorFunction}
\alias{checkHexColorFunction}
\alias{hexColorFunction}
\alias{hex_color_function}
\alias{check_hex_color_function}
\alias{testHexColorFunction}
\alias{test_hex_color_function}
\alias{assertHexColorFunction}
\alias{assert_hex_color_function}
\alias{expect_hex_color_function}
\title{Does the Argument Contain a Function That Returns Hexadecimal Colors?}
\usage{
checkHexColorFunction(x, null.ok = FALSE)

check_hex_color_function(x, null.ok = FALSE)

testHexColorFunction(x, null.ok = FALSE)

test_hex_color_function(x, null.ok = FALSE)

assertHexColorFunction(x, null.ok = FALSE,
  .var.name = checkmate::vname(x), add = NULL)

assert_hex_color_function(x, null.ok = FALSE,
  .var.name = checkmate::vname(x), add = NULL)

expect_hex_color_function(x, null.ok = FALSE, info = NULL,
  label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{null.ok}{\code{logical(1)}. If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{.var.name}{\code{character(1)}. Name of the checked object to print in
assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}

\item{label}{\code{character(1)}. Name of the checked object to print in messages.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
This assert check is intended primarily to check for RColorBrewer or viridis
hexadecimal color value return.
}
\examples{
## Pass ====
x <- viridis::viridis
checkHexColorFunction(x)

## Fail ====
x <- ggplot2::scale_colour_manual
checkHexColorFunction(x)
}
\seealso{
RColorBrewer, viridis
}
