% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkURL.R
\name{checkURL}
\alias{checkURL}
\alias{url}
\alias{urls}
\alias{check_url}
\alias{testURL}
\alias{test_url}
\alias{assertURL}
\alias{assert_url}
\alias{expect_url}
\alias{checkURLs}
\alias{check_urls}
\alias{testURLs}
\alias{test_urls}
\alias{assertURLs}
\alias{assert_urls}
\alias{expect_urls}
\title{Does the Argument Contain a URL?}
\usage{
checkURL(x)

check_url(x)

testURL(x)

test_url(x)

assertURL(x, .var.name = checkmate::vname(x), add = NULL)

assert_url(x, .var.name = checkmate::vname(x), add = NULL)

expect_url(x, info = NULL, label = vname(x))

checkURLs(x)

check_urls(x)

testURLs(x)

test_urls(x)

assertURLs(x, .var.name = checkmate::vname(x), add = NULL)

assert_urls(x, .var.name = checkmate::vname(x), add = NULL)

expect_urls(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{Object.}

\item{.var.name}{\code{character(1)}. Name of the checked object to print in
assertions.}

\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}

\item{label}{\code{character(1)}. Name of the checked object to print in messages.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Does the Argument Contain a URL?
}
\section{Functions}{
\itemize{
\item \code{checkURL}: Requires a single URL.

\item \code{checkURLs}: Requires multiple URLs, which is less common.
}}

\examples{
urls <- c("https://www.r-project.org", "ftp://r-project.org")

## Pass ====
checkURL(urls[[1L]])
checkURLs(urls)

## Fail ====
checkURL("xxx")
checkURL(urls)
checkURLs(urls[[1L]])
}
