% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containsAlpha.R
\name{containsAlpha}
\alias{containsAlpha}
\title{Does the Argument Contain an Alpha Level?}
\usage{
containsAlpha(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
An alpha level must be a \code{scalar numeric} greater than 0 and less than 1.
}
\examples{
## Pass ====
containsAlpha(0.05)

## Fail ====
containsAlpha("xxx")
containsAlpha(1L)
}
