% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containsHeaderLevel.R
\name{containsHeaderLevel}
\alias{containsHeaderLevel}
\title{Does the Argument Contain a Markdown Header Level?}
\usage{
containsHeaderLevel(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Markdown supports header levels \code{1}-\code{7} (\code{<H1>}-\code{<H7>}).
}
\examples{
## Pass ====
containsHeaderLevel(1)

## Fail ====
containsHeaderLevel(0)
}
