% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containsHexColors.R
\name{containsHexColors}
\alias{containsHexColors}
\title{Does the Argument Contain Hexadecimal Colors?}
\usage{
containsHexColors(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Argument Contain Hexadecimal Colors?
}
\examples{
## Pass ====
x <- viridis::viridis(n = 2L)
containsHexColors(x)

## Fail ====
x <- ggplot2::scale_colour_manual
containsHexColors(x)
}
\seealso{
\code{assertive::is_hex_color()}.
}
