% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-containsHexColors.R
\name{containsHexColors}
\alias{containsHexColors}
\alias{areHexColors}
\title{Does the Input Contain Hexadecimal Colors?}
\usage{
containsHexColors(x, .xname = getNameInParent(x))

areHexColors(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
Does the Input Contain Hexadecimal Colors?
}
\examples{
## Pass ====
x <- viridis::viridis(n = 2L)
class(x)
containsHexColors(x)

## Fail ====
x <- ggplot2::scale_colour_manual
class(x)
containsHexColors(x)
}
\seealso{
\code{assertive::is_hex_color()}.
}
