% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-containsURL.R
\name{containsURL}
\alias{containsURL}
\alias{containsAURL}
\alias{isURL}
\alias{isAURL}
\title{Does the input contain a URL?}
\usage{
containsURL(x)

containsAURL(x)

isURL(x)

isAURL(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{logical}.
}
\description{
Does the input contain a URL?
}
\section{Functions}{
\itemize{
\item \code{containsURL}: Supports multiple URLs.

\item \code{containsAURL}: Requires a single URL.
}}

\examples{
urls <- c("https://www.r-project.org", "ftp://r-project.org")

## Pass ====
containsURL(urls)
containsAURL(urls[[1L]])

## Fail ====
containsURL("xxx")
containsAURL(urls)
}
