% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containsURL.R
\name{containsURL}
\alias{containsURL}
\title{Does the Argument Contain a URL?}
\usage{
containsURL(x, string = FALSE)
}
\arguments{
\item{x}{Object.}

\item{string}{\code{logical(1)}. Require match against string?}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Argument Contain a URL?
}
\examples{
urls <- c("https://www.r-project.org", "ftp://r-project.org")

## Pass ====
containsURL(urls[[1L]], string = TRUE)
containsURL(urls, string = FALSE)

## Fail ====
containsURL("xxx")
}
