% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{assertIsAnnotable}
\alias{assertIsGFF}
\alias{assertAreGeneAnnotations}
\alias{assertAreTranscriptAnnotations}
\alias{assertIsCharacterOrNULL}
\alias{assertIsDataFrameOrNULL}
\alias{assertIsGene2symbol}
\alias{assertIsTx2gene}
\title{Defunct Functions}
\usage{
assertIsAnnotable(...)

assertIsGFF(...)

assertAreGeneAnnotations(...)

assertAreTranscriptAnnotations(...)

assertIsCharacterOrNULL(...)

assertIsDataFrameOrNULL(...)

assertIsGene2symbol(...)

assertIsTx2gene(...)
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\keyword{internal}
