% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{containsAlpha}
\alias{containsHeaderLevel}
\alias{containsHexColors}
\alias{containsURL}
\alias{containsAURL}
\title{Deprecated functions}
\usage{
containsAlpha(x, .xname = getNameInParent(x))

containsHeaderLevel(x, .xname = getNameInParent(x))

containsHexColors(x, .xname = getNameInParent(x))

containsURL(x, .xname = getNameInParent(x))

containsAURL(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\examples{
a <- function(...) {
    .Deprecated("b")
    b(...)
}

b <- function(x) {
    x + 1L
}

withCallingHandlers(
    expr = a(1L),
    warning = function(w) {
        print(w)
        invokeRestart("muffleWarning")
    }
)
}
\seealso{
\link[base:Deprecated]{Deprecated}.
}
\keyword{internal}
