% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{containsAlpha}
\alias{containsHeaderLevel}
\alias{containsHexColors}
\alias{containsURL}
\alias{containsAURL}
\title{Deprecated functions}
\usage{
containsAlpha(x, .xname = getNameInParent(x))

containsHeaderLevel(x, .xname = getNameInParent(x))

containsHexColors(x, .xname = getNameInParent(x))

containsURL(x, .xname = getNameInParent(x))

containsAURL(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
