% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{assertAllAreNonExisting}
\alias{assertAllAreValidNames}
\alias{assertIsAHeaderLevel}
\alias{assertIsAStringOrNULL}
\title{Deprecated Functions}
\usage{
assertAllAreNonExisting(...)

assertAllAreValidNames(...)

assertIsAHeaderLevel(...)

assertIsAStringOrNULL(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Deprecated()}.
}
\description{
Deprecated Functions
}
\keyword{internal}
