% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{assertAllAreValidNames}
\alias{assertFormalGene2symbol}
\alias{assertIsAHeaderLevel}
\alias{assertIsAStringOrNULL}
\title{Deprecated Functions}
\usage{
assertAllAreValidNames(...)

assertFormalGene2symbol(...)

assertIsAHeaderLevel(...)

assertIsAStringOrNULL(...)
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\keyword{internal}
