% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-existing.R
\name{existing}
\alias{existing}
\alias{isExisting}
\alias{areExisting}
\alias{isNonExisting}
\alias{areNonExisting}
\title{Does the requested input exist in the environment?}
\usage{
isExisting(x, envir = parent.frame(), inherits = FALSE,
  .xname = getNameInParent(x))

areExisting(x, envir = parent.frame(), inherits = FALSE,
  .xname = getNameInParent(x))

isNonExisting(x, envir = parent.frame(), inherits = FALSE,
  .xname = getNameInParent(x))

areNonExisting(x, envir = parent.frame(), inherits = FALSE,
  .xname = getNameInParent(x))
}
\arguments{
\item{x}{\code{character}.
Variable names to check in \code{environment}.}

\item{envir}{\code{environment}.
Environment.}

\item{inherits}{\code{logical(1)}.
Should the enclosing frames of the \code{environment} be searched?}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the requested input exist in the environment?
}
\note{
\code{exists} only supports \code{character(1)}, so we are exporting
\code{isExisting} as a convenience function to check multiple variables in a
single call.
}
\examples{
suppressWarnings(rm(x, y))
a <- 1L
b <- 2L

## Pass ====
isExisting(c("a", "b"))
areNonExisting(c("x", "y"))

## Fail ====
isExisting(c("x", "y"))
areNonExisting(c("a", "b"))
}
\seealso{
\code{assertive.code::is_existing()}.
}
