% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{extPattern}
\alias{extPattern}
\alias{compressExtPattern}
\title{File extension pattern matching}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
extPattern

compressExtPattern
}
\description{
File extension pattern matching
}
\note{
Updated 2020-01-04.
}
\examples{
extPattern
compressExtPattern
}
\keyword{datasets}
\keyword{internal}
