% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\name{extPattern}
\alias{extPattern}
\alias{compressExtPattern}
\title{File extension pattern matching}
\description{
File extension pattern matching
}
\note{
Updated 2019-10-21.
}
\examples{
extPattern
compressExtPattern
}
