% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-false.R
\name{false}
\alias{false}
\alias{falseFromVector}
\title{Return \code{FALSE} scalar with cause of failure}
\usage{
false(...)

falseFromVector(x)
}
\arguments{
\item{...}{Passed to \code{\link[base:sprintf]{gettextf()}} to create a \link{cause} of
failure message.}

\item{x}{Object.}
}
\value{
\code{goalie}/\code{logical(1L)}.
}
\description{
Always returns the value \code{FALSE}, with a \link{cause}
\link[base:attributes]{attribute}.
}
\note{
Updated 2019-07-29.
}
\seealso{
\code{assertive.base::false()}.
}
\keyword{internal}
