% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false.R, R/fixme-falseFromVector.R
\name{false}
\alias{false}
\alias{falseFromVector}
\title{Set a false goalie check with cause attribute}
\usage{
false(...)

falseFromVector(x)
}
\arguments{
\item{...}{Elements to pass to internal \code{\link[base:sprintf]{sprintf()}} call.}
}
\value{
\code{goalie}.
}
\description{
Set a false goalie check with cause attribute
}
\note{
Updated 2021-02-23.
}
\examples{
x <- false("'\%s' is invalid.", "xxx")
cause(x)
}
