% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-false.R
\name{false}
\alias{false}
\title{Return \code{FALSE} scalar with cause of failure}
\usage{
false(...)
}
\arguments{
\item{...}{Passed to \link[base:gettextf]{gettextf()} to create a \link{cause} of
failure message.}
}
\value{
\code{goalie}/\code{logical(1L)}.
}
\description{
Always returns the value \code{FALSE}, with a \code{cause}
\link[base:attributes]{attribute}.
}
\seealso{
\code{assertive.base::false()}.
}
