% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/false.R
\name{falseFromVector}
\alias{falseFromVector}
\alias{false}
\alias{falseFromVector,goalie}
\alias{falseFromVector,goalie-method}
\title{Set a false goalie check with cause attribute}
\usage{
falseFromVector(x, ...)

false(...)

`falseFromVector,goalie`(x)

\S4method{falseFromVector}{goalie}(x)
}
\arguments{
\item{...}{Elements to pass to internal \code{\link[base:sprintf]{sprintf()}} call.}
}
\value{
\code{goalie}.
}
\description{
Set a false goalie check with cause attribute
}
\note{
Updated 2021-02-23.
}
\examples{
x <- false("'\%s' is invalid.", "xxx")
print(x)
print(cause(x))

## falseFromVector ====
x <- goalie(
    object = c("aaa" = FALSE, "bbb" = TRUE, "ccc" = FALSE),
    cause = c("xxx", NA, "yyy")
)
print(x)
print(cause(x))
xx <- falseFromVector(x)
print(xx)
print(cause(xx))
}
