% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/falseFromVector-methods.R
\name{falseFromVector}
\alias{falseFromVector}
\alias{falseFromVector,goalie-method}
\title{Set a scalar false goalie check with cause attribute}
\usage{
falseFromVector(object, ...)

\S4method{falseFromVector}{goalie}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{goalie}.
}
\description{
Set a scalar false goalie check with cause attribute
}
\note{
Updated 2023-10-02.
}
\examples{
x <- goalie(
    object = c(FALSE, TRUE, FALSE),
    cause = c("xxx", NA, "yyy")
)
print(x)
print(cause(x))
xx <- falseFromVector(x)
print(xx)
print(cause(xx))
}
