% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formalCompress.R
\name{formalCompress}
\alias{formalCompress}
\title{Check the "compress" Formal Argument}
\usage{
formalCompress(compress)
}
\arguments{
\item{compress}{\code{logical(1)} or \code{character(1)}. These character strings are
currently allowed for \code{save}: \code{"gzip"}, \code{"bzip2"``, or }"xz"`.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop}) on failure.
}
}
\description{
Check the "compress" Formal Argument
}
\examples{
## Pass ====
formalCompress("gzip")
formalCompress(TRUE)

## Fail ====
formalCompress("xxx")
}
