% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-assertive-getNameInParent.R
\name{getNameInParent}
\alias{getNameInParent}
\alias{vname}
\title{Get the Name of a Variable in the Parent Frame}
\usage{
getNameInParent(x, escapePercent = TRUE)

vname(x, escapePercent = TRUE)
}
\arguments{
\item{x}{\code{symbol}.
Variable to get the name of.}

\item{escapePercent}{\code{logical(1)}.
If \code{TRUE}, percent signs are doubled, making the value suitable for use
with \code{sprintf}.}
}
\description{
Get the Name of a Variable in the Parent Frame
}
\examples{
getNameInParent(test)
}
\seealso{
\itemize{
\item \code{assertive.base::get_name_in_parent}.
\item \code{checkmate::vname}.
}
}
