% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNameInParent.R
\name{getNameInParent}
\alias{getNameInParent}
\title{Get the name of a variable defined in the parent frame}
\usage{
getNameInParent(x, escapePercent = TRUE, width = 100L)
}
\arguments{
\item{x}{\code{symbol}.
Variable to get the name of.}

\item{escapePercent}{\code{logical(1)}.
If \code{TRUE}, percent signs are doubled, making the value suitable for use
with \code{\link[base:sprintf]{sprintf()}}.}

\item{width}{\code{integer(1)}.
Maximum width of the name.}
}
\value{
\code{character(1)}.
}
\description{
Get the name of a variable defined in the parent frame
}
\note{
Updated 2019-09-06.
}
\examples{
getNameInParent(test)
}
\seealso{
\itemize{
\item \code{assertive.base::get_name_in_parent()}.
\item \code{checkmate::vname()}.
}
}
