% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{goalie-class}
\alias{goalie-class}
\title{goalie logical assert check return}
\value{
\code{goalie}.
}
\description{
Contains a \code{logical} with \code{cause} attributes.
}
\note{
Updated 2023-10-02.
}
