% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{goalie-package}
\alias{goalie}
\alias{goalie-package}
\title{goalie}
\description{
Assertive check functions for defensive R programming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://goalie.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/goalie/}
  \item Report bugs at \url{https://github.com/acidgenomics/goalie/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
