% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{goalie}
\alias{goalie}
\title{Define a goalie check classed return}
\usage{
goalie(object, cause)
}
\arguments{
\item{object}{\code{logical}.}

\item{cause}{\code{character}.
Corresponding cause attribute for logical \code{object}.
Note that \code{TRUE} values must contain \code{NA_character_} cause.
This value should not be named.}
}
\value{
\code{goalie}, which extends \code{logical}.
}
\description{
Define a goalie check classed return
}
\note{
Updated 2021-02-23.
}
\examples{
x <- goalie(
    object = c("aaa" = FALSE, "bbb" = TRUE),
    cause = c(
        "'aaa' is FALSE.",
        NA_character_
    )
)
print(x)
print(cause(x))
}
