% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-hasAccess.R
\name{hasAccess}
\alias{hasAccess}
\title{Does R Have File System Access Rights?}
\usage{
hasAccess(x, access = "r")
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{access}{[\code{character(1)}]\cr
Single string containing possible characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights, respectively.
Write and executable rights are not checked on Windows.}
}
\description{
Works for either file or directory paths.
}
\examples{
hasAccess("~")
}
\seealso{
\code{checkmate::testAccess}.
}
