% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasDims.R
\name{hasDims}
\alias{hasDims}
\alias{hasDimnames}
\alias{hasRows}
\alias{hasRownames}
\alias{hasCols}
\alias{hasColnames}
\title{Does the Input Have Dimensions?}
\usage{
hasDims(x, .xname = get_name_in_parent(x))

hasDimnames(x, .xname = get_name_in_parent(x))

hasRows(x, .xname = get_name_in_parent(x))

hasRownames(x, .xname = get_name_in_parent(x))

hasCols(x, .xname = get_name_in_parent(x))

hasColnames(x, .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Input Have Dimensions?
}
