% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasDims.R
\name{hasDims}
\alias{hasDims}
\alias{hasRows}
\alias{hasCols}
\alias{hasDimnames}
\alias{hasColnames}
\title{Does the Input Have Dimensions?}
\usage{
hasDims(x, .xname = get_name_in_parent(x))

hasRows(x, .xname = get_name_in_parent(x))

hasCols(x, .xname = get_name_in_parent(x))

hasDimnames(x, .xname = get_name_in_parent(x))

hasColnames(x, .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
Does the Input Have Dimensions?
}
\examples{
## Pass ====
x <- datasets::mtcars

hasDims(x)
hasRows(x)
hasCols(x)

hasDimnames(x)
hasRownames(x)
hasColnames(x)

## Note that dims don't have to be non-zero, just not NULL.
hasDims(data.frame())

## Fail ====
x <- data.frame()

hasDims(list())
hasRows(x)
hasCols(x)

hasDimnames(x)
hasRownames(x)
hasColnames(x)
}
