% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasElements.R
\name{hasElements}
\alias{hasElements}
\alias{isEmpty}
\alias{isNonEmpty}
\alias{isOfDimension}
\alias{isOfLength}
\title{Does the Input Have Elements?}
\usage{
isEmpty(x, metric = c("length", "elements"),
  .xname = get_name_in_parent(x))

isNonEmpty(x, metric = c("length", "elements"),
  .xname = get_name_in_parent(x))

hasElements(x, n, .xname = get_name_in_parent(x))

isOfDimension(x, n, .xname = get_name_in_parent(x))

isOfLength(x, n, .xname = get_name_in_parent(x))

isNonEmpty(x, metric = c("length", "elements"),
  .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{Object.}

\item{metric}{A string. Should be length or the number of elements be used to
determine if the object is empty/non-empty/scalar?}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{n}{Non-negative integer(s) of the expected length/number of elements/
lengths of dimensions.  See note.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Input Have Elements?
}
\examples{
## Pass ====
hasElements("hello")
hasElements(TRUE)
hasElements(list(a = 1, b = 2))

## Fail ====
hasElements(list())
}
