% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-flag-hasInternet.R
\name{hasInternet}
\alias{hasInternet}
\title{Does the current session have an internet connection?}
\usage{
hasInternet()
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the current session have an internet connection?
}
\examples{
hasInternet()
}
