% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasLength.R
\name{hasLength}
\alias{hasLength}
\title{Does the Input Have a Non-Zero or Defined Length?}
\usage{
hasLength(x, n = NULL)
}
\arguments{
\item{x}{Object.}

\item{n}{\code{NULL} or \code{integer}. If \code{NULL} (default), the function will
check to see if the input length is non-zero.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Input Have a Non-Zero or Defined Length?
}
\examples{
## Pass ====
hasLength(1L)
hasLength(FALSE)
hasLength(datasets::mtcars)
hasLength("")

## Fail ====
hasLength(NULL)
hasLength(character())
hasLength(data.frame())
}
