% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-hasLength.R
\name{hasLength}
\alias{hasLength}
\title{Does the input have a non-zero or defined length?}
\usage{
hasLength(x, n = NULL, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{n}{\code{NULL} or \code{integer}.
If \code{NULL} (default), the function will check to see if the input length is
non-zero.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input have a non-zero or defined length?
}
\examples{
## Pass ====
hasLength(1L)
hasLength(FALSE)
hasLength(datasets::mtcars)
hasLength("")

## Fail ====
hasLength(NULL)
hasLength(character())
hasLength(data.frame())
}
