% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasNames.R
\name{hasNames}
\alias{hasNames}
\title{Does the Input Have Names?}
\usage{
hasNames(x, .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop}) on failure.
}
}
\description{
Does the Input Have Names?
}
\examples{
x <- datasets::mtcars
hasNames(x)
}
