% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasRownames.R
\name{hasRownames}
\alias{hasRownames}
\alias{assertHasRownames}
\title{Has Rownames?}
\usage{
hasRownames(x)

assertHasRownames(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{base::stop()} or \code{base::stopifnot()} for details).
}
}
\description{
A stricter alternative to the assertive version that works properly with
data frames.
}
\note{
Also, \code{tibble::has_rownames()} appears to be more consistent than
\code{assertive.properties::has_rownames()} for \code{DataFrame} and \code{tbl_df} class.
}
\examples{
x <- S4Vectors::DataFrame(
    "sample1" = c(1L, 2L),
    "sample2" = c(3L, 4L),
    row.names = c("gene1", "gene2")
)
print(x)
hasRownames(x)
}
