% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/hasUniqueCols-methods.R
\docType{methods}
\name{hasUniqueCols}
\alias{hasUniqueCols}
\alias{has_unique_cols}
\alias{hasUniqueCols,ANY-method}
\alias{hasUniqueCols,sparseMatrix-method}
\alias{hasUniqueCols,SummarizedExperiment-method}
\alias{assertHasUniqueCols}
\title{Does the Object Contain Unique Columns?}
\usage{
hasUniqueCols(x, ...)

hasUniqueCols(x, ...)

\S4method{hasUniqueCols}{ANY}(x)

\S4method{hasUniqueCols}{sparseMatrix}(x)

\S4method{hasUniqueCols}{SummarizedExperiment}(x)

assertHasUniqueCols(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Checks a matrix for duplicated columns, which reprent samples. Duplicate
rows are allowed here, because many genes (rows) can contain all zeros.
}
\section{SummarizedExperiment}{


If the object is a \code{SummarizedExperiment}, then the primary \code{assay()} matrix
is checked for duplicated columns.
}

\examples{
## Unique.
x <- matrix(data = seq_len(20L), ncol = 2L)
hasUniqueCols(x)

## Duplicated.
x <- matrix(data = rep(seq_len(10L), times = 2L), ncol = 2L)
hasUniqueCols(x)
}
\seealso{
\itemize{
\item \code{checkmate::makeAssertionFunction()}.
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
