% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAll.R
\name{isAll}
\alias{isAll}
\title{Does the Input Belong to or Inherit All of These Classes?}
\usage{
isAll(x, classes)
}
\arguments{
\item{x}{Object.}

\item{classes}{\code{character}.
Object classes.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
Does the Input Belong to or Inherit All of These Classes?
}
\examples{
x <- 1L

## Pass ====
isAll(x, classes = c("integer", "numeric"))

## Fail ====
isAll(x, classes = c("integer", "NULL"))
}
