% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-flag-isAlpha.R
\name{isAlpha}
\alias{isAlpha}
\alias{containsAlpha}
\title{Does the input contain an alpha level?}
\usage{
isAlpha(x, .xname = getNameInParent(x))

containsAlpha(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
An alpha level must be \code{numeric(1)} greater than 0 and less than 1.
}
\examples{
## Pass ====
isAlpha(0.05)

## Fail ====
isAlpha("xxx")
isAlpha(1L)
}
