% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAny.R
\name{isAny}
\alias{isAny}
\title{Does the Object Belong to or Inherit Any of These Classes?}
\usage{
isAny(x, classes)
}
\arguments{
\item{x}{Object.}

\item{classes}{\code{character}. Object classes.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Object Belong to or Inherit Any of These Classes?
}
\examples{
x <- 1L
isAny(x, classes = c("integer", "character"))
isAny(x, classes = c("numeric", "character"))
}
