% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isAny.R
\name{isAny}
\alias{isAny}
\title{Does the object belong to or inherit any of these classes?}
\usage{
isAny(x, classes, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{classes}{\code{character}.
Object classes.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the object belong to or inherit any of these classes?
}
\examples{
x <- 1L

## Pass ====
isAny(x, classes = c("integer", "NULL"))
isAny(x, classes = c("numeric", "NULL"))
isAny(x, classes = c("atomic", "NULL"))

## Fail ====
isAny(x, classes = c("character", "data.frame"))
}
