% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDir.R
\name{isDir}
\alias{isDir}
\alias{assertIsDir}
\alias{areDirs}
\alias{assertAreDirs}
\title{Is Directory?}
\usage{
isDir(x)

assertIsDir(x)

areDirs(x)

assertAreDirs(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions  (see \code{stop()}for details).
}
}
\description{
Is Directory?
}
\examples{
x <- "~"

## Requires scalar.
isDir(x)

## Parameterized.
areDirs(rep(x, times = 2L))
}
