% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isDirectory.R
\name{isDirectory}
\alias{isDirectory}
\alias{isDir}
\alias{isADirectory}
\alias{isADir}
\title{Does the input contain a directory?}
\usage{
isDirectory(x)

isDir(x)

isADirectory(x)

isADir(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{logical}.
}
\description{
Does the input contain a directory?
}
\examples{
isDirectory(c("~", "~"))
isADirectory("~")
}
\seealso{
\code{R.utils::isDirectory()}.\preformatted{getS3method(
    f = "isDirectory",
    class = "default",
    envir = asNamespace("R.utils")
)
}
}
