% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDirectory.R
\name{isDirectory}
\alias{isDirectory}
\alias{isDir}
\alias{isADirectory}
\alias{isADir}
\title{Does the Argument Contain a Directory?}
\usage{
isDirectory(x)

isDir(x)

isADirectory(x)

isADir(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the Argument Contain a Directory?
}
\examples{
isADirectory("~")
}
