% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isDirectory.R
\name{isDirectory}
\alias{isDirectory}
\alias{isDir}
\alias{isADirectory}
\alias{isADir}
\alias{allAreDirectories}
\alias{allAreDirs}
\title{Does the input contain a directory?}
\usage{
isDirectory(x)

isDir(x)

isADirectory(x)

isADir(x)

allAreDirectories(x)

allAreDirs(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{logical}.
}
\description{
Does the input contain a directory?
}
\section{Functions}{
\itemize{
\item \code{isDirectory}: Vectorized.

\item \code{isDir}: Short alias for \code{\link[=isDirectory]{isDirectory()}}.

\item \code{isADirectory}: Scalar.

\item \code{isADir}: Short alias for \code{\link[=isADirectory]{isADirectory()}}.

\item \code{allAreDirectories}: Scalar variant.

\item \code{allAreDirs}: Short alias for \code{\link[=allAreDirectories]{allAreDirectories()}}.
}}

\examples{
isDirectory(c("~", "~"))
isADirectory("~")
}
\seealso{
\code{R.utils::isDirectory()}.\preformatted{getS3method(
    f = "isDirectory",
    class = "default",
    envir = asNamespace("R.utils")
)
}
}
