% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-isEqualTo.R
\name{isEqualTo}
\alias{isEqualTo}
\alias{isNotEqualTo}
\alias{isGreaterThan}
\alias{isGreaterThanOrEqualTo}
\alias{isLessThan}
\alias{isLessThanOrEqualTo}
\title{How Does the Input Relate to a Value?}
\usage{
isEqualTo(x, y, tol = 100 * .Machine$double.eps,
  .xname = get_name_in_parent(x), .yname = get_name_in_parent(x))

isNotEqualTo(x, y, tol = 100 * .Machine$double.eps,
  .xname = get_name_in_parent(x), .yname = get_name_in_parent(x))

isGreaterThan(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(x))

isGreaterThanOrEqualTo(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(x))

isLessThan(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(x))

isLessThanOrEqualTo(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(x))
}
\arguments{
\item{x}{Object.}

\item{y}{Secondary object.}

\item{tol}{Values within \code{tol} are considered equal.}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{.yname}{\emph{Not intended to be used directly.}}
}
\value{
\code{logical}.
}
\description{
How Does the Input Relate to a Value?
}
\note{
These functions return \code{logical}, not necessarily \code{logical(1)}.
}
\examples{
## Pass ====
isEqualTo(x = 1L, y = 1)
isNotEqualTo(x = 2, y = 1)
isGreaterThan(x = 1, y = 0)
isGreaterThanOrEqualTo(x = seq_len(2), y = 1)
isLessThan(x = -1, y = 0)
isLessThanOrEqualTo(x = seq_len(2), y = 3)

## Fail ====
isEqualTo(x = seq_len(2), y = 1)
}
