% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-isEqualTo.R
\name{isEqualTo}
\alias{isEqualTo}
\alias{isNotEqualTo}
\alias{isGreaterThan}
\alias{isGreaterThanOrEqualTo}
\alias{isLessThan}
\alias{isLessThanOrEqualTo}
\title{How Does the Input Relate to a Value?}
\usage{
isEqualTo(x, y)

isNotEqualTo(x, y)

isGreaterThan(x, y)

isGreaterThanOrEqualTo(x, y)

isLessThan(x, y)

isLessThanOrEqualTo(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Secondary object.}
}
\value{
\code{logical}.
}
\description{
How Does the Input Relate to a Value?
}
\note{
These functions return \code{logical}, not necessarily \code{logical(1)}.
}
\examples{
## Pass ====
isEqualTo(x = 1L, y = 1)
isNotEqualTo(x = 2, y = 1)
isGreaterThan(x = 1, y = 0)
isGreaterThanOrEqualTo(x = seq_len(2), y = 1)
isLessThan(x = -1, y = 0)
isLessThanOrEqualTo(x = seq_len(2), y = 3)

## Fail ====
isEqualTo(x = seq_len(2), y = 1)
}
\seealso{
\itemize{
\item Primitives: \code{==}, \code{>}, \code{>=}, \code{<}, \code{<=}.
\item \code{assertive.numbers::is_equal_to()}.
\item \code{assertive.numbers::is_not_equal_to()}.
\item \code{assertive.numbers::is_greater_than()}.
\item \code{assertive.numbers::is_greater_than_or_equal_to()}.
\item \code{assertive.numbers::is_less_than()}.
\item \code{assertive.numbers::is_less_than_or_equal_to()}.
}
}
