% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFile.R
\name{isFile}
\alias{isFile}
\title{Does the Argument Contain a Single (Existing) File?}
\usage{
isFile(x)
}
\arguments{
\item{x}{\code{character(1)}. String specifying a local file path that must exist
and be accessible.}
}
\description{
Does the Argument Contain a Single (Existing) File?
}
\examples{
x <- "example.txt"
file.create(x)
isFile(x)
unlink(x)
}
