% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFile.R
\name{isFile}
\alias{isFile}
\alias{isAFile}
\title{Does the Input Contain a File?}
\usage{
isFile(x)

isAFile(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{logical}.
}
\description{
Does the Input Contain a File?
}
\section{Functions}{
\itemize{
\item \code{isFile}: Supports multiple files.

\item \code{isAFile}: Check for a single file.
}}

\examples{
x <- "example.txt"
file.create(x)
isAFile(x)
unlink(x)
}
\seealso{
\code{R.utils::isFile()}.
}
