% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isFile.R
\name{isFile}
\alias{isFile}
\alias{isAFile}
\title{Does the input contain a file?}
\usage{
isFile(x)

isAFile(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{logical}.
}
\description{
Does the input contain a file?
}
\section{Functions}{
\itemize{
\item \code{isFile}: Vectorized.

\item \code{isAFile}: Scalar variant.
}}

\examples{
x <- "example.txt"
file.create(x)
isAFile(x)
unlink(x)
}
\seealso{
\code{R.utils::isFile()}.\preformatted{getS3method(
    f = "isFile",
    class = "default",
    envir = asNamespace("R.utils")
)
}
}
