% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-isFlag.R
\name{isFlag}
\alias{isFlag}
\title{Does the input contain a boolean flag?}
\usage{
isFlag(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input contain a boolean flag?
}
\examples{
## Pass ====
isFlag(TRUE)
isFlag(FALSE)

## Fail ====
isFlag(c(TRUE, TRUE))
isFlag(1)
isFlag(NA)
}
