% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isHeaderLevel.R
\name{isHeaderLevel}
\alias{isHeaderLevel}
\alias{assertIsHeaderLevel}
\title{Is Markdown Header Level?}
\usage{
isHeaderLevel(x)

assertIsHeaderLevel(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Markdown supports header levels 1-7 (\code{<H1>}-\code{<H7>}).
}
\examples{
isHeaderLevel(1L)
}
\seealso{
\itemize{
\item \code{checkmate::makeAssertionFunction()}.
}
}
