% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isScalar.R,
%   R/check-vector-isIntegerish.R
\name{isInt}
\alias{isInt}
\alias{isIntegerish}
\title{Is the input integer(ish)?}
\usage{
isInt(x)

isIntegerish(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Check for valid input of either explicit (e.g. \code{1L}) and/or implict
(e.g. \code{1}) \code{integer}.
}
\section{Functions}{
\itemize{
\item \code{isInt}: Alias for \code{\link[=isScalarIntegerish]{isScalarIntegerish()}}.
}}

\examples{
isIntegerish(seq_len(2L))
isIntegerish(c(1, 2))
}
\seealso{
\itemize{
\item \code{isInt()} or \code{isScalarIntegerish()} for scalar.
\item \code{rlang::is_integerish()}.
\item \code{checkmate::checkIntegerish()}.
}
}
