% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-isIntegerish.R
\name{isIntegerish}
\alias{isIntegerish}
\title{Is the Input Integerish?}
\usage{
isIntegerish(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Check for valid input of either explicit (e.g. \code{1L}) and/or implict
(e.g. \code{1}) \code{integer}.
}
\examples{
isIntegerish(seq_len(2L))
isIntegerish(c(1, 2))
}
\seealso{
\itemize{
\item \code{rlang::is_integerish()}.
\item \code{checkmate::checkIntegerish()}.
}
}
