% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-isPattern.R
\name{isPattern}
\alias{isPattern}
\alias{isMatchingFixed}
\alias{isNotMatchingFixed}
\alias{isMatchingRegex}
\alias{isNotMatchingRegex}
\title{Does the string match a pattern?}
\usage{
isMatchingFixed(x, pattern, .xname = getNameInParent(x))

isNotMatchingFixed(x, pattern, .xname = getNameInParent(x))

isMatchingRegex(x, pattern, .xname = getNameInParent(x))

isNotMatchingRegex(x, pattern, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{logical}.
}
\description{
Does the string match a pattern?
}
\examples{
isMatchingFixed(x = "foobar", pattern = "bar")
isNotMatchingFixed(x = "foo", pattern = "bar")
isMatchingRegex(x = "foobar", pattern = "^f")
isNotMatchingRegex(x = "foobar", pattern = "^b")
}
\seealso{
\itemize{
\item \code{assertive.strings::is_matching_fixed()}.
\item \code{assertive.strings::is_not_matching_fixed()}.
\item \code{assertive.strings::is_matching_regex()}.
\item \code{assertive.strings::is_not_matching_regex()}.
}
}
