% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPattern.R
\name{isPattern}
\alias{isPattern}
\alias{isMatchingFixed}
\alias{isNotMatchingFixed}
\alias{isMatchingRegex}
\alias{isNotMatchingRegex}
\title{Does the String Match a Pattern?}
\usage{
isMatchingFixed(x, pattern, opts_fixed = NULL,
  .xname = get_name_in_parent(x))

isNotMatchingFixed(x, pattern, opts_fixed = NULL,
  .xname = get_name_in_parent(x))

isMatchingRegex(x, pattern, opts_regex = NULL,
  .xname = get_name_in_parent(x))

isNotMatchingRegex(x, pattern, opts_regex = NULL,
  .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{Object.}

\item{pattern}{\code{character(1)}. Pattern to use for matching.}

\item{opts_fixed}{Passed to \code{\link[stringi]{stri_detect_fixed}}.}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{opts_regex}{Passed to \code{\link[stringi]{stri_detect_regex}}.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Does the String Match a Pattern?
}
