% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-isNumber.R,
%   R/check-scalar-isScalar.R
\name{isNumber}
\alias{isNumber}
\alias{isScalar}
\alias{isNonScalar}
\alias{isScalarList}
\alias{isScalarAtomic}
\alias{isScalarVector}
\alias{isScalarNumeric}
\alias{isScalarInteger}
\alias{isScalarIntegerish}
\alias{isScalarDouble}
\alias{isScalarCharacter}
\alias{isScalarLogical}
\title{Is the input scalar?}
\usage{
isNumber(x, nullOK = FALSE)

isScalar(x)

isNonScalar(x)

isScalarList(x)

isScalarAtomic(x)

isScalarVector(x)

isScalarNumeric(x)

isScalarInteger(x)

isScalarIntegerish(x)

isScalarDouble(x)

isScalarCharacter(x)

isScalarLogical(x)
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Scalar represents a length of 1.
}
\section{Functions}{
\itemize{
\item \code{isNumber}: Alias for \code{\link[=isScalarNumeric]{isScalarNumeric()}}.
}}

\examples{
## Pass ====
isScalar("a")
isScalarInteger(1L)
isScalarIntegerish(1)

## Fail ====
isScalar(NULL)
isScalar(c("a", "b"))
}
\seealso{
\code{help(topic = "scalar-type-predicates", package = "rlang")}.
\itemize{
\item \code{assertive.properties::is_scalar()}.
\item \code{rlang::is_scalar_list()}.
\item \code{rlang::is_scalar_atomic()}.
\item \code{rlang::is_scalar_vector()}.
\item \code{rlang::is_scalar_integer()}.
\item \code{rlang::is_scalar_integerish()}.
\item \code{rlang::is_scalar_double()}.
\item \code{rlang::is_scalar_character()}.
\item \code{rlang::is_scalar_logical()}.
}
}
