% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isString.R
\name{isString}
\alias{isString}
\title{Is the Argument a String?}
\usage{
isString(x, na.ok = FALSE, min.chars = NULL, pattern = NULL,
  fixed = NULL, ignore.case = FALSE, null.ok = FALSE)
}
\arguments{
\item{x}{Object.}

\item{pattern}{\code{character(1)}. Pattern to use for matching.}

\item{null.ok}{\code{logical(1)}. If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert()} function.
\item \code{check*}:
\code{TRUE} on success or \code{character} on failure.
Intended to be used in combination with \code{validate()} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop()}) on failure.
}
}
\description{
Is the Argument a String?
}
\examples{
## Pass ====
isString("a")

## Fail ====
isString(c("a", "b"))
}
