% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-isString.R
\name{isString}
\alias{isString}
\title{Does the input contain a (non-empty) character string?}
\usage{
isString(x, nullOK = FALSE, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Does the input contain a (non-empty) character string?
}
\examples{
## Pass ====
isString("hello")

## Fail ====
isString(1)
isString("")
isString(NA_character_)
}
