% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-vector-isURL.R
\name{isURL}
\alias{isURL}
\alias{isAURL}
\alias{containsURL}
\alias{containsAURL}
\title{Does the input contain a URL?}
\usage{
isURL(x, .xname = getNameInParent(x))

isAURL(x)
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{logical}.
}
\description{
Does the input contain a URL?
}
\section{Functions}{
\itemize{
\item \code{isURL}: Supports multiple URLs.

\item \code{isAURL}: Requires a single URL.
}}

\examples{
urls <- c("https://www.r-project.org", "ftp://r-project.org")

## Pass ====
isURL(urls)
isAURL(urls[[1L]])

## Fail ====
isURL("xxx")
isAURL(urls)
}
