% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-scalar-matchesUniqueGeneNames.R
\name{matchesUniqueGeneNames}
\alias{matchesUniqueGeneNames}
\title{Do the input gene names match unique values in the corresponding object?}
\usage{
matchesUniqueGeneNames(x, genes, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{.xname}{Name of object defined in \code{x}.
\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
This assert check determines if a user-defined gene name query is using only
unique (non-amgibuous) symbols. It is designed to be used for gene plotting
particularly when performing single-cell RNA-seq marker analysis.
}
\note{
Updated 2019-07-29.
}
\examples{
x <- SummarizedExperiment::SummarizedExperiment(
    assays = matrix(
        data = seq_len(16L),
        nrow = 4L,
        ncol = 4L,
        dimnames = list(
            paste0("gene", seq_len(4L)),
            paste0("sample", seq_len(4L))
        )
    ),
    rowData = S4Vectors::DataFrame(
        geneID = paste0("ENSG0000000000", seq_len(4L)),
        geneName = paste0("SYMBOL", seq_len(4L))
    )
)
genes <- SummarizedExperiment::rowData(x)$geneName

## TRUE ====
matchesUniqueGeneNames(x = x, genes = genes)
}
