% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{object}{Object.}

\item{envir}{\code{environment}.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}

\item{url}{\code{character}. Uniform Resource Locators (URLs). HTTP or FTP
addresses.}
}
\value{
Always stop on error.
}
\description{
Parameters
}
\keyword{internal}
