% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{envir}{\code{environment}. Environment.}

\item{genes}{\code{character}. Genes.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}

\item{names}{\code{character}. Names (e.g. rownames, colnames).}

\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{base::stop()} or \code{base::stopifnot()} for details).
}
}
\description{
Parameters
}
\keyword{internal}
