% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{add}{\code{AssertCollection}. Extra information to be included in the
message for the testthat reporter. See \code{testthat::expect_that()}.}

\item{classes}{\code{character}. Object classes.}

\item{envir}{\code{environment}. Environment.}

\item{genes}{\code{character}. Genes.}

\item{info}{\code{character}. Extra information to be included in the message for
the testthat reporter. See \code{testthat::expect_that()}.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}

\item{name}{\code{character}. Element name.}

\item{names}{\code{character}. Element names (e.g. row names, column names).}

\item{x}{Object.}

\item{.var.name}{\code{string}. Name of the checked object to print in assertions.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{boolean flag} for \code{are/has/is*()} functions.
\item Stop on error for \code{assert*()} functions
(see \code{stop()} or \code{stopifnot()} for details).
}
}
\description{
Parameters
}
\seealso{
\itemize{
\item \code{checkmate::makeTestFunction()}.
\item \code{checkmate::makeAssertionFunction()}.
\item \code{checkmate::makeExpectationFunction()}.
}
}
\keyword{internal}
