% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{dir}{\code{string}. Directory path.}

\item{envir}{\code{environment}. Environment.}

\item{file}{\code{string}. File path.}

\item{genes}{\code{character}. Genes.}

\item{inherits}{\code{boolean}. Should the enclosing frames of the \code{environment}
be searched?}

\item{names}{\code{character}. Names (e.g. rownames, colnames).}

\item{object}{Object.}

\item{url}{\code{character}. Uniform Resource Locators (URLs). HTTP or FTP
addresses.}
}
\value{
Always stop on error.
}
\description{
Parameters
}
\keyword{internal}
