% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{add}{\code{AssertCollection}.
Extra information to be included in the message for the testthat reporter.
See \code{testthat::expect_that()}.}

\item{classes}{\code{character}.
Object classes.}

\item{envir}{\code{environment}.
Environment.}

\item{genes}{\code{character}.
Genes.}

\item{info}{\code{character}.
Extra information to be included in the message for the testthat reporter.
See \code{testthat::expect_that()}.}

\item{inherits}{\code{logical(1)}.
Should the enclosing frames of the \code{environment} be searched?}

\item{label}{\code{character(1)}.
Name of the checked object to print in messages.}

\item{name}{\code{character}.
Element name.}

\item{names}{\code{character}.
Element names (e.g. row names, column names).}

\item{nullOK}{\code{logical(1)}.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}

\item{object}{Object.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{x}{Object.}

\item{y}{Secondary object.}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{.yname}{\emph{Not intended to be used directly.}}

\item{...}{Additional arguments.}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Parameters
}
\keyword{internal}
