% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print}
\alias{print}
\alias{print.goalie}
\title{Print methods for objects with a cause attribute}
\usage{
\method{print}{goalie}(x, ...)
}
\arguments{
\item{x}{an object of class \link{goalieScalar} or \link{goalieVector}.}

\item{...}{\emph{Currently unused.}}
}
\description{
Print methods for objects with a cause attribute
}
