% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print}
\alias{print}
\alias{print.goalie}
\title{Print methods for objects with a cause attribute}
\usage{
\method{print}{goalie}(x, n = 10L, ignoreNA = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{n}{\code{integer(1)}.
Number to include.}

\item{ignoreNA}{\code{logical(1)}.
Should NA values pass (\code{TRUE}) or fail (\code{FALSE})?
\itemize{
\item \code{TRUE}: OK can be \code{TRUE} or \code{NA}; \code{FALSE} is bad.
\item \code{FALSE}: OK can be \code{TRUE}; \code{FALSE} or \code{NA} is bad.
}}

\item{...}{Additional arguments.}
}
\value{
Print command and return invisibly.
}
\description{
Print methods for objects with a cause attribute
}
\note{
Updated 2019-08-05.
}
