% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietly.R
\name{quietly}
\alias{quietly}
\title{Perform an action quietly}
\usage{
quietly(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
Invisible \code{NULL}.
}
\description{
Suppress all warnings, messages, and console output.
}
\note{
Updated 2023-04-26.
}
\examples{
quietly({
    message("hello world")
    object <- c("aaa", "bbb")
})
print(object)
}
