% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietly.R
\name{quietly}
\alias{quietly}
\title{Perform an action quietly}
\usage{
quietly(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
Invisible \code{NULL}.
}
\description{
Suppress all warnings, messages, and console output.
}
\note{
Updated 2023-11-27.
}
\examples{
quietly({
    message("hello world")
    object <- c("aaa", "bbb")
})
print(object)
}
\seealso{
\itemize{
\item \code{help("conditions")}
\item \code{sink()} and \code{nullfile()}.
\item \code{utils::capture.output()}.
\item https://stackoverflow.com/questions/2723034/
\item https://stackoverflow.com/questions/18730491/
\item https://stackoverflow.com/questions/25320381/
\item https://stackoverflow.com/questions/4948361/
}
}
