% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ranges.R
\name{ranges}
\alias{ranges}
\alias{isInRange}
\alias{isInClosedRange}
\alias{isInOpenRange}
\alias{isInLeftOpenRange}
\alias{isInRightOpenRange}
\alias{isNegative}
\alias{isNonNegative}
\alias{isPositive}
\alias{isNonPositive}
\alias{isPercentage}
\alias{isProportion}
\title{Is the Input in Range?}
\usage{
isInRange(x, lower = -Inf, upper = Inf, lower_is_strict = FALSE,
  upper_is_strict = FALSE, .xname = get_name_in_parent(x))

isInClosedRange(x, lower = -Inf, upper = Inf,
  .xname = get_name_in_parent(x))

isInOpenRange(x, lower = -Inf, upper = Inf,
  .xname = get_name_in_parent(x))

isInLeftOpenRange(x, lower = -Inf, upper = Inf,
  .xname = get_name_in_parent(x))

isInRightOpenRange(x, lower = -Inf, upper = Inf,
  .xname = get_name_in_parent(x))

isNegative(x, .xname = get_name_in_parent(x))

isNonNegative(x, .xname = get_name_in_parent(x))

isPositive(x, .xname = get_name_in_parent(x))

isNonPositive(x, .xname = get_name_in_parent(x))

isPercentage(x, lower_is_strict = FALSE, upper_is_strict = FALSE,
  .xname = get_name_in_parent(x))

isProportion(x, lower_is_strict = FALSE, upper_is_strict = FALSE,
  .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{Object.}

\item{lower}{Lower bound for the interval.}

\item{upper}{Upper bound for the interval.}

\item{lower_is_strict}{If \code{TRUE}, the lower bound is open (strict) 
otherwise it is closed.}

\item{upper_is_strict}{If \code{TRUE}, the upper bound is open (strict)
otherwise it is closed.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
Is the Input in Range?
}
\examples{
## Pass ====
isInRange(x = 1, lower = 0, upper = 1)

isInClosedRange(x = 0, lower = 0, upper = 1)
isInClosedRange(x = 1, lower = 0, upper = 1)

isInOpenRange(x = 0.5, lower = 0, upper = 1)

isInLeftOpenRange(x = 1, lower = 0, upper = 1)
isInRightOpenRange(x = 0, lower = 0, upper = 1)

## Fail ====
isInRange(x = 2, lower = 0, upper = 1)

isInOpenRange(x = 0, lower = 0, upper = 1)
isInOpenRange(x = 1, lower = 0, upper = 1)

isInLeftOpenRange(x = 0, lower = 0, upper = 1)
isInRightOpenRange(x = 1, lower = 0, upper = 1)
}
