% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/verbs-has.R, R/verbs-is.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assert_that}
\alias{validate_that}
\alias{\%>\%}
\alias{hasName}
\alias{has_name}
\alias{hasNames}
\alias{has_names}
\alias{is_empty_file}
\alias{isEmptyFile}
\alias{is_existing}
\alias{isExisting}
\alias{is_integerish}
\alias{isIntegerish}
\alias{is_negative}
\alias{isNegative}
\alias{is_non_empty_file}
\alias{isNonEmptyFile}
\alias{is_not_na}
\alias{isNotNA}
\alias{is_positive}
\alias{isPositive}
\alias{is_scalar_double}
\alias{isScalarDouble}
\alias{is_scalar_integerish}
\alias{isScalarIntegerish}
\alias{is_string}
\alias{isString}
\title{Objects exported from other packages}
\section{Functions}{
\itemize{
\item \code{hasNames}: \code{checkmate::testNames}.

\item \code{has_names}: \code{hasNames} snake alias.

\item \code{isEmptyFile}: \code{is_empty_file} camel alias.
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{assertive.base}{\code{\link[assertive.base]{is_not_na}}}

  \item{assertive.code}{\code{\link[assertive.code]{is_existing}}}

  \item{assertive.files}{\code{\link[assertive.files]{is_empty_file}}, \code{\link[assertive.files]{is_non_empty_file}}}

  \item{assertive.numbers}{\code{\link[assertive.numbers]{is_negative}}, \code{\link[assertive.numbers]{is_positive}}}

  \item{assertthat}{\code{\link[assertthat]{assert_that}}, \code{\link[assertthat]{validate_that}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{is_integerish}}, \code{\link[rlang]{is_scalar_double}}, \code{\link[rlang]{is_scalar_integerish}}, \code{\link[rlang]{is_string}}}

  \item{utils}{\code{\link[utils]{hasName}}}
}}

