% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_verbs.R, R/is_verbs.R, R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{has_name}
\alias{hasName}
\alias{has_names}
\alias{hasNames}
\alias{is_scalar_double}
\alias{isScalarDouble}
\alias{is_string}
\alias{isString}
\alias{is_integerish}
\alias{isIntegerish}
\alias{is_scalar_integerish}
\alias{isScalarIntegerish}
\alias{assert_that}
\alias{validate_that}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{assertive.properties}{\code{\link[assertive.properties]{has_names}}}

  \item{assertthat}{\code{\link[assertthat]{assert_that}}, \code{\link[assertthat]{validate_that}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{has_name}}, \code{\link[rlang]{is_scalar_double}}, \code{\link[rlang]{is_string}}, \code{\link[rlang]{is_integerish}}, \code{\link[rlang]{is_scalar_integerish}}}
}}

