% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-assertthat.R,
%   R/reexports-checkmate.R, R/reexports-has_verbs.R, R/reexports-is_verbs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assert_that}
\alias{validate_that}
\alias{validate}
\alias{assert}
\alias{assertClass}
\alias{checkClass}
\alias{testClass}
\alias{assertDirectoryExists}
\alias{checkDirectoryExists}
\alias{testDirectoryExists}
\alias{assertFactor}
\alias{checkFactor}
\alias{testFactor}
\alias{assertFileExists}
\alias{checkFileExists}
\alias{testFileExists}
\alias{assertFlag}
\alias{checkFlag}
\alias{testFlag}
\alias{assertLogical}
\alias{checkLogical}
\alias{testLogical}
\alias{assertNames}
\alias{checkNames}
\alias{testNames}
\alias{assertScalar}
\alias{checkScalar}
\alias{testScalar}
\alias{assertString}
\alias{checkString}
\alias{testString}
\alias{assert_has_cols}
\alias{assertHasCols}
\alias{has_cols}
\alias{hasCols}
\alias{assert_has_dimnames}
\alias{assertHasDimnames}
\alias{has_dimnames}
\alias{hasDimnames}
\alias{assert_has_dims}
\alias{assertHasDims}
\alias{has_dims}
\alias{hasDims}
\alias{hasName}
\alias{has_name}
\alias{assert_has_names}
\alias{assertHasNames}
\alias{has_names}
\alias{hasNames}
\alias{assert_has_rows}
\alias{assertHasRows}
\alias{has_rows}
\alias{hasRows}
\alias{is_empty_file}
\alias{isEmptyFile}
\alias{is_existing}
\alias{isExisting}
\alias{is_integerish}
\alias{isIntegerish}
\alias{is_negative}
\alias{isNegative}
\alias{is_non_empty_file}
\alias{isNonEmptyFile}
\alias{is_not_na}
\alias{isNotNA}
\alias{is_positive}
\alias{isPositive}
\alias{is_scalar_double}
\alias{isScalarDouble}
\alias{is_scalar_integerish}
\alias{isScalarIntegerish}
\alias{is_string}
\alias{isString}
\title{Objects exported from other packages}
\usage{
validate(..., env = parent.frame(), msg = NULL)
}
\section{Functions}{
\itemize{
\item \code{validate}: Shorter name variant of \code{validate_that}.

\item \code{assertHasCols}: camel alias.

\item \code{hasCols}: camel alias.

\item \code{assertHasDimnames}: camel alias.

\item \code{hasDimnames}: camel alias.

\item \code{assertHasDims}: camel alias.

\item \code{hasDims}: camel alias.

\item \code{has_name}: snake alias.

\item \code{assertHasNames}: camel alias.

\item \code{hasNames}: camel alias.

\item \code{assertHasRows}: camel alias.

\item \code{hasRows}: camel alias.

\item \code{isEmptyFile}: camel alias.

\item \code{isExisting}: camel alias.

\item \code{isIntegerish}: camel alias.

\item \code{isNegative}: camel alias.

\item \code{isNonEmptyFile}: camel alias.

\item \code{isNotNA}: camel alias.

\item \code{isPositive}: camel alias.

\item \code{isScalarDouble}: camel alias.

\item \code{isScalarIntegerish}: camel alias.

\item \code{isString}: camel alias.
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{assertive.base}{\code{\link[assertive.base]{is_not_na}}}

  \item{assertive.code}{\code{\link[assertive.code]{is_existing}}}

  \item{assertive.files}{\code{\link[assertive.files]{is_empty_file}}, \code{\link[assertive.files]{is_non_empty_file}}}

  \item{assertive.numbers}{\code{\link[assertive.numbers]{is_negative}}, \code{\link[assertive.numbers]{is_positive}}}

  \item{assertive.properties}{\code{\link[assertive.properties]{assert_has_cols}}, \code{\link[assertive.properties]{has_cols}}, \code{\link[assertive.properties]{assert_has_dimnames}}, \code{\link[assertive.properties]{has_dimnames}}, \code{\link[assertive.properties]{assert_has_dims}}, \code{\link[assertive.properties]{has_dims}}, \code{\link[assertive.properties]{assert_has_names}}, \code{\link[assertive.properties]{has_names}}, \code{\link[assertive.properties]{assert_has_rows}}, \code{\link[assertive.properties]{has_rows}}}

  \item{assertthat}{\code{\link[assertthat]{assert_that}}, \code{\link[assertthat]{validate_that}}}

  \item{checkmate}{\code{\link[checkmate]{assert}}, \code{\link[checkmate]{assertClass}}, \code{\link[checkmate]{checkClass}}, \code{\link[checkmate]{testClass}}, \code{\link[checkmate]{assertDirectoryExists}}, \code{\link[checkmate]{checkDirectoryExists}}, \code{\link[checkmate]{testDirectoryExists}}, \code{\link[checkmate]{assertFactor}}, \code{\link[checkmate]{checkFactor}}, \code{\link[checkmate]{testFactor}}, \code{\link[checkmate]{assertFileExists}}, \code{\link[checkmate]{checkFileExists}}, \code{\link[checkmate]{testFileExists}}, \code{\link[checkmate]{assertFlag}}, \code{\link[checkmate]{checkFlag}}, \code{\link[checkmate]{testFlag}}, \code{\link[checkmate]{assertLogical}}, \code{\link[checkmate]{checkLogical}}, \code{\link[checkmate]{testLogical}}, \code{\link[checkmate]{assertNames}}, \code{\link[checkmate]{checkNames}}, \code{\link[checkmate]{testNames}}, \code{\link[checkmate]{assertScalar}}, \code{\link[checkmate]{checkScalar}}, \code{\link[checkmate]{testScalar}}, \code{\link[checkmate]{assertString}}, \code{\link[checkmate]{checkString}}, \code{\link[checkmate]{testString}}}

  \item{rlang}{\code{\link[rlang]{is_integerish}}, \code{\link[rlang]{is_scalar_double}}, \code{\link[rlang]{is_scalar_integerish}}, \code{\link[rlang]{is_string}}}

  \item{utils}{\code{\link[utils]{hasName}}}
}}

