% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAtomic.R, R/checkCharacter.R,
%   R/checkClass.R, R/checkDataFrame.R, R/checkDirectoryExists.R,
%   R/checkEnvironment.R, R/checkFactor.R, R/checkFileExists.R, R/checkFlag.R,
%   R/checkList.R, R/checkLogical.R, R/checkMultiClass.R, R/checkNames.R,
%   R/checkScalar.R, R/checkSetEqual.R, R/checkString.R, R/checkSubset.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{checkAtomic}
\alias{check_atomic}
\alias{testAtomic}
\alias{test_atomic}
\alias{assertAtomic}
\alias{assert_atomic}
\alias{expect_atomic}
\alias{checkCharacter}
\alias{check_character}
\alias{testCharacter}
\alias{test_character}
\alias{assertCharacter}
\alias{assert_character}
\alias{expect_character}
\alias{checkClass}
\alias{check_class}
\alias{testClass}
\alias{test_class}
\alias{assertClass}
\alias{assert_class}
\alias{expect_class}
\alias{checkDataFrame}
\alias{check_data_frame}
\alias{testDataFrame}
\alias{test_data_frame}
\alias{assertDataFrame}
\alias{assert_data_frame}
\alias{expect_data_frame}
\alias{checkDirectoryExists}
\alias{check_directory_exists}
\alias{testDirectoryExists}
\alias{test_directory_exists}
\alias{assertDirectoryExists}
\alias{assert_directory_exists}
\alias{expect_directory_exists}
\alias{checkEnvironment}
\alias{check_environment}
\alias{testEnvironment}
\alias{test_environment}
\alias{assertEnvironment}
\alias{assert_environment}
\alias{expect_environment}
\alias{checkFactor}
\alias{check_factor}
\alias{testFactor}
\alias{test_factor}
\alias{assertFactor}
\alias{assert_factor}
\alias{expect_factor}
\alias{checkFileExists}
\alias{check_file_exists}
\alias{testFileExists}
\alias{test_file_exists}
\alias{assertFileExists}
\alias{assert_file_exists}
\alias{expect_file_exists}
\alias{checkFlag}
\alias{check_flag}
\alias{testFlag}
\alias{test_flag}
\alias{assertFlag}
\alias{assert_flag}
\alias{expect_flag}
\alias{checkList}
\alias{check_list}
\alias{testList}
\alias{test_list}
\alias{assertList}
\alias{assert_list}
\alias{expect_list}
\alias{checkLogical}
\alias{check_logical}
\alias{testLogical}
\alias{test_logical}
\alias{assertLogical}
\alias{assert_logical}
\alias{expect_logical}
\alias{checkMultiClass}
\alias{check_multi_class}
\alias{testMultiClass}
\alias{test_multi_class}
\alias{assertMultiClass}
\alias{assert_multi_class}
\alias{expect_multi_class}
\alias{checkNames}
\alias{check_names}
\alias{testNames}
\alias{test_names}
\alias{assertNames}
\alias{assert_names}
\alias{expect_names}
\alias{checkScalar}
\alias{check_scalar}
\alias{testScalar}
\alias{test_scalar}
\alias{assertScalar}
\alias{assert_scalar}
\alias{expect_scalar}
\alias{checkSetEqual}
\alias{check_set_equal}
\alias{testSetEqual}
\alias{test_set_equal}
\alias{assertSetEqual}
\alias{assert_set_equal}
\alias{expect_set_equal}
\alias{checkString}
\alias{check_string}
\alias{testString}
\alias{test_string}
\alias{assertString}
\alias{assert_string}
\alias{expect_string}
\alias{checkSubset}
\alias{check_subset}
\alias{testSubset}
\alias{test_subset}
\alias{assertSubset}
\alias{assert_subset}
\alias{expect_subset}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{checkmate}{\code{\link[checkmate]{checkAtomic}}, \code{\link[checkmate]{check_atomic}}, \code{\link[checkmate]{testAtomic}}, \code{\link[checkmate]{test_atomic}}, \code{\link[checkmate]{assertAtomic}}, \code{\link[checkmate]{assert_atomic}}, \code{\link[checkmate]{expect_atomic}}, \code{\link[checkmate]{checkCharacter}}, \code{\link[checkmate]{check_character}}, \code{\link[checkmate]{testCharacter}}, \code{\link[checkmate]{test_character}}, \code{\link[checkmate]{assertCharacter}}, \code{\link[checkmate]{assert_character}}, \code{\link[checkmate]{expect_character}}, \code{\link[checkmate]{checkClass}}, \code{\link[checkmate]{check_class}}, \code{\link[checkmate]{testClass}}, \code{\link[checkmate]{test_class}}, \code{\link[checkmate]{assertClass}}, \code{\link[checkmate]{assert_class}}, \code{\link[checkmate]{expect_class}}, \code{\link[checkmate]{checkDataFrame}}, \code{\link[checkmate]{check_data_frame}}, \code{\link[checkmate]{testDataFrame}}, \code{\link[checkmate]{test_data_frame}}, \code{\link[checkmate]{assertDataFrame}}, \code{\link[checkmate]{assert_data_frame}}, \code{\link[checkmate]{expect_data_frame}}, \code{\link[checkmate]{checkDirectoryExists}}, \code{\link[checkmate]{check_directory_exists}}, \code{\link[checkmate]{testDirectoryExists}}, \code{\link[checkmate]{test_directory_exists}}, \code{\link[checkmate]{assertDirectoryExists}}, \code{\link[checkmate]{assert_directory_exists}}, \code{\link[checkmate]{expect_directory_exists}}, \code{\link[checkmate]{checkEnvironment}}, \code{\link[checkmate]{check_environment}}, \code{\link[checkmate]{testEnvironment}}, \code{\link[checkmate]{test_environment}}, \code{\link[checkmate]{assertEnvironment}}, \code{\link[checkmate]{assert_environment}}, \code{\link[checkmate]{expect_environment}}, \code{\link[checkmate]{checkFactor}}, \code{\link[checkmate]{check_factor}}, \code{\link[checkmate]{testFactor}}, \code{\link[checkmate]{test_factor}}, \code{\link[checkmate]{assertFactor}}, \code{\link[checkmate]{assert_factor}}, \code{\link[checkmate]{expect_factor}}, \code{\link[checkmate]{checkFileExists}}, \code{\link[checkmate]{check_file_exists}}, \code{\link[checkmate]{testFileExists}}, \code{\link[checkmate]{test_file_exists}}, \code{\link[checkmate]{assertFileExists}}, \code{\link[checkmate]{assert_file_exists}}, \code{\link[checkmate]{expect_file_exists}}, \code{\link[checkmate]{checkFlag}}, \code{\link[checkmate]{check_flag}}, \code{\link[checkmate]{testFlag}}, \code{\link[checkmate]{test_flag}}, \code{\link[checkmate]{assertFlag}}, \code{\link[checkmate]{assert_flag}}, \code{\link[checkmate]{expect_flag}}, \code{\link[checkmate]{checkList}}, \code{\link[checkmate]{check_list}}, \code{\link[checkmate]{testList}}, \code{\link[checkmate]{test_list}}, \code{\link[checkmate]{assertList}}, \code{\link[checkmate]{assert_list}}, \code{\link[checkmate]{expect_list}}, \code{\link[checkmate]{checkLogical}}, \code{\link[checkmate]{check_logical}}, \code{\link[checkmate]{testLogical}}, \code{\link[checkmate]{test_logical}}, \code{\link[checkmate]{assertLogical}}, \code{\link[checkmate]{assert_logical}}, \code{\link[checkmate]{expect_logical}}, \code{\link[checkmate]{checkMultiClass}}, \code{\link[checkmate]{check_multi_class}}, \code{\link[checkmate]{testMultiClass}}, \code{\link[checkmate]{test_multi_class}}, \code{\link[checkmate]{assertMultiClass}}, \code{\link[checkmate]{assert_multi_class}}, \code{\link[checkmate]{expect_multi_class}}, \code{\link[checkmate]{checkNames}}, \code{\link[checkmate]{check_names}}, \code{\link[checkmate]{testNames}}, \code{\link[checkmate]{test_names}}, \code{\link[checkmate]{assertNames}}, \code{\link[checkmate]{assert_names}}, \code{\link[checkmate]{expect_names}}, \code{\link[checkmate]{checkScalar}}, \code{\link[checkmate]{check_scalar}}, \code{\link[checkmate]{testScalar}}, \code{\link[checkmate]{test_scalar}}, \code{\link[checkmate]{assertScalar}}, \code{\link[checkmate]{assert_scalar}}, \code{\link[checkmate]{expect_scalar}}, \code{\link[checkmate]{checkSetEqual}}, \code{\link[checkmate]{check_set_equal}}, \code{\link[checkmate]{testSetEqual}}, \code{\link[checkmate]{test_set_equal}}, \code{\link[checkmate]{assertSetEqual}}, \code{\link[checkmate]{assert_set_equal}}, \code{\link[checkmate]{expect_set_equal}}, \code{\link[checkmate]{checkString}}, \code{\link[checkmate]{check_string}}, \code{\link[checkmate]{testString}}, \code{\link[checkmate]{test_string}}, \code{\link[checkmate]{assertString}}, \code{\link[checkmate]{assert_string}}, \code{\link[checkmate]{expect_string}}, \code{\link[checkmate]{checkSubset}}, \code{\link[checkmate]{check_subset}}, \code{\link[checkmate]{testSubset}}, \code{\link[checkmate]{test_subset}}, \code{\link[checkmate]{assertSubset}}, \code{\link[checkmate]{assert_subset}}, \code{\link[checkmate]{expect_subset}}}
}}

