% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertive-assertAreDisjointSets.R,
%   R/assertive-assertHasNames.R, R/assertive-assertHasNoDuplicates.R,
%   R/checkmate-checkAtomic.R, R/checkmate-checkCharacter.R,
%   R/checkmate-checkClass.R, R/checkmate-checkDataFrame.R,
%   R/checkmate-checkDataTable.R, R/checkmate-checkDirectoryExists.R,
%   R/checkmate-checkEnvironment.R, R/checkmate-checkFactor.R,
%   R/checkmate-checkFileExists.R, R/checkmate-checkFlag.R,
%   R/checkmate-checkFunction.R, R/checkmate-checkInt.R,
%   R/checkmate-checkInteger.R, R/checkmate-checkList.R,
%   R/checkmate-checkLogical.R, R/checkmate-checkMatrix.R,
%   R/checkmate-checkMultiClass.R, R/checkmate-checkNames.R,
%   R/checkmate-checkNull.R, R/checkmate-checkScalar.R,
%   R/checkmate-checkSetEqual.R, R/checkmate-checkString.R,
%   R/checkmate-checkSubset.R, R/checkmate-checkTibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{are_disjoint_sets}
\alias{areDisjointSets}
\alias{assert_are_disjoint_sets}
\alias{assertAreDisjointSets}
\alias{has_names}
\alias{hasNames}
\alias{assert_has_names}
\alias{assertHasNames}
\alias{has_no_duplicates}
\alias{hasNoDuplicates}
\alias{assert_has_no_duplicates}
\alias{assertHasNoDuplicates}
\alias{checkAtomic}
\alias{check_atomic}
\alias{testAtomic}
\alias{test_atomic}
\alias{assertAtomic}
\alias{assert_atomic}
\alias{expect_atomic}
\alias{checkCharacter}
\alias{check_character}
\alias{testCharacter}
\alias{test_character}
\alias{assertCharacter}
\alias{assert_character}
\alias{expect_character}
\alias{checkClass}
\alias{check_class}
\alias{testClass}
\alias{test_class}
\alias{assertClass}
\alias{assert_class}
\alias{expect_class}
\alias{checkDataFrame}
\alias{check_data_frame}
\alias{testDataFrame}
\alias{test_data_frame}
\alias{assertDataFrame}
\alias{assert_data_frame}
\alias{expect_data_frame}
\alias{checkDataTable}
\alias{check_data_table}
\alias{testDataTable}
\alias{test_data_table}
\alias{assertDataTable}
\alias{assert_data_table}
\alias{expect_data_table}
\alias{checkDirectoryExists}
\alias{check_directory_exists}
\alias{testDirectoryExists}
\alias{test_directory_exists}
\alias{assertDirectoryExists}
\alias{assert_directory_exists}
\alias{expect_directory_exists}
\alias{checkEnvironment}
\alias{check_environment}
\alias{testEnvironment}
\alias{test_environment}
\alias{assertEnvironment}
\alias{assert_environment}
\alias{expect_environment}
\alias{checkFactor}
\alias{check_factor}
\alias{testFactor}
\alias{test_factor}
\alias{assertFactor}
\alias{assert_factor}
\alias{expect_factor}
\alias{checkFileExists}
\alias{check_file_exists}
\alias{testFileExists}
\alias{test_file_exists}
\alias{assertFileExists}
\alias{assert_file_exists}
\alias{expect_file_exists}
\alias{checkFlag}
\alias{check_flag}
\alias{testFlag}
\alias{test_flag}
\alias{assertFlag}
\alias{assert_flag}
\alias{expect_flag}
\alias{checkFunction}
\alias{check_function}
\alias{testFunction}
\alias{test_function}
\alias{assertFunction}
\alias{assert_function}
\alias{expect_function}
\alias{checkInt}
\alias{check_int}
\alias{testInt}
\alias{test_int}
\alias{assertInt}
\alias{assert_int}
\alias{expect_int}
\alias{checkInteger}
\alias{check_integer}
\alias{testInteger}
\alias{test_integer}
\alias{assertInteger}
\alias{assert_integer}
\alias{expect_integer}
\alias{checkList}
\alias{check_list}
\alias{testList}
\alias{test_list}
\alias{assertList}
\alias{assert_list}
\alias{expect_list}
\alias{checkLogical}
\alias{check_logical}
\alias{testLogical}
\alias{test_logical}
\alias{assertLogical}
\alias{assert_logical}
\alias{expect_logical}
\alias{checkMatrix}
\alias{check_matrix}
\alias{testMatrix}
\alias{test_matrix}
\alias{assertMatrix}
\alias{assert_matrix}
\alias{expect_matrix}
\alias{checkMultiClass}
\alias{check_multi_class}
\alias{testMultiClass}
\alias{test_multi_class}
\alias{assertMultiClass}
\alias{assert_multi_class}
\alias{expect_multi_class}
\alias{checkNames}
\alias{check_names}
\alias{testNames}
\alias{test_names}
\alias{assertNames}
\alias{assert_names}
\alias{expect_names}
\alias{checkNull}
\alias{check_null}
\alias{testNull}
\alias{test_null}
\alias{assertNull}
\alias{assert_null}
\alias{checkScalar}
\alias{check_scalar}
\alias{testScalar}
\alias{test_scalar}
\alias{assertScalar}
\alias{assert_scalar}
\alias{expect_scalar}
\alias{checkSetEqual}
\alias{check_set_equal}
\alias{testSetEqual}
\alias{test_set_equal}
\alias{assertSetEqual}
\alias{assert_set_equal}
\alias{expect_set_equal}
\alias{checkString}
\alias{check_string}
\alias{testString}
\alias{test_string}
\alias{assertString}
\alias{assert_string}
\alias{expect_string}
\alias{checkSubset}
\alias{check_subset}
\alias{testSubset}
\alias{test_subset}
\alias{assertSubset}
\alias{assert_subset}
\alias{expect_subset}
\alias{checkTibble}
\alias{check_tibble}
\alias{testTibble}
\alias{test_tibble}
\alias{assertTibble}
\alias{assert_tibble}
\alias{expect_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{assertive}{\code{\link[assertive]{are_disjoint_sets}}, \code{\link[assertive]{assert_are_disjoint_sets}}, \code{\link[assertive]{has_names}}, \code{\link[assertive]{assert_has_names}}}

  \item{assertive.properties}{\code{\link[assertive.properties]{has_no_duplicates}}, \code{\link[assertive.properties]{assert_has_no_duplicates}}}

  \item{checkmate}{\code{\link[checkmate]{checkAtomic}}, \code{\link[checkmate]{check_atomic}}, \code{\link[checkmate]{testAtomic}}, \code{\link[checkmate]{test_atomic}}, \code{\link[checkmate]{assertAtomic}}, \code{\link[checkmate]{assert_atomic}}, \code{\link[checkmate]{expect_atomic}}, \code{\link[checkmate]{checkCharacter}}, \code{\link[checkmate]{check_character}}, \code{\link[checkmate]{testCharacter}}, \code{\link[checkmate]{test_character}}, \code{\link[checkmate]{assertCharacter}}, \code{\link[checkmate]{assert_character}}, \code{\link[checkmate]{expect_character}}, \code{\link[checkmate]{checkClass}}, \code{\link[checkmate]{check_class}}, \code{\link[checkmate]{testClass}}, \code{\link[checkmate]{test_class}}, \code{\link[checkmate]{assertClass}}, \code{\link[checkmate]{assert_class}}, \code{\link[checkmate]{expect_class}}, \code{\link[checkmate]{checkDataFrame}}, \code{\link[checkmate]{check_data_frame}}, \code{\link[checkmate]{testDataFrame}}, \code{\link[checkmate]{test_data_frame}}, \code{\link[checkmate]{assertDataFrame}}, \code{\link[checkmate]{assert_data_frame}}, \code{\link[checkmate]{expect_data_frame}}, \code{\link[checkmate]{checkDataTable}}, \code{\link[checkmate]{check_data_table}}, \code{\link[checkmate]{testDataTable}}, \code{\link[checkmate]{test_data_table}}, \code{\link[checkmate]{assertDataTable}}, \code{\link[checkmate]{assert_data_table}}, \code{\link[checkmate]{expect_data_table}}, \code{\link[checkmate]{checkDirectoryExists}}, \code{\link[checkmate]{check_directory_exists}}, \code{\link[checkmate]{testDirectoryExists}}, \code{\link[checkmate]{test_directory_exists}}, \code{\link[checkmate]{assertDirectoryExists}}, \code{\link[checkmate]{assert_directory_exists}}, \code{\link[checkmate]{expect_directory_exists}}, \code{\link[checkmate]{checkEnvironment}}, \code{\link[checkmate]{check_environment}}, \code{\link[checkmate]{testEnvironment}}, \code{\link[checkmate]{test_environment}}, \code{\link[checkmate]{assertEnvironment}}, \code{\link[checkmate]{assert_environment}}, \code{\link[checkmate]{expect_environment}}, \code{\link[checkmate]{checkFactor}}, \code{\link[checkmate]{check_factor}}, \code{\link[checkmate]{testFactor}}, \code{\link[checkmate]{test_factor}}, \code{\link[checkmate]{assertFactor}}, \code{\link[checkmate]{assert_factor}}, \code{\link[checkmate]{expect_factor}}, \code{\link[checkmate]{checkFileExists}}, \code{\link[checkmate]{check_file_exists}}, \code{\link[checkmate]{testFileExists}}, \code{\link[checkmate]{test_file_exists}}, \code{\link[checkmate]{assertFileExists}}, \code{\link[checkmate]{assert_file_exists}}, \code{\link[checkmate]{expect_file_exists}}, \code{\link[checkmate]{checkFlag}}, \code{\link[checkmate]{check_flag}}, \code{\link[checkmate]{testFlag}}, \code{\link[checkmate]{test_flag}}, \code{\link[checkmate]{assertFlag}}, \code{\link[checkmate]{assert_flag}}, \code{\link[checkmate]{expect_flag}}, \code{\link[checkmate]{checkFunction}}, \code{\link[checkmate]{check_function}}, \code{\link[checkmate]{testFunction}}, \code{\link[checkmate]{test_function}}, \code{\link[checkmate]{assertFunction}}, \code{\link[checkmate]{assert_function}}, \code{\link[checkmate]{expect_function}}, \code{\link[checkmate]{checkInt}}, \code{\link[checkmate]{check_int}}, \code{\link[checkmate]{testInt}}, \code{\link[checkmate]{test_int}}, \code{\link[checkmate]{assertInt}}, \code{\link[checkmate]{assert_int}}, \code{\link[checkmate]{expect_int}}, \code{\link[checkmate]{checkInteger}}, \code{\link[checkmate]{check_integer}}, \code{\link[checkmate]{testInteger}}, \code{\link[checkmate]{test_integer}}, \code{\link[checkmate]{assertInteger}}, \code{\link[checkmate]{assert_integer}}, \code{\link[checkmate]{expect_integer}}, \code{\link[checkmate]{checkList}}, \code{\link[checkmate]{check_list}}, \code{\link[checkmate]{testList}}, \code{\link[checkmate]{test_list}}, \code{\link[checkmate]{assertList}}, \code{\link[checkmate]{assert_list}}, \code{\link[checkmate]{expect_list}}, \code{\link[checkmate]{checkLogical}}, \code{\link[checkmate]{check_logical}}, \code{\link[checkmate]{testLogical}}, \code{\link[checkmate]{test_logical}}, \code{\link[checkmate]{assertLogical}}, \code{\link[checkmate]{assert_logical}}, \code{\link[checkmate]{expect_logical}}, \code{\link[checkmate]{checkMatrix}}, \code{\link[checkmate]{check_matrix}}, \code{\link[checkmate]{testMatrix}}, \code{\link[checkmate]{test_matrix}}, \code{\link[checkmate]{assertMatrix}}, \code{\link[checkmate]{assert_matrix}}, \code{\link[checkmate]{expect_matrix}}, \code{\link[checkmate]{checkMultiClass}}, \code{\link[checkmate]{check_multi_class}}, \code{\link[checkmate]{testMultiClass}}, \code{\link[checkmate]{test_multi_class}}, \code{\link[checkmate]{assertMultiClass}}, \code{\link[checkmate]{assert_multi_class}}, \code{\link[checkmate]{expect_multi_class}}, \code{\link[checkmate]{checkNames}}, \code{\link[checkmate]{check_names}}, \code{\link[checkmate]{testNames}}, \code{\link[checkmate]{test_names}}, \code{\link[checkmate]{assertNames}}, \code{\link[checkmate]{assert_names}}, \code{\link[checkmate]{expect_names}}, \code{\link[checkmate]{checkNull}}, \code{\link[checkmate]{check_null}}, \code{\link[checkmate]{testNull}}, \code{\link[checkmate]{test_null}}, \code{\link[checkmate]{assertNull}}, \code{\link[checkmate]{assert_null}}, \code{\link[checkmate]{checkScalar}}, \code{\link[checkmate]{check_scalar}}, \code{\link[checkmate]{testScalar}}, \code{\link[checkmate]{test_scalar}}, \code{\link[checkmate]{assertScalar}}, \code{\link[checkmate]{assert_scalar}}, \code{\link[checkmate]{expect_scalar}}, \code{\link[checkmate]{checkSetEqual}}, \code{\link[checkmate]{check_set_equal}}, \code{\link[checkmate]{testSetEqual}}, \code{\link[checkmate]{test_set_equal}}, \code{\link[checkmate]{assertSetEqual}}, \code{\link[checkmate]{assert_set_equal}}, \code{\link[checkmate]{expect_set_equal}}, \code{\link[checkmate]{checkString}}, \code{\link[checkmate]{check_string}}, \code{\link[checkmate]{testString}}, \code{\link[checkmate]{test_string}}, \code{\link[checkmate]{assertString}}, \code{\link[checkmate]{assert_string}}, \code{\link[checkmate]{expect_string}}, \code{\link[checkmate]{checkSubset}}, \code{\link[checkmate]{check_subset}}, \code{\link[checkmate]{testSubset}}, \code{\link[checkmate]{test_subset}}, \code{\link[checkmate]{assertSubset}}, \code{\link[checkmate]{assert_subset}}, \code{\link[checkmate]{expect_subset}}, \code{\link[checkmate]{checkTibble}}, \code{\link[checkmate]{check_tibble}}, \code{\link[checkmate]{testTibble}}, \code{\link[checkmate]{test_tibble}}, \code{\link[checkmate]{assertTibble}}, \code{\link[checkmate]{assert_tibble}}, \code{\link[checkmate]{expect_tibble}}}
}}

