% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-assertive-safeDeparse.R
\name{safeDeparse}
\alias{safeDeparse}
\title{Expression Deparsing}
\usage{
safeDeparse(expr, ...)
}
\arguments{
\item{expr}{\code{expression}.
Any R expression.}

\item{...}{Passed to \code{deparse}.}
}
\description{
Turn unevaluated expressions into character strings.
}
\details{
\code{safeDeparse} is modified version of \code{deparse} that is guaranteed to always
return \code{character(1)}.
}
\examples{
safeDeparse(is.character("a"))
}
\seealso{
\itemize{
\item \code{assertive.base::safe_deparse}.
\item \code{deparse}.
}
}
