% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setCause-methods.R
\name{setCause}
\alias{setCause}
\alias{setCause,logical-method}
\title{Return \code{FALSE} vector with causes of failure}
\usage{
setCause(x, ...)

\S4method{setCause}{logical}(x, false = "false", missing = "missing")
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{false}{\code{character}.
A character vector to set the cause to, when \code{x} is \code{FALSE}.}

\item{missing}{\code{character}.
A character vector to set the cause to, when \code{x} is \code{NA}.}
}
\value{
\code{goalie}.
}
\description{
Sets the \code{cause} \link[base:attributes]{attribute} of an object and returns that
object.
}
\note{
Updated 2021-02-23.
}
\examples{
x <- setCause(x = FALSE, false = "test")
print(x)
cause(x)
}
\seealso{
\itemize{
\item \code{cause()}.
\item \code{assertive.base::set_cause()}.
\item \code{stats::setNames()}.
}
}
