% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCause.R
\name{setCause}
\alias{setCause}
\title{Return \code{FALSE} vector with causes of failure}
\usage{
setCause(x, false = "false", missing = "missing")
}
\arguments{
\item{x}{Object.}

\item{false}{\code{character}.
A character vector to set the cause to, when \code{x} is \code{FALSE}.}

\item{missing}{\code{character}.
A character vector to set the cause to, when \code{x} is \code{NA}.}
}
\value{
\code{goalie}/\code{logical}.
}
\description{
Sets the \code{cause} \link[base:attributes]{attribute} of an object and returns that
object.
}
\examples{
setCause(x = FALSE, false = "test")
}
\seealso{
\itemize{
\item \code{cause()}.
\item \code{assertive.base::set_cause()}.
\item \code{stats::setNames()}.
}
}
