% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-assertive-setCause.R
\name{setCause}
\alias{setCause}
\title{Set a cause and return the input}
\usage{
setCause(x, falseValue, missingValue = "missing")
}
\arguments{
\item{x}{Object.}

\item{falseValue}{\code{character}.
A character vector to set the cause to, when \code{x} is \code{FALSE}.}

\item{missingValue}{\code{character}.
A character vector to set the cause to, when \code{x} is \code{NA}.}
}
\description{
Sets the cause attribute of an object and returns that object.
}
\examples{
setCause(FALSE, falseValue = "test")
}
\seealso{
\itemize{
\item \code{cause()}.
\item \code{assertive.base::set_cause()}.
\item \code{stats::setNames()}.
}
}
