% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{sets}
\alias{sets}
\alias{areDisjointSets}
\alias{areIntersectingSets}
\alias{areSetEqual}
\alias{isSubset}
\alias{isSuperset}
\title{Set Comparisons}
\usage{
areDisjointSets(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(y))

areIntersectingSets(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(y))

areSetEqual(x, y, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(y))

isSubset(x, y)

isSuperset(x, y, strictly = FALSE, .xname = get_name_in_parent(x),
  .yname = get_name_in_parent(y))
}
\arguments{
\item{x}{Object.}

\item{y}{Secondary object.}

\item{.xname}{\emph{Not intended to be used directly.}}

\item{.yname}{\emph{Not intended to be used directly.}}

\item{strictly}{Logical.  If \code{TRUE}, \code{x} and \code{y} should not
be set equal.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
Set Comparisons
}
\examples{
## Pass ====
areDisjointSets(x = c("a", "b"), y = c("c", "d"))
areIntersectingSets(x = c("a", "b"), y = c("b", "c"))
areSetEqual(x = c("a", "b"), y = c("b", "a"))
isSubset(x = "a", y = c("a", "b"))

## This assert is particularly useful for checking required columns.
isSuperset(
    x = colnames(datasets::ChickWeight),
    y = c("Time", "weight", "Diet")
)

## Fail ====
areDisjointSets(x = c("a", "b"), y = c("b", "a"))
areIntersectingSets(x = c("a", "b"), y = c("c", "d"))
areSetEqual(x = c("a", "b"), y = c("b", "c"))
isSubset(x = "c", y = c("a", "b"))
isSuperset(
    x = c("Time", "weight", "Diet"),
    y = colnames(datasets::ChickWeight)
)
}
