% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\name{show}
\alias{show}
\alias{show,goalie-method}
\title{Show an object}
\usage{
show(object)

\S4method{show}{goalie}(object)
}
\value{
Invisible \code{NULL}.
}
\description{
Show an object
}
\note{
Updated 2022-02-07.
}
\examples{
## Match the default logical print method.
x <- goalie(
    object = c("aaa" = FALSE, "bbb" = TRUE),
    cause = c(
        "'aaa' is FALSE.",
        NA_character_
    )
)
show(x)
}
\author{
Michael Steinbaugh
}
