% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show}
\alias{show}
\alias{show,goalie-method}
\title{Show an object}
\usage{
\S4method{show}{goalie}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output.
}
\description{
Show an object
}
\note{
Updated 2021-02-23.
}
\examples{
## Match the default logical print method.
x <- goalie(
    x = c("aaa" = FALSE, "bbb" = TRUE),
    cause = "'aaa' is FALSE."
)
show(x)
}
\author{
Michael Steinbaugh
}
