% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate an S4 Class}
\usage{
validate(...)
}
\arguments{
\item{...}{unnamed expressions that describe the conditions to be tested.
Rather than combining expressions with \code{&&}, separate them by commas
so that better error messages can be generated.}
}
\description{
\code{\link[=validate]{validate()}} is a variant of \code{\link[=assert]{assert()}} that is specifically intended to be
used inside of an S4 validity method definition.
}
\details{
Like \code{\link[=assert]{assert()}}, \code{\link[=validate]{validate()}} returns \code{TRUE} on success. However, on failure
it returns a \code{character} instead of a \link[base:stop]{stop()} call. This is the
current recommended practice for defining S4 validity methods inside of a
\link[methods:setValidity]{setValidity()} call. Refer to the documentation in the
methods package, specifically \link[methods:validObject]{validObject()} for
detailed information on S4 validity methods.
}
\examples{
validate(
    is.atomic("example"),
    is.character("example")
)
}
\seealso{
\itemize{
\item \code{methods::setValidity()}.
\item \code{methods::validObject()}.
\item \code{assertthat::validate_that()}.
}
}
