% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateClasses.R
\name{validateClasses}
\alias{validateClasses}
\title{Validate Expected Classes}
\usage{
validateClasses(object, expected, subset = FALSE)
}
\arguments{
\item{object}{Object.}

\item{expected}{\code{list}. Named list of expected classes per slot.}

\item{subset}{\code{logical(1)}. Only check a subset of slots in the object.}
}
\value{
\code{logical(1)} (\code{TRUE}) on sucess or \code{character(1)} containing
informative message on failure.
}
\description{
Validity check capable of validating multiple slots in a single call.
}
\details{
To be used inside S4 \code{\link[methods:setValidity]{methods::setValidity()}} call or with \code{\link[=validate]{validate()}}.
Particularly useful for checking multiple slotted objects inside
\link[S4Vectors:metadata]{metadata}.
}
\examples{
validateClasses(
    object = list(
        a = character(),
        b = integer(),
        c = factor()
    ),
    expected = list(
        a = "character",
        b = "integer",
        c = "factor"
    )
)
}
\seealso{
\itemize{
\item \code{validate}.
\item \code{methods::setValidity}.
\item \code{methods::validObject}.
}
}
