% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-flag-validNames.R
\name{validNames}
\alias{validNames}
\title{Are these valid names?}
\usage{
validNames(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\code{TRUE} on success;
\code{FALSE} on failure, with \code{cause} \link[base:attributes]{attribute} set.
}
\description{
Are these valid names?
}
\examples{
## Dots (periods) and underscores are valid.
validNames(c("sample.1", "sample_1"))

## Can't begin with a number.
validNames("293cells")

## Spaces, dashes (hyphens), and other non-alphanumerics aren't valid.
validNames("sample 1")
validNames("cell-AAAAAAAA")
validNames("GFP+")
}
\seealso{
\itemize{
\item \code{make.names()}.
\item \code{basejump::makeNames()}.
}
}
