% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validNames.R
\name{validNames}
\alias{validNames}
\title{Are These Valid Names?}
\usage{
validNames(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}:
\code{logical(1)}. Boolean flag (\code{TRUE}/\code{FALSE}).
Intended to be used in combination with \code{assert} function.
\item \code{check*}:
\code{TRUE} on success or \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}:
\code{TRUE} on success or error (\code{stop}) on failure.
}
}
\description{
Are These Valid Names?
}
\examples{
## Dots (periods) and underscores are valid.
validNames(c("sample.1", "sample_1"))

## Can't begin with a number.
validNames("293cells")

## Spaces, dashes (hyphens), and other non-alphanumerics aren't valid.
validNames("sample 1")
validNames("cell-AAAAAAAA")
validNames("GFP+")
}
\seealso{
\itemize{
\item \code{make.names}.
\item \code{basejump::makeNames}.
}
}
