% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-validNames.R
\name{validNames}
\alias{validNames}
\title{Are these valid names?}
\usage{
validNames(x, .xname = getNameInParent(x))
}
\arguments{
\item{x}{Object.}

\item{.xname}{\emph{Not intended to be used directly.}}
}
\value{
\itemize{
\item \code{are*}/\code{contains*}/\code{has*}/\code{is*}: \code{logical(1)}.
Boolean flag (\code{TRUE}/\code{FALSE}). Intended to be used in combination with
\code{assert} function.
\item \code{check*}: \code{TRUE} on success, \code{character(1)} on failure.
Intended to be used in combination with \code{validate} function.
\item \code{assert*}: \code{TRUE} on success, error (\code{stop}) on failure.
}
}
\description{
Are these valid names?
}
\examples{
## Dots (periods) and underscores are valid.
validNames(c("sample.1", "sample_1"))

## Can't begin with a number.
validNames("293cells")

## Spaces, dashes (hyphens), and other non-alphanumerics aren't valid.
validNames("sample 1")
validNames("cell-AAAAAAAA")
validNames("GFP+")
}
\seealso{
\itemize{
\item \code{make.names()}.
\item \code{basejump::makeNames()}.
}
}
